[NU-39]: Introduce ICP class.
The bug reported in NU-39 is due to us using doubles to represent ICP.

There are a number of problems with using doubles:
* It's subject to rounding errors, as already reported in NU-39.
* ICP can hold invalid values (e.g. negative numbers, < 1e8)

To fix these issues, I am creating a class `ICP` that:
* Encodes ICPs using `BigInt` to solve the rounding issues we're seeing
  with doubles.
* Adds validation checks to ensure ICPs are initialized with sane values.

In this commit, I am only modifying the `BalanceWidgetTest` to minimize
the scope. In a follow-up PR I'll expand the use of `ICP` to other widgets.

Additionally, I started adding some unit tests, which I think are much
needed for this codebase.