fix: Fix changing controllers flow  (#53)
* Make cycles conversion description more clear

* fix: Fix changing controllers flow

The flow for changing controllers was broken, as the `js-agent` was
expecting a `Principal` but the dart code was passing a `string`. This
PR fixes this bug and allows controllers to be modified. Additionally,
this PR contains the following enhancements:

* Updates the frontend to use the new multiple controllers API. This is
necessary to support multiple controllers.
* Added a command to automatically regenerate the TS and JS associated
with the management canister by running `npm run update_candid` in
`js-agent`.