NU-146: Replace CanvasKit renderer with HTML renderer.
The CanvasKit renderer provides very little gain and causes an additional
download of ~6.4MiB on firefox and 2.5MiB on Chrome.

All the pages look and behave identically with the HTML rendered, with
the exception of the voting page, which page feels slightly sluggish,
but removing the unnecessary animations on that page would fix this issue.