EXC-405: Use protobuf endpoint when adding hot keys. (#145)
The Ledger wallet will only support the protobuf interface, so moving
to the protobuf API is a pre-requisite to integrating this transaction
with the ledger wallet.

Since there's shared proto between the governance and ledger, I moved
all the proto into a `proto` folder. Additionally, I deleted the generated
proto files and are instead generated locally when building. This guarantees
that the generated files will always be based off of the latest proto
 files.