EXC-456: Improved error handling for ledger wallet transactions.
This PR is the first of several that improve the error handling flow
of ledger wallets specifically, and of the dart code in general.

I introduced the rust notion of "Result". Rather than relying on exceptions
where it's easy to forget wrapping things in try/catch, I'm proposing
we move closer to the Rust model where we put things in a Result object
to force callers to handle errors.

In this PR I've updated stake neuron and disburse to return `Result`,
and I've updated the callers to display the error to the user. Similar
changes will follow for the other manage neuron operations in upcoming
PRs.