EXC-566: Add "certification" to Ledger.getBalances calls. (#230)
This PR adds additional security when fetching ICP balances from
Ledger.getBalances.

Instead of only having a query call, we now do two calls in parallel:

1. A query call, which is fast but uncertified. We show this result
   immediately.
2. An update call, which is certified. We update the page with this
   result as soon as it is received.

In 99% of the cases, the result of both calls will be the same and this
change will be completely unnoticeable to the user. There's a slight
chance of some race condition, but in that case, updating the UI with
the certified response remains a sensible approach.