[EXC-587, EXC-590] Implement PUQ for `get_account` and `get_transactions`
For lack of a better term, I'll be calling the approach of certifying
by doing parallel update/query calls "PUQ" (Parallel update query).

In this PR I add PUQ to get_accounts and get_transactions.

Note that, while this is a step in the right direction, the work is not
considered complete, as there can be race conditions.

For example:

* `get_account` query call fetches a malicious account identifier.
* Before the result of the `get_account` update call is returned, the
  `ledger_sync_service` starts PUQing the ICP balance of the malicious
  account identifier.
* Response from the update call of `get_account` is returned, updating
  the UI to show the correct account identifier, but the balance
  remains that of the malicious account identifier.

I'll be restructuring the syncing logic in another PR to ensure that
these race conditions are handled gracefully.