Add a vanilla svelte app (#271)
# Motivation
The front end is to be rewritten in svelte.  There is some svelte boiler-plate code.  It seems preferable to add the code that has no meaning beyond "because this is svelte" in a separate step.

# Changes
Added a blank svelte app as described here: https://svelte.dev/blog/the-easiest-way-to-get-started#2_Use_degit
```
cd frontend
npx degit sveltejs/template svelte
cd svelte
node scripts/setupTypeScript.js
npm install
git add .
git commit
```

Note: This svelte code is not added into the canister as part of the build step.  It can be used as a base for local development, no more.