Make the fonts compiled into flutter accessible to svelte (#286)
# Motivation
The svelte app should use the same fonts as flutter.  However the flutter fonts don't appear to be available from download from the canister, they are probably compiled into the flutter app, so we need a separate copy for svelte.  Eventually the flutter app will be gone anyway so this copy will be the only copy remaining.

# Changes
- Copied the flutter fonts into the svelte codebase.
- Provided the urls to the fonts in the global.css, so that they can be used.