Allow specifying OWN_CANISTER_ID (#365)
The `update_config.sh` script needs to know which canister ID we build
for, and so far has called out to `dfx` to know which canisters have
been deployed locally (for local deploys). This is a problem in the
docker build since we don't deploy canisters. Instead, we allow
specifying the canister ID. Stop gap measure while we find a better
solution.