Sort CSS using rollup plugin
This reverts to the original plugin (pre
https://github.com/dfinity/nns-dapp/pull/409) and actually performs the
sorting in the rollup config as opposed to using an extra bash script.

I've also created an issue upstream and waiting for more info on the
maintainer re. what the root cause may be, and potentially how to avoid
the `mkdir` step here:
https://github.com/thgh/rollup-plugin-css-only/issues/42