build: CSP and other adjustements to make deploy happens (#444)
# Motivation

Define the Content Security Policies (CSP) that should be applied in production.

Make the svelte app deployment to canisters - i.g. testnet and mainnet - happens (without adjustements the app can indeed be deployed but is not functionable).

# Changes

- prerender the CSP meta rules statically within the `index.html`
- apply rules (following [Internet Identity](https://github.com/dfinity/internet-identity/blob/c5709518ce3daaf7fdd9c7994120b66bd613f01b/src/internet_identity/src/main.rs#L824) example)
- refactor `favicon.png` to use the same asset in both svelte and flutter app
- remove CSP rule for local development only
- fix "no sourcemap in production build"
- extract logic and default environment variables to a new `env.config.mjs` script 

# Note

This PR does **not** add new env variables to the rollup configuration and per extension to the JS code. In addition to CSP, it only cleans up and group variables within a same configuration that can be use for the svelte build.