style: fix modal overflow (#552)
# Motivation

When the wizard is used in the modal, the transition are glitchy because an horizontal scroll is briefly displayed on transitions.

When the content of the modal fits, a vertical scrollbar is displayed. That should not be the case. Such a scroll should be displayed only if the content can be scrolled.

# Changes

- style `overflow-x` and `overflow-y` of the modal's content