Flutter and Svelte in parallel (#627)
# Motivation
When testing Svelte it is useful to be able to compare with flutter.  At the moment there are several deployments, including prod and staging.  However it is not enough.  Sometimes there is a strange state and it is useful to be able to examine that state on the same canister, so there need to be redirects from neither svelte nor flutter.

References:
- https://dfinity.atlassian.net/browse/L2-440

# Changes
- Defined a new value for `REDIRECT_TO_LEGACY`.
- Removed old values such as `1`
- Documented the values of `REDIRECT_TO_LEGACY`.