feat: ignore error "Neuron already voted on proposal." (#726)
# Motivation

If user has for example two neurons with one neuron (B) following another neuron (A). Then if user select both A and B to cast a vote on a proposal, A will first vote for itself and then vote for the followee B. Then our function `Promise.allSettled` that register all votes process next neuron B and try to vote again. This leads to an error `"Neuron already voted on proposal."` thrown by the [Governance canister](https://github.com/dfinity/ic/blob/936d234f2e6179cbaf6e5628df0f77e594e03505/rs/nns/governance/src/governance.rs#L5264) because it has indeed already been registered by A.

Vote are correctly cast therefore we can safely ignore this particular error.

# Note

This has been thankfully found by [justmythoughts](https://forum.dfinity.org/u/justmythoughts) on the [forum](https://forum.dfinity.org/t/nns-voting-error-nns-app-ui-shows-error-when-voting-every-time/12212/12).

# Why this solution

After some more research, we noticed that this error was not new and already existed in the Flutter version of the app. However, the error was ignored, therefore not presented to the user and silent. As it was not explicitly coded that way, we couldn't have noticed it in our reverse engineering that served the development of the Svelte rewrite.

We discussed  different way to process the neurons tree on the frontend side to query the backend with only unique neurons but any solution also need some UX improvements as user somehow should also be made aware of the relations between the neurons on the proposal detail's page - voting card.

That's why we decided to postpone such a tasky enhancements for the future as we aim to improve the overall UX and for now on, to ignore safely the error message that is actually an information.

# Changes

- ignore "Neuron already voted on proposal." when processing the responses of the register of all votes
- remove debugging information and workaround that made the selected neuron ids unique