Rm DEPLOY_ENV (#1034)
# Motivation
`DEPLOY_ENV` duplicates `DFX_NETWORK`.  The latter is set by `dfx` so there is no need of the former.

The only complication is that `DEPLOY_ENV` has one value that is not a network: `nobuild`.  This uses the existing WASM.  That `nobuild` functionality would have to be replaced by something else.

# Changes
* Replace `DEPLOY_ENV` with `DFX_NETWORK`, simplifying code where previously both were used.
* Introduce the `NNS_DAPP_WASM` environment variable.  If set, it points to a wasm file to use instead of building.