build: remove https: from CSP script-src (#1492)
# Motivation

We use the `https:` in the `script-src` rules of the CSP for backwards compatibility as displayed by [MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src). However we have a report from a user who cannot access NNS-dapp on his Android v10 device. We are not sure what his issue is but, using Android Simulator, we can reproduce a blank screen on OS <= Android v12. That's why, in accordance with Security team, we remove the `https:` rule that solves the issue in the simulator and which is in any case not used in modern browsers that support `strict-dynamic`.

# Changes

- remove `https:` rules from CSP `script-src`