Support testnets that make use of the fixed SNS-WASM canister deployed via icos_deploy.sh (#1475)
# Motivation

The SNS-WASM canister deployed via icos_deploy.sh (which use
ic-nns-init) should now be used instead of a custom SNS-WASM canister.
This PR includes two new propose-to-scripts that will help set up the
SNS-WASM canister to work with a newly deployed testnet.
# Changes

- The `propose-to-add-wasm-to-sns-wasm` will use the
test-neuron-proposer to submit NNS Governance proposals to add the
various downloaded SNS canister WASMs. These should automatically pass
if no other neurons are created in the testnet.
- The `propose-to-update-sns-wasm-sns-subnet-ids` will use the
test-neuron-proposer to submit an NNS Governance proposal to update the
SNS-WAMS canister to be able to deploy to the newly created testnet
subnets.
- The Dockerfile has been updated to support the new location of the
`icp_ledger/ledger.did` which was just moved in the monorepo.
- dfx.json now points to the fixed canister id of the SNS-WASM canister
(`qaa6y-5yaaa-aaaaa-aaafa-cai`). Now when deploying nns-dapp via the
deploy script, the dapp will be configured to point at this canister
instead of the custom one.