feat: public private routing (#1474)
# Motivation

Instead of an up-front sign-in screen, we want user to be to sign-in on each screen. Doing so we will be able to display meaningful static information on each pages and maybe be even able to display public information while not being signed - i.e. render proposal even if not signed in.

# Breaking changes

This PR switch the custom routing implementation to a SvelteKit base routing. This is a **BREAKING CHANGES**.
We switch from an hasbang base routing system to a query params system. In addition, we migrate from a SPA to a multi page app.

There is not backwards compatibility for the urls. Hashbang params will be ignored.

However, to ease the launch on mainnet without the need to synchronize the proposal execution with the dashboard, the PR contains a workaround to route `/#/proposal/{id}` from login page to proposal detail page. Workaround we aim to ultimately, rather soon than later, delete.

# Approach

We use [+layout.ts](https://kit.svelte.dev/docs/routing#layout-layout-js) file to extract the parameters (universe and details like "&proposal=123") request.

The "universe" param being a global information, it gets derived in a custom `pageStore` - i.e. we use SvelteKit [$app/stores](https://kit.svelte.dev/docs/modules#$app-stores) to derive our own page information.

Other fine grained params - e.g &proposal=123" - a passed down from the `+page` to its content.

# Limitation

This PR does not support unknown routes. i.e. if user would enter `nns.ic0.app/yolo` it will face a 500 error uncatched. See "Open questions".

# Performance

Optimizing the performance of the dapp is a battle between loading everything in a single JS bundle files and finding the right amount of chunks.

This PR does the following:

- it uses Rollup `manualchunks` to bundle all `node_modules` - except Svelte related components and code - dependencies in a `vendor.js` file and a `dapp` file for the app

- SvelteKit chunks the rest of the app

- sometimes we might want to limit the number of chunks for fine tuning. That's why the `login` page does not use a `+layout.svelte` file but includes the layout as a component or why the route are still lazy loaded with our `RouteModule`

# Security

The CSP parser has been extended to parse the rules in each and every single `.html` file that is generated.