fix: init auth worker race condition (#1519)
# Motivation

When deployed on the IC, it is possible that fetching the worker takes more time than fetching the code to sync the identity. This can lead to a race condition in the case the user proceed as following:

- user access nns-dapp
- user sign-in
- user refresh the browser

# Changes

- synchronize the auth worker not only when the auth changes but also right after the worker has been loaded

# Tests

This has been debuged locally by setting a delegation that least only 20s and tested manually in various scenario.