feat: increase stake of a sns neuron (#1639)
# Motivation

Add a feature to increase the stake of a Sns neuron.

# ⚠️ Warning

As for "staking Sns neuron", this PR follows the existing same Sns pattern that propose a payment flow which happens on the frontend side only. Therefore, it isn't fully reliable. e.g. if transfer succeed and refresh isn't called, the stake of the neuron is not increase while the token haven been transfered. This would need refreshing the neuron again, the user would have to increase the stake again. The resulting would be the sum of both actions. I decline any responsability about this.


Worth to note that we will soon begin to strenghtness the payment flows therefore this will be ultimately handled in a better way.

# Changes

- implement service and api to increase stake of a sns neuron
- add a button to the sns neuron detail page to open the modal to process the increase of the stake (similar to nns)
- implementation transaction modal flow and call increase stake neuron on execution (similar flow to nns)
- load params and accounts - `snsSyncAccounts` - if necessay on button click and display a spinner while loading the data
- misc. files name refactoring to make obvious which component is Nns and which is Sns