feat: display TVL (total value locked) in USD  (#1725)
# Motivation

Display the total value locked (tvl) in US dollar.

# ⚠️ Warning

This feature does not use certified data. It fetches exchange rate from Binance API (web2) and the TVL from the `.raw` domain of the governance canister.

# Future

Our plan is to use the new exchange canister (XRC) through a proxy - that caches the ICP<->USD rate - and get the TVL information with a certified or query+update calls.

# Rules

- if any of these API crashes an error, it should be silently* ignored and nothing - no TVL - should be displayed to the user
- if zero values are provided, likewise, nothing shold be displayed

* This feature is not decisive for user flow but stricly informative therefore we can/should silent the potential issues (not speaking of the fact that the endpoints are not certified)

# Notes

This feature performs the call for following reasons:

1. UI performance
2. Security

## UI performance

I personally think that it is not the job of the frontend UI to keep track and run cronjobs necesserary to fetch realtime data. I think that deferring these to workers reduce the load of the UI on the window side and therefore leads to better performance.

## Security

By executing the requests on the workers side we ensure that these are sandboxed outside the window. In addition, by doing so, we do not have to open our Content-Security-Policy to introduce exception for some web2 or raw domains.

# Changes

- new `<TotalValueLocked />` component and usage
- implementation of `rest` calls and interpretation of the results (in case of the governance)
- adds worker to fetch and re-evaluate the value every minute as cronjob
- remove ic link to spare space (kind of a duplicate of nns link to ic.org)
- few stylish adjustements on the login screen (like hover color)