More formatting and linting (#1817)
# Motivation
The formatting and lint commands are incomplete. Take 10 minutes to make
the situation better.

# Changes
- Added a yaml formatter (the one I use locally but it is not in CI)
- Format the YAML files
- ~Update the e2e-test lockfile, as it is out of sync with packages.json
and cannot be used as is.~ actually, it's just generated with an older
nodejs.
- Create a single top-level lint command, similar to the top-level fmt
command. The list of lint shell scripts is incomplete, so the top level
command is necessarily also incomplete.

# Not included
These are not included in this PR:
- frontend linting (encoded in ci but lot as a local shell script)

# Tests
- See CI for the formatting update
- I have run the linter locally. While convenient, the other lint
commands should be added before it makes sense to use it in CI, in my
opinion.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>