feat: use "Link canister" instead of "Attach" for consistency reason (#1797)
# Motivation

There is a lack of consistency about the terms "Attaching" or "Linking" canisters. Sometimes we use "Attach", sometimes "Link".

As "Link" seems more userfriendly than "Attach", we decided to go on with the term "Link" and to rename "Detach" with "Unlink" for consistency reason as well. 

# Notes

No functional changes. Wordings and refactoring only is provided in this PR.

# Changes

- review `en.json`, generate new keys and adapt code accordingly
- rename few components and their usage