Factor out need checking (#1767)
# Motivation
This is a follow up on conversations in
https://github.com/dfinity/nns-dapp/pull/1759 In particular, the logic
that goes at the end of each workflow was being fine-tuned. Rather than
making the changes in many places, it possibly makes sense to factor out
that final logic and tweak it in one place.

Why not to merge this PR: The common code is hardly longer than the code
used to call up the common code, so factoring out seems a bit pointless,
however it does make sure that the common code stays in sync. A cost of
such refactoring is that the repo needs to be checked out to get the
file with the common logic and run it. The checkout takes about 1s in
the gh action. Still, overall, this change may be worthwhile.

# Changes
- Create a local saved github action that checks that all needed jobs
have completed successfully.
- Incorporate some ideas from #1759 into the common code.
- Replace the final check in each workflow file with a call to the
common code.

# Tests
No additional checks should be needed.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>