Collect some data early (#1861)
# Motivation
The aggregator update interval is configurable and can be set to a long
time. Fine, but that means that after deployment, it can take quite a
while before there is any sign of life. So, on installation, make sure
that one SNS is scraped promptly.

# Changes
* Start pulling data quickly after installation.
* Slow down the default update frequency to a more prod-like value.

# Tests
CI should suffice. Also, this has been used for deployments to testnets
from a feature branch.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>