Allow settings API_HOST and STATIC_HOST instead of HOST in dfx.json (#1900)
# Motivation

When we fetch data from the aggregator canister, this does not go
through a candid API.
This allows us to certify the data without using an update call.
With the domain migration, candid API calls will have to go through a
different domain.
So we need to be able to specific separate domains for candid API calls
vs. static content fetches.

# Changes

Change config.sh to read API_HOST or STATIC_HOST depending on whether
the host is used for candid API calls or static content fetches. Both
fall back on HOST if not set.

# Tests

Set different values for API_HOST and STATIC_HOST on medium07 and
noticed the difference in the `.env` file.
The `.env` file is the same as before when the `dfx.json` file is
unchanged but the new hosts are used when present.
Also verified in the browser console that the failed aggregator canister
call goes to the new URL.