Bump ic commit 003 (#1903)
# Motivation
The sns aggregator needs to handle newer SNSs. The latest IC commit
supported by the snsdemo is 0a5b8387 from 5 Feb. So let's use that.

# Changes
- Run: `./scripts/update_ic_commit --ic_commit
0a5b8387179d2dec8269cc48b907ab0b334fdac1`
  - Note: This:
- Updates the IC_COMMIT in dfx.json and so also updates the canister
versions used in tests.
    - Updates the SNS .did files
    - Updates the rust code derived from the did files
- Update the patch file for dev features such as fast wait for quiet.
- Note: The patch is applied to the IC codebase. The patch is the same
but had to be pointed at a different location, as the code had moved.
- Tweak the candid -> rust conversion that leads to a compiler error out
of the box.
- Note: the didc candid -> Rust conversion has trouble with empty enum
variants. It creates them in a valid way, but in a way that is
incompatible with the CandidType macro.

# Tests
CI should suffice

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>