GIX-1340: Retry get open ticket (#1941)
# Motivation

Many users trying to participate at the same time might overload the sns
sale canister.

We want to be the ones managing that experience, instead of showing
errors.

Therefore, we retry getting the open ticket from the sale canister if we
get a not identified error.

# Changes

Main changes to review:
* New helper `pollGetOpenTicket` used in `loadOpenTicket`.
* New property `keepPolling` in the SnsTicketsStore.
* New methods in `stopPolling` and `enablePolling` in SnsTicketsStore.

Other changes:
* Changes in error handling in `loadOpenTicket`.
* New parameter in `poll` util: `millisecondsToWait`
* Move `pollingLimit` next to the `poll` util.
* Changes in the onDestroy of the `ParticipateButton`.
* Refactor `hasOpenTicketInProcess` to pass the store so that it's
recalculated when store changes.

# Tests

* More tests for `loadOpenTicket` service.
* Add missing test for `restoreSnsSaleParticipation`.
* New tests for snsTicketsStore
* Fix broken tests after changes.