In poll(), don't wait after the last retry (#1980)
# Motivation

We have a `poll()` function which retries a call until it passes or some
other condition is met.
If you reach the maximum number of retries, it returns an error but
currently it still waits another timeout before doing so.

# Changes

Increase and check the counter before waiting for the timeout instead of
after calling the recursion.

# Tests

Added a unit test and ran existing unit tests.