Feat: get_open_ticket update call (#2002)
# Motivation

Avoid unnecessary update calls.

Our calls to SNS canisters use the sns wrapper. The sns wrapper calls
the SNS-W and Root canister on initialization.

Yet, we don't really need to because the canister ids for each SNS
project are already provided by the SNS Aggregator.

Therefore, this PR builds the wrappers after getting the data from SNS
Aggregator and adds them to the cache.

Otherwise, there would be two extra update calls to SNS canisters when
we get the open ticket with the update call.

# Changes

* New util `buildAndStoreWrapper` in `sns-wrapper.api`.
* Use the new util `buildAndStoreWrapper` after fetching the data from
the SNS Aggregator.
* In ParticipateButton, wait until summary is present to load the
ticket. This way the cached wrapper will be used.
* Changes in `loadSnsSwapCommitment`. Check whether we already have the
user commitment to avoid getting it again.
* Change `rootCanisterId` for `swapCanisterId` in the `getOpenTicket`
sns sale api to use the SnsSwapCanister directly instead of the wrapper.
* Fix calls to `getOpenTicket` and parent calls to add the
`swapCanisterId` as param

# Tests

* Add test for `buildAndStoreWrapper`.
* Add test case for `wrappers`.
* Add missing test for `loadSnsSwapCommitment`.
* Fix tests after refactoring getOpenTicket to use SnsSwapCanister.
* Add test cases in app.services