Clean up polling tests (#2007)
# Motivation

https://github.com/dfinity/nns-dapp/pull/2004 added a number of `await
null` to tests which use polling to make tests work without it being
fulling clear what's going on.
This PR replaces all the `await null` with `advanceTime` which does not
need to be called an arbitrary number of times.
I also made additional cleanups to make it clearer what's really going
on in the tests.

<!-- Describe the motivation that lead to the PR -->

# Changes

1. Add a test util to advance fake timers while also running all already
resolved promise in order to get the full affect of advancing time.
2. Use the test util to allow removing all the `await null` statements.
3. Reset mocks to avoid one failure resulting in additional failure
because of left over mock state.
4. Don't wrap expectations in `waitFor` when not necessary.
5. Instead of making additional rounds in a loop, to additional time
advancement outside the loop.
4. Rename `retriesBefore...` to `callsBefore...`. "Retries" is confusing
because you could think that a retry only happens after the first try so
it could be 1 less than the number of calls.
3. In `sns-sale.services.spec.ts` rename `sendICPSpy` to `spyOnSendICP`
for consistency with the other spies in the same file.

<!-- List the changes that have been developed -->

# Tests

Tests only

<!-- Please provide any information or screenshots about the tests that
have been done -->