Add $tests as import root for "./src/tests". (#2035)
# Motivation

Currently we use a variable number of "../" to refer to "src/tests" but
it's hard to tell if the number of "../" actually refers to "src/tests"
or some deeper directory.

# Changes

Introduce $tests to refer to "src/tests" and use it in all `.ts` files
under src/tests.
Apart from the changes in `frontend/jest.config.cjs` and
`frontend/tsconfig.spec.json`, the changes were made mechanically as
follows:
```
1. vim `grep -l '"../[^.]' frontend/src/tests/*/*.ts`
2. In vim do `:%s/from "\zs\.\.\ze\/[^.]/$tests` in every buffer.
3. vim `grep -l '"../../[^.]' frontend/src/tests/*/*/*.ts`
4. In vim do `:%s/from "\zs\.\.\/\.\.\ze\/[^.]/$tests` in every buffer.
5. vim `grep -l '"../../../[^.]' frontend/src/tests/*/*/*/*.ts`
6. In vim do `:%s/from "\zs\.\.\/\.\.\/\.\.\ze\/[^.]/$tests` in every buffer.
7. vim `grep -l '"../../../../[^.]' frontend/src/tests/*/*/*/*/*.ts`
8. In vim do `:%s/from "\zs\.\.\/\.\.\/\.\.\/\.\.\ze\/[^.]/$tests` in every buffer.
```

# Tests

`npm run test`