Feat: Improve merge neurons flow (#2051)
# Motivation

Merge neurons is a source of confusing and bad UX for users.

While we work on a better and more sophisticated solution, this PR
addresses main pain points and tries to explain better what's happening.

It uses only text to do so. No new UX is added in this PR.

# Changes

* Add explanation text in SelectNeuronsToMerge.svelte
* Add explanation text in ConfirmNeuronsMerge.svelte

# Tests

Only UI changes.