cleanup: Remove setLoading on snsQueryStore and snsProposalsStore (#2065)
# Motivation

snsQueryStore and snsProposalsStore use `undefined` to mean reset and
`null` to mean loading. But then when checking whether they are loading
we just use isNullish. This is confusing and it's also not super clear
that setLoadingState actually resets all the data.

To make this obvious we use reset() instead of setLoadingState() and
expose separate stores to check whether the store is loading via
isNullish.

# Changes

1. Remove setLoading from snsQueryStore and snsProposalsStore.
2. Add snsQueryStoreIsLoading and snsProposalsStoreIsLoading
3. Use `.reset()` instead of `.setLoading()`.
4. Use `snsQueryStoreIsLoading` instead of
`isNullish($snsSummariesStore) || isNullish($snsQueryStore)`.
[snsSummariesStore](https://github.com/dfinity/nns-dapp/blob/4588644d282a9a3bbe4df3bac35215021f277361/frontend/src/lib/stores/sns.store.ts#L271)
can't actually be non-nullish and `isNullish($snsQueryStore)` is what
`snsQueryStoreIsLoading` does.
5. Use $snsProposalsStoreIsLoading instead of
isNullish($snsProposalsStore)

# Tests

1. Changed an existing unit test for snsQueryStoreIsLoading and added a
new test for snsProposalsStoreIsLoading.
2. Manual testing.