GIX-1328 Only show sns neurons for committed projects (#2069)
# Motivation

Currently it's possible to see founder neurons before the sns project is
committed by manually change a URL parameter.
The current code checks that the URL parameter represents an existing
project but not that the project is committed.

# Changes

In Neurons.svelte, instead of only rendering SnsNeurons if the selected
project exists, also check that it's committed.

# Tests

Unit tests and manual testing.