GIX-1401: Setup new helper `loadbalance` (#2066)
# Motivation

After each transaction we were calling `syncAccounts`, which loads the
balance from nns-dapp canister plus gets the balance of all the
accounts.

This is not necessary. We only need to update the balance of the account
that made the transaction.

# Changes

Main changes:
* New account service `loadBalance` that queries the balance of one
account and loads in the store.
* New method `setBalance` in `accountsStore`.

Minor changes:
* Use new `loadBalance` after staking an nns neurons.
* Use new `loadBalance` after disbursing an nns neuron.

# Tests

* Test new account service `loadBalance`.
* Test new accountsStore `setBalance`.
* Test that staking neuron queries balance.
* Test that disbursing neuron queries balance.