Introducing fake api modules for easier testing (#2077)
# Motivation

Setting up mocks for api calls during tests can be tedious and the
result is often not easy to read.

A fake holds state and responds to calls in a consistent way based on
the internal state, which is set up for the test.
It can also provide sane defaults for values that don't matter during
the test.
This means the test can just set up the state it cares about and the
fake will take care of the rest.

# Changes

1. Add partial fakes for governance.api, sns-aggregator.api and
sns-ledger.api.
* Meaning, not all of the functions on these apis are implemented and
some just return empty or default values which can't yet be controlled
by the test.
2. Add a fake for sns-governance.api which also fakes some functions on
sns,api, which should really be moved to sns-governance.api because they
are backed by the same state.
3. Update NeuronDetail.spec.ts and Neurons.spec.ts to use the new fakes.
4. Update module.test-utils.ts to allow installing partial
implementations of modules in addition to blocking calls to the
functions for which no implementations are provided.
5. Update tsconfig.json to allow using "$tests" imports in fakes.

# Tests

Tests only.

<!-- Please provide any information or screenshots about the tests that
have been done -->