Downgrade/upgrade test for aggregator (#2088)
# Motivation
Basic test that the canister can be downgraded and upgraded without
bricking the canister.

Note: This does not catch subtle issues such as data corruption when
deserializing and reserializing. It just checks whether the canister can
transition and is still alive after each transition.

Note: The test does a downgrade then an upgrade rather than the other
way around as it is faster. When testing, the new code is typically
already installed and running, so downgrade+upgrade requires two
installations, whereas upgrade downgrade needs four:
downgrade-upgrade-downgrade-upgrade

# Changes
- Add a script to perform the downgrade-upgrade test
- Exercise the script in CI

# Tests
See the new tests in CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>