GIX-1372: Update elected replica versions payload (#2059)
🚧 replaced with https://github.com/dfinity/nns-dapp/pull/2079
`accounts_store.rs` was upgraded but downgrade-upgrade test isn't happy
about changes.

# Motivation

Add proposal to update elected replica versions support.
https://dfinity.atlassian.net/browse/GIX-1372

# Changes

- `UpdateElectedReplicaVersions` description and support
- fixed typo in `BlessReplicaVersion` description
- Old types can be decoded as new types, but new types couldn't be
decoded as new types. Therefore, we decode the accounts store decodes to
the old types for backwards compatibility.

# Test

- Add transaction in decoding and encoding test.

# Next steps

Once this PR is in production and we know there won't be a rollback, we
want to encode with new types and remove all the old types.

---------

Co-authored-by: Llorenç <llorenc.muntaner@dfinity.org>