Create SNSs in parallel (#2117)
# Motivation
Faster CI. There are two things that take a lot of time in the current
CI. One of those is creating many SNS for the aggregator test.

# Changes
- Use the snsdemo command that creates SNSs in parallel

# Tests
This is a change to a test; see CI for the result...

In particular, at present the aggregator test currently takes between 18
and 21 minutes. That includes several other steps but 7m of that is
creating 11 SNSs. If successful, this PR should take a few minutes off
that. It is hard to estimate how much less as the github runners are
small single core VMs and that limits the value of running things in
parallel.

Result: The aggregator test took 14 minutes, so the speedup is
worthwhile.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>