Chore: Fail SNS aggregator with less than 2 SNSes (#2120)
# Motivation

Current SNS Aggregator doesn't persist data between upgrades. Therefore,
it might be that for a few minutes the users of the NNS Dapp do not see
OC and SNS-1 in the UI.

As a temporary solution, we'll revert to the fallback if the number of
projects returned from the aggregator is less than 2.

# Changes

* Add a check after receiving the response from the aggregator. If less
than 2 projects, raise an error to use the fallback.

# Tests

* Fix tests so that they don't run into the issue of less than 2.
* Add a test case to check that the fallback is used in case the number
of SNSes is less than 2.