Remove attribute checking from page object constructors. (#2136)
# Motivation

We want to reuse page objects for e2e tests.
Some things, such as getting element attributes, have to be async in e2e
and so they will have to be async in shared page objects.
This means we can't do it in a constructor.
Currently, page objects check that the element has the expected data-tid
in the constructor.

# Changes

1. Remove attribute checking in page object constructors.
2. Make constructors private to make sure page object classes still have
full control over what can be used as their root elements.

# Tests

`npm run test`