Add Jest specific implementation of PageObjectElement (#2152)
# Motivation

We want to share page objects between Jest unit tests and Playwright e2e
tests.
So we provide an interface abstracting the element to use by page
objects.
This interface will have different implementations per platform.
This PR adds a first version of the Jest implementation.

# Changes


1. Add JestPageObjectElement and use it in all the tests to create page
objects.
2. Change PageObjectElement to an interface implemented by
JestPageObjectElement.
3. Add just a few methods on PageObjectElement to make current tests
pass.

In the future we will change the interface to be more async so that it
can be compatible with e2e but this will require more changes to the
tests and page objects to accommodate so it will be different PRs.

# Tests

npm run test
npm run check