Make the IC commit check fast (#2149)
# Motivation
Making CI faster.

Getting the did files from the e2e docker build is fast if we already
created that build for the e2e tests. And it makes sense to do it that
way to make sure that the did files match the canisters we test against.
But we no longer use those builds for e2e tests, so now we have the
did->rust conversion taking 16-20m because of the NNS canister build:

![Screenshot from 2023-03-22
11-18-25](https://user-images.githubusercontent.com/5982633/226893078-34110d2a-0305-4c9a-81f9-b36eeb62d03f.png)

# Changes
- Get the NNS and SNS did files from github.

# Tests
* See CI - it will break quite spectacularly if the did file URLs are
wrong, for example, and they should request rust changes if the did
files don't match the previous values.

* I changed the ic commit to verify that this does indeed make changes
when the upstream did files change, and then changed it back.

* Regarding speed, the "IC Commit" check that took 16m in the screenshot
above now takes 6s:

![Screenshot from 2023-03-22
13-28-16](https://user-images.githubusercontent.com/5982633/226905074-91fdbee4-a679-4b88-bdc0-323e3b707689.png)

* I verified that the logs of the step look plausible, because that is
quite fast.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>