Use cdk for init and upgrade hooks (#2166)
# Motivation
We wish to pass init arguments to the nns-dapp canister. This requires
the latest cdk. However there are e2e test failures when using the
latest CDK macros on some of the API methods. The test failures look
like flakes, maybe timing is a bit different or something like that as
the functionality being tested seems to work fine, but tracking down
this strange e2e test result is taking a lot of time and delaying the
work on arguments. And for the arguments we need to update only the
macros used in the init and upgrade hooks.

So let's upgrade just those hooks first.

# Changes
* Add the latest cdk macros.
* Use the latest cdk macros for init, pre_upgrade and post_upgrade.

# Tests
See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>