feat: get BTC address once Receive modal displayed (#2165)
# Motivation

In ckBTC, the default segment is now "ckBTC" and not "Bitcoin" anymore.
Therefore, instead of loading the BTC address before opening the modal,
we can fetch it once the modal is displayed.

# Changes

- move the "get address" logic from the caller to the on mount of the
modal
- display a spinner while loading (not the prettiest but the ultimate
goal is to not make a call to derive the address anyway)

# Screenshots

Loading (rest of the modal does not change):

<img width="1536" alt="Capture d’écran 2023-03-22 à 18 52 28"
src="https://user-images.githubusercontent.com/16886711/226998480-5c07df7e-e7de-4605-9e98-f152f0af30c4.png">