Make presence checking on page objects async (#2159)
# Motivation

We want to reuse page objects for e2e tests.
Playwright uses "Locator"s which refer to specific elements.
But to be able to chain them synchronously (such set
`getTab().getCard().getButton()`), they might refer to elements that
don't exist. And you can only find out if they exist via a synchronous
call.
For compatibility we change the Jest page objects to work the same.

# Changes

1. `JestPageObjectElement` may now hold `null` elements but its
`querySelector()` will never return `null`. Instead it might return a
`JestPageObjectElement` with a null element.
2. Checking if a `JestPageObjectElement` refers to an element that
exists is now done async via `isPresent()`.
3. `JestPageObjectElement.querySelectorAll()` is now async. It can't be
deferred because there is no way to know the size of the list in
advance.
4. `PageObjectElement.getText()` now returns `null` if the object
doesn't exist.
5. The `under` static function on page object classes now always returns
non-null but it might refer to an element that turns out not to exist
when checked for presence.
6. The `allUnder` static function on page object classes is now async.
It can't be deferred because there is no way to know the size of the
list in advance.
7. Add an `isPresent()` method to `BasePageObject` so every page object
can be checked for presence, which would have previously been done with
`nonNullish()`.
8. If a page object has an `isContentLoaded()` function that checks if a
list of skeletons is empty, make sure that the page object itself is
present. Otherwise the list of skeletons will be empty while the object
itself isn't even there and certainly not loaded.
9. Add necessary awaits on presence checks.

# Tests

`npm run test`
`npm run check`

---------

Co-authored-by: Llorenç <llorenc.muntaner@dfinity.org>