GIX-1336: Refactor restoring participation flow to project detail (#2112)
# Motivation

Move logic to restore participation from the button component to the
page component.

There is no new functionality in this PR. This is just refactoring.

# Changes

Main changes
* Move logic to restore participation from ParticipateButton to
ProjectDetail.
* Refactor and organize the code in ProjectDetail.svelte to make it more
understandable.
* It's worth taking a look at the ProjectDetail.svelte alone instead of
the changes to better understand the flow. I'd like to simplify a bit
this, but that will come in another PR.

# Tests

* In ParticipateButton, I removed all the tests regarding the restoring
participation flow.
* In ProjectDetail page, I mocked apis instead of service and adapted
tests to it.
* Main and biggest test: make sure that the commitment is updated if
there is an open ticket in ProjectDetail page.