GIX-1220: SNS proposal - Render system info (#2171)
# Motivation

User should be able to see the metadata of an SNS proposal.

# Changes

Main change:
* Implement SnsProposalSystemInfoSection.svelte.

Minor changes:
* New component SnsProposerEntry.svelte to handle the logic of the
proposal key value. Not now, it will be in another PR, but that will
open the neuron modal if the user is logged in.
* Add new prop `className` to the Hash.svelte component.
* Minor changes in the SnsProposalDetail.svelte to add `rootCanisterId`
in the scope so that it can be passed to SnsProposalSystemInfoSection.
No change in functionality. The other component that will need it is the
SnsProposalVotingSection.

# Tests

* New scenario in SnsProposalDetail that tests presence of the metadata.
* Test all the functionality of SnsProposerEntry
* New POs: KeyValuePairPo and SnsProposalSystemInfoSectionPo