fix: settings title and visual quirks (#2214)
# Motivation

This PR fixes two things:

1. The "Settings" title should be set in the header because if you
navigate from a page that sets a title (not the Accounts page) the title
in the header remains the one of the previous page

2. When navigating there is a brief moment where both previous and new
pages are in the DOM. The layout being what the design requires, when
navigating from a page that display universe to a page with island,
there could be a brief moment where the title store is updated and
therefore the title might be displayed twice.

# Notes

Regarding 2. I did not find better solution for now. It does not solve
the issue but, at least workaround to not display twice the title.

Note against better solution at all!

# Changes

- display "Settings" title in header
- add a CSS rule to hide one of the title if two spit panes are rendered
in the DOM

Not related:

- add style `value` to display principal on settings page

# Screenshots of the issues

<img width="1536" alt="Capture d’écran 2023-03-30 à 20 26 33"
src="https://user-images.githubusercontent.com/16886711/228931682-16ac15e9-b87c-4e15-a1e9-a3093a2ed94f.png">


![demo](https://user-images.githubusercontent.com/16886711/228931739-fa9aa44d-a63a-47ff-88c4-1b1ed8ac5467.gif)

# Screenshots with fix/workaround


![fix](https://user-images.githubusercontent.com/16886711/228932137-3aa94176-3ee2-4924-a0e1-cdafc971db46.gif)

---------

Co-authored-by: mstrasinskis <98811342+mstrasinskis@users.noreply.github.com>