Feat: Use only type for SNS proposals, no topic (#2271)
# Motivation

Align NNS Dapp with the agreements reached within the team on how to
deal with Type and Topic for SNS proposals.

Result:
* Topic does not exist yet for SNS proposals.
* The type is the name of the nervous function.

# Changes

* Change mapProposalInfo in sns proposal utils to return the type and
typeDescription instead of topic and topicDescription.
* Do not render topic and type items in the
SnsProposalSystemInfoSection. Render only the type item.
* Render the type as title in the SnsProposalSystemInfoSection
component.

# Tests

* Adapt tests to new functionality and also change the test descriptions
to match the changes.
* Remove getTopicText from SnsProposalSystemInfoSection page object.
* Change testId in SnsProposalCard and add it to the type. This is used
to test that the nervous functions are properly loaded in the
SnsProposals component test.