Populate CSP and Robots in the canister (#2267)
# Motivation
We need to populate the robots meta tag and parts of the CSP in the
canister, based on arguments.

Given that this is mostly a matter of substituting argument values in
the index.html, it seems sensible to make a general rule for replacing,
so here it is:

* `${{FOO}}` is replaced by the value of argument keyed by FOO, for any
all-caps name FOO.
* `<!-- FOO -->` is replaced similarly.

# Changes
* Add a simple substitution engine to the nns-dapp.
* Define the`ROBOTS` tag as part of the config, to keep complexity out
of the canister.
* Remove the robots injection code from the frontend; the comment tag
will be replaced by the canister instead.
* Add variable strings for the CSP, for the canister to translate.

# Tests
Tests are with this PR combined with #2230 , as otherwise correct
behaviour can be due to the frontend code populating the correct values
and this can hide a bug in this code.

First a stock network is created from an snsdemo snapshot:
```
dfx-snapshot-restore
```
Then this script was used to combine the two PRs, build and deploy:
```
$ cat ,make-it-so 
#!/usr/bin/env bash
set -euxo pipefail
git reset --hard canister-csp
git merge origin/GIX-1440_LM_nns-dapp-uses-inserted-arguments
dfx build nns-dapp
dfx canister install nns-dapp --argument "$(cat nns-dapp-arg.did)" --wasm nns-dapp.wasm  --mode upgrade
```

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>