Wait for aggregator (#2306)
# Motivation
Apparently e2e tests are flaking out because the aggregator doesn't have
enough data when the e2e tests run.

# Changes
* Configure the aggregator collect data quickly.
* Note: The local replica has an artificial delay to make timing
realistic. In the next dfx that delay is configurable, so it can be made
short for fast CI tests or it can be made long to simulate usage in a
stressed environment. This would help here as well, once we upgrade.
*  Wait for the aggregator to have 12 SNS before running e2e tests

# Tests
See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>