Log pre-upgrade cycle count (#2335)
# Motivation
We are concerned that the pre-upgrade cycle count may be too high and
may cause the canister to fail to upgrade.

Storing pre-upgrade cycle counts in the stats structure is a bit tricky,
as data would have to be stored in stable memory and the cycles are
consumed by storing data in stable memory. This can be solved if we
upgrade the memory structure (extremely carefully!) to the new stable
memory library. That also supports stable structures, which makes this
problem go away.

In the meantime, a simple solution is to print out the instruction
counter counter and look at the replica logs; we run one of the servers
on the nns subnet so have access to logs there. Longer term, logs will
be available to all canisters via a ring buffer but that is work in
progress.

# Changes
- Print the instruction count at the beginning and end of the
`pre_upgrade` hook.

# Tests

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>