Add testId to ProjectCard (#2363)
# Motivation

I'm adding a testId to `ProjectCard` in
https://github.com/dfinity/nns-dapp/pull/2361 but it overrides the
testId of the `Card` component, which causes an existing test to fail.
Because the test is unrelated to that PR, I'm updating it in this
separate PR.

# Changes

1. Add testId on `Card` in
`frontend/src/lib/components/launchpad/ProjectCard.svelte`.
2. Update `frontend/src/tests/lib/components/launchpad/Projects.spec.ts`
to use that testId instead of the default `"card"` testId.

# Tests

`npm run test`