Build only one wasm (#2369)
# Motivation
The wasm build no longer depends on network, so there is no need to
build it for two networks.

The only difference between the docker build outputs is whether the
arguments are prepared for mainnet or localhost. The mainnet arguments
can be used for release verification. If we really need the localhost
arguments we can generate them by running ./config.sh.

# Changes
* run the docker build for mainnet only.

# Tests
See CI

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>