GIX-1505: Fix prod preview (#2377)
# Motivation

Fix `npm run preview` which runs a production-like environment.

# Changes

Main change:
* Logic inside `getEnvVars`. Force to use the build env vars if the port
is set. This is the case for local dev server and preview.

Minor changes:
* Add `raw` to the SNS aggregator url when running from localhost as
well. This is skipped if the `url` of the aggregator already includes
`localhost`.
* New env util `isLocalhost`.

# Tests

* Test new env util.

I tested the three ways:
* Deploy to testnet
* Local dev server
* Preview

# Docs

Add a comment on how to run `npm run preview`.