Fix nightly tagging (#2375)
# Motivation
Triggering builds with tags is unreliable. Presumably there is a race
condition between adding a tag to git and triggering the run, so the
list of tags known to git may or may not include the tag that triggered
an event.

# Changes
- Check whether a docker-build workflow was triggered by a tag. If so,
add the tag to the list of tags for this commit, if it is not already
there.

# Tests
Only time will tell for sure. But pushing a tag does show the tag being
identified in the github workflow.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>