Chore: Derived store with neurons with distributed maturity (#2393)
# Motivation

We show to the user when was the last maturity distributed.

Yet, that doesn't mean the neuron received any maturity on that date.

We want to show to the user whether the specific neuron received or not
maturity on the latest distribution date.

This is the first of two PRs.

# Changes

* New derived store `neuronsLastDistributedMaturityStore` that returns
the neuronIds of the neurons that participated in the last reward event.

# Tests

* Test new derived store.