Chore: Add exchange rate proposal (#2431)
# Motivation

Add a dummy proposal to create Exchange Rate proposals.

There is an issue with those proposals where the user is always able to
vote, even after voting.

The problem seems in the data coming from the backend.

This PR will help us debug the issue faster by being able to create the
proposal responsible for the problem.

# Changes

* Add a new dummy proposal to the list of dummy proposals for NNS.

# Tests

Not applicable