Check ckBTC minter balance and display call to action (#2430)
# Motivation

On ckBTC accounts page, once per session, check the minter canister
balance for the user withdrawal account.

Display a related button in form of a card with an information if data
are loading or loaded.

# Changes

- add new `AccountType` with the type of account: `withdrawalAccount`
- new component `CkBTCWithdrawalAccount` (without meaningful content
yet)
- new function `getCkBTCWithdrawalAccount` that loads withdrawal account
and account and handles case in which one of the other should be skipped
and values from store should be used instead
- new `loadCkBTCWithdrawalAccount` for query and update loading
- new store `ckBTCWithdrawalAccountsStore`