GIX-1546: Integrate geo location service (#2447)
# Motivation

SNSs projects might want to restrict participation in some countries.

To do so, the NNS Dapp needs to get the location of the user and compare
that with the restriction list.

In this PR: I setup the service and store that will have the user data

# Changes

* New location api. It returns a hardcoded value for now. Once we know
which service we need to use, we can implement it.
* New locationStore that has the location of the user as a country code.
* New location services that uses the api to load the store.
* New type `CountryCode` that we might want to improve. We might need to
translate the country code to human readable and this type might help
us. For now, this is just a string.

# Tests

* Test the new location service.
* Test the new location store.