Script to start dfx with snapshot (#2455)
# Motivation

We have `scripts/dev-local-state.sh` which starts a replica (`dfx
start`) *and* the frontend (`npm run dev`) with a given snapshot. I
originally thought this would be convenient but it hasn't been for 2
reasons:
1. `npm ci` often causes the vite server to kill itself. This then makes
the script also kill the replica, which is unnecessary.
2. Some scripts (e.g. in snsdemo) run a replica and start by killing any
existing running replicas. Because `dev-local-state.sh` runs the replica
in the background, it doesn't notice that it's killed and then it
doesn't get an opportunity to restore the state.

# Changes

Add `scripts/dfx-snapshot-start` which is just like
[dev-local-state.sh](https://github.com/dfinity/nns-dapp/blob/main/scripts/dev-local-state.sh)
but it runs `dfx start` in the foreground instead of the background, and
doesn't run `npm run dev`.



# Tests

Ran the script manually.