feat: Add ic-wasm to the toolchain (#2497)
# Motivation
We would like to include the git commit and API did in the nns-dapp and
sns_aggregator wasm metadata sections. The ic-wasm tool is currently the
preferred method for adding custom sections.

Unfortunately its pre-built binaries do not work on the standard
production linux (Ubuntu 20.04), due to the same issue that plagued a
number of other ic projects in recent months. So we have to fall back to
cargo install, which works reliably but is slow. CI will similarly be
slow until ic-wasm is in the main branch docker cache. So let's get it
there soon.

# Changes
- Add ic-wasm to the toolchain

# TODO
- Use ic-wasm to add the desired custom sections.

# Tests
I have used ic-wasm locally to add custom sections

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>