Delay ckBTC account and transactions reload (#2532)
# Motivation

The `index` canister needs 30sec or so and soon less to populate new
transaction but still, takes a bit of time. That's why after a "ckBTC
refresh balance" currently only the balance is updated and not the list
of transactions.

In the future we will fetch the list of transactions and balance from
the index but, it isn't there yet. Likewise we plan to implement some
crontab refresh but it's tasky.

That's why this PR just adds a bit of delay.

# Changes

- delay call of `reload` function in `updateBalance` of few seconds

---------

Co-authored-by: mstrasinskis <maksims.strasinskis@dfinity.org>
Co-authored-by: mstrasinskis <98811342+mstrasinskis@users.noreply.github.com>