Test: Add Participation Test in Unit Tests (#2542)
# Motivation

Improve testing.

To easily test the participation flow under different scenarios, it's
better if we can use the same page object used in e2e for unit tests.

In this PR: Add a test to ProjectDetail.svelte using the ProjectDetailPo
to participate in a swap.

# Changes

Main change:
* Add new test scenario in ProjectDetail.svelte for a user to go through
the participation flow. Use same as in e2e test.

Minor changes:
* Add a couple of `waitFor` in the jest object before performing some
actions.

# Tests

Only changes in tests.