Bump snsdemo commit in workflows (#2583)
# Motivation

`dfx` often fails to install on CI because it gets the install script
from a canister which updates often.
The script that downloads the install script comes from snsdemo and has
been changed to retry for longer.

# Changes

Use a newer snsdemo commit for `actions/checkout@v3` in CI workflows.

# Tests

CI