Remove token from "Account" and replace with e8s value (#2594)
# Motivation

We aim to integrate ICP with ICRC-1 standard. When we introduced a store
for the tokens, we did set the duplicate "token" information in
`Account` as a deprecated field. This PR uses the `tokensStore` or the
`ICPToken` constant for Nns and resolve the deprecation by using a
balance set as e8s value (as provided by the API).

That way, in future PRs, we will be able to remove some duplicate calls
in ckBTC and Sns that are use to fetch the token twice.

# Related PRs

Few related PRs were needed to implement this refactoring.

- [x] #2596
- [x] #2597
- [x] #2598
- [x] #2599

# Changes

- main change: in `Account` change `balance` from a `Token` to a
`bigint`
- adapt code base and test accordingly