Re-enable Firefox on Playwright tests (#2609)
# Motivation

We disabled Firefox tests because the tests were flaky.
https://github.com/dfinity/nns-dapp/pull/2445 should have fixed most or
all of the flakiness.
If the tests are flaky again, I'll investigate and disable if necessary.

# Changes

1. Enable Firefox in `frontend/playwright.config.ts`
2. Increase the timeout on CI because the tests are close to the limit.
Lower the timeout on non-CI because CI machines are really quite slow
compared to our development machines.
3. Set the number of workers for non-CI to 6 so all (now 6) tests can
run in parallel by default. Apparently `undefined` defaults to 5
workers.

# Tests

`npm run test-e2e`