Test that the mainnet config is not accidentally changed (#2553)
# Motivation
When changing dfx.json or config.sh there is a risk of accidentally
changing the mainnet config, causing a broken deployment. That is
probably the biggest risk of changing the config.sh. Having a reference
for what these files are expected to contain gives some assurance that
the values are not changed accidentally.

# Changes
- Add a stability test for the frontend args and deploy args.

Note: In theory didc could break, causing an invalid binary argument
even when the text argument is correct. That possibility is not covered
by this test.

# Tests
This PR is a test, exercised in CI.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>