Use standard arg parsing for docker-build (#2636)
# Motivation
Consistency. I would like to add another flag but feel as if I am
reinventing the wheel when modifying the docker-build arguments. So why
not use the standard argument parser in docker-build?

# Changes
- Use standard argument parsing in `docker-build`
- Surface the `--verbose` flag in the `clap.bash` environment variables,
so that it can be used to make the docker output verbose, as before.

# Tests
See CI.  I have also used it locally.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>