One wasm to rule them all (#2635)
# Motivation
The wasm we have on mainnet should be identical to the one used on local
networks. Let's test that.

# Changes
- Add a test that compares wasm builds for mainnet and local.

# Notes
- There is a separate issue of whether all parts of the code are using
the dynamic configuration passed in with arguments. That is not covered
in this PR.
- There is a question of where the test should live. I am personally of
the opinion that we should build the wasm once, then use that one wasm
across many tests. Why? Why waste CPU cycles repeating work? That means
that we should not have separate e2e and testing workflows but rather
one workflow with separate jobs inside it.

# Tests
See above.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>