Page objects in MergeNeuronsModal.spec.ts (#2649)
# Motivation

I'm planning to make changes to `MergeNeuronsModal.svelte` so cleaning
up the test so it's easier to add tests later.
In particular, migrating the test to using page objects.

# Changes

1. Added several page objects.
2. Also added `CardPo` which serves as a base class for other card page
objects. And use it for new `NeuronCardContainerPo` and existing
`ProjectCardPo` which is otherwise unrelated to this PR.
3. Added test IDs to make components accessible to page objects.
4. Change
`frontend/src/tests/lib/modals/neurons/MergeNeuronsModal.spec.ts` to use
page objects.
5. Also use a top-level `beforeEach` instead of a nested `afterEach`.
6. And fixed an issue in `JestPageObjectElement.isPreset()` because it
only worked for descendant elements. `:scope` is a selector referring to
the element itself but apparently only works when followed by descendant
selectors.

# Tests

`npm run test`

Manually checked that the modified components look normal.