Fix Sns proposal reload with correct ID but NNS as universe (#2663)
# Motivation

When navigating from a Sns proposal detail to a Nns universe, such as
for example the launchpad, the reload feature is triggered again with
the same Id but with the a universe Id extracted from the route that is
Nns.

# Changes

- add a check to load the Sns proposal only if not Nns
- added a `TODO` to use another source of information to load the
universe

# Screenshot

I found the issue with a breakpoint in the browser.

<img width="1536" alt="Capture d’écran 2023-06-08 à 07 33 01"
src="https://github.com/dfinity/nns-dapp/assets/16886711/3ab1820c-3771-4c24-968e-57bdd68f6e3d">