Create a generic timer (scheduler) for web workers (#2651)
# Motivation

In `cycles.worker` we have implemented a scheduler that loads the
identity and performs a repetitive task.
We are going to use this pattern in PR #2639 as well and probably future
web workers.

That is why this PR extract the existing code to a generic timer that
can be re-used to implement new workers.

# Note

This PR does not introduce new logic or code except modifying the
variable `syncInProgress` from a `boolean` to typed `string`. For the
rest, it refactored existing pattern and code into a new reusable code.

# Changes

- extract timer and sync logic from `cycles.worker` to create
`worker.timer`
- move `loadIdentity` from `cycles.worker` to `auth.utils`