Use fake governance API in MergeNeuronsModal.spec.ts (#2671)
# Motivation

I'll be making changes to this test when implementing the simulate
merging neurons UI so now seems like a good time to change it from
mocking the service and store to using a fake.

# Changes

In `frontend/src/tests/fakes/governance-api.fake.ts`:
1. Add `mergeNeurons`. It only adds the stakes together in the target
neuron and sets the stake to 0 in the source neuron. This is very
simplistic but should be good enough to verify in the test that neurons
have been merged.
2. Add `getNeuron({identity, neuronId})` for use in the fake itself as
well as in tests for convenience.
3. Allow setting the `stake` on neurons in the fake.
4. Add `addNeurons` to insert multiple fake neurons with a single call.

In `frontend/src/tests/lib/modals/neurons/MergeNeuronsModal.spec.ts`:
1. Create and use a test identity to make sure the authenticated
identity, the controller of the neurons and the principal of the main
account all align.
2. Insert neurons into the fake API instead of in the store.
3. Verify the result of the mergeNeurons call instead of only expecting
`mergeNeurons` to be called on the service.


# Tests

Test only