Fix Sns proposal lazy loading (#2664)
# Motivation

The Sns proposals list ends in an infinite loop. The root cause of the
issue is the fact that we paginate the list of proposals but fails at
providing the correct index to search next proposals - i.e. the UI ends
providing always the same index and therefore the query is repeated over
and over again.

Worth to note that the issue happens when filters are selected - i.e.
when only a subset of the proposals are displayed.

# Changes

- resolve index (`beforeProposalId`) for the next query with list of all
Sns proposals that have been queried for a universe instead of using the
derived store that contains only those that are displayed

# Screenshots

Bug


![bug](https://github.com/dfinity/nns-dapp/assets/16886711/02e8cd24-5beb-45f9-af29-6effb8bb38c5)

Fix. Note the brief glitch in the screenshot is expected with current
UI. When fetching a spinner/skeleton is briefly displayed therefore
until the pagination is over, these glitchs appears. This PR does not
improve the UI.


![fix](https://github.com/dfinity/nns-dapp/assets/16886711/3f9d4622-8677-4ef0-a221-d0a552e970b2)