Use binstall (#2675)
# Motivation
`cargo install` builds from source, making it slow. So we pre-build many
tools and download the pre-built binaries. Now we have the problem that
we have a bunch of scripts to find available versions and install them
by curl. `cargo binstall` does that for us. For the versions, it uses
the list of versions published as rust crates. And it installs pre-built
binaries, if available, else builds from source.

Pros:
- Using `binstall` saves us from having to write a bunch of custom
curl-based installers and version finders.

Cons:
- We have to install `binstall`
- We have to put our pre-built executables at standardized locations.
- I have done this for `idl2json`, so `cargo binstall idl2json_cli`
works. It isn't that hard.
  - I have made a PR to support binstall for `ic-wasm`.

# Changes
- Change every `cargo install` to `cargo binstall --no-confirm`

# Tests
See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>