Fix some setup issues on a new M1 installation (#2672)
# Motivation

I tried to build outside of Docker on my M1 and required a few
interventions. This is only a partial list of the issues I discovered
with @dskloetd help, but I'll make another PR once I test some more.

# Changes

- `sns` binary url for Darwin was wrong
- `realpath` on my M1 exists on `/bin` so `./setup` didn't `brew install
coreutils`, so I changed the condition to `sha256sum`

# Tests

Manual tests