Fix: Use ballot voting power in SNS proposals (#2683)
# Motivation

There was a bug where the voting power shows in the SNS proposals
matched the voting power of the neuron, instead of the voting power of
the ballot.

Voting power of a neuron for a proposal is set when the proposal is
created. If the voting power of the neuron changes afterwards, it
doesn't affect the voting power for earlier proposals.

# Changes

* New sns proposal util `ballotVotingPower` to get the voting power of a
neuron for a specific proposal.
* Change the voting power calculation in the sns vote registration
service `proposalAfterVote`.
* Change the voting power calculation in the sns neuron util
`votedSnsNeuronDetails`.
* Change the voting power calculation in the sns proposal util
`snsNeuronToVotingNeuron`.
* Remove the unused property `parameters` in the changed functions and
in the dependencies where they are called.

# Tests

* Add two new cases in SnsVotingCard spec file to check that voting
power comes from ballots.
* Edit the tests for the utils after change in calculation.
* New test for `ballotVotingPower` util.