Import canister url command (#2627)
# Motivation
The snsdemo repo has a command for getting canister URLs. It is useful
and can replace some of the logic in config.sh with something more
maintainable.

Note:
* We could use `dfx-canister-url` directly from the snsdemo repo,
however that assumes that we have the snsdemo repo bin in the path every
time we call `config.sh` and that is not the case and can not trivially
be made so.
* I have proposed putting the polyfill upstream in dfx proper, however
that might take a while.
https://app.slack.com/client/T43F9UHS5/threads/thread/CGA566TPV-1685628863.824769

# Changes
- Import `dfx-canister-url`
- Tweak due to slightly different linter settings.

# Tests
The command is tested upstream

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>