Feat: Change functionality empty NNS Proposal filters (#2691)
# Motivation

Bring consistency in proposal filters.

Inconsistency: SNS filters showed all proposals if no filter was
selected. Whereas NNS had a hack coming from flutter to not render any
proposal if a filter was empty.

Solution: Along with product, we agreed to remove the legacy hack and
consider no filter selected as "any". This is also the behavior of the
backend endpoint.

# Changes

* Remove hack in `findProposals` proposal service that would set an
empty store if a filter was empty.
* Change the behavior of the queryProposals api when the topics filter
is empty.
* Change behavior of the `matchFilters` helper inside the proposal
utils.

# Tests

* Add a test case for the `queryProposals` api new logic.
* Change the test case for the findProposals service.
* Change the test case for the `hide` proposal util.