Faster wasm network independence test (#2699)
# Motivation
We have a test that checks that the wasm is the same, regardless of
which network we are building for. This test is commonly the last to
finish in CI:

![Screenshot from 2023-06-15
09-43-37](https://github.com/dfinity/nns-dapp/assets/5982633/5d3b2eb4-97f2-4ad8-acdd-a8f759a9b123)

It would be nice if CI were to finish sooner.

# Changes
* That the wasm does not depend on the network is guaranteed by the fact
that the network is not an input to any of the wasm build steps in the
Dockerfile. As such, it probably suffices to perform the check only if
the dockerfile has changed.
* If the test needs to be performed, use a larger runner.

# Tests
* See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>