Feat: Store data utils and type (#2701)
# Motivation

We want stores to hold the error, loading and success states.

This PR: Introduces a type that should be used by the stores and a
helper to convert the type to the success type.

The type is already used in the useCountryStore.

The util will be used in the util from PR #2700
[swapSaleBuyerCount](https://github.com/dfinity/nns-dapp/pull/2700/files#diff-aa44a48b94ac01459ed5ebb708a264a70bbd0b1c1afc4616da02d00053131738R4)

# Changes

* New type `StoreData`.
* Use new type in `userCountryStore`.
* New store util `isStoreData`. Not used yet, only tested.

# Tests

* New test file for new store util `isStoreData`.