Utils to extend JSON stringify and parse (#2719)
# Motivation

In #2639 we will pass complex objects back and forth between the UI and
web workers. While the browser can serialize most data, their some types
that are not supported by the browser (`bigint` or `Uint8Array`) and
some that can be serialized but because they are ojbect, looses their
functions during the process (for example `Principal.getText()`).

To solve the serialization issue we can stringify and parse objects when
we exchange data between UI and web workers and provide special replacer
and reviver.

# Changes

- new `jsonReplacer` and `jsonReviver`
- add a TODO / question for an existing replacer