Fix: Deploy to testnet script (#2705)
# Motivation

Script to deploy to testnets wasn't working `./deploy.sh`.

The problem was that dfx didn't know that the wasm was gzipped.

Solution: rename the wasm in dfx.json

More about the discussion:
https://dfinity.slack.com/archives/CGA566TPV/p1686829378681289

# Changes
* [Commit
1](https://github.com/dfinity/nns-dapp/pull/2705/commits/88360aadbb3a180088b639db97b3aa24067bffe4):
  * Rename nns-dapp wasm in `dfx.json`.
  * Add new file extension to git ignore.
*
[Commit2](https://github.com/dfinity/nns-dapp/pull/2705/commits/a270f06f5bec6aa049efdfa8462b45a1ac3251eb):
Make build-rs.sh emit .wasm.gz instead of .wasm (tested manually, it
works)
* [Commit
3](https://github.com/dfinity/nns-dapp/pull/2705/commits/f10ef1d4fe80e182a5774b209f9552fe4561dd68):
Update the dockerfile so that it too emits .wasm.gz (tested manually, it
works)
* [Commit
4](https://github.com/dfinity/nns-dapp/pull/2705/commits/5ec7bf79c3b6949e98d1094b4c453de346f5b690):
Bulk rename .wasm to .wasm.gz, but done manually checking every case as
there are quite a few exceptions. E.g. external .wasm files are not
renamed and jq paths ending in .wasm must not be renamed. (tested by CI;
this verifies that users of the above changes within this codebase have
been updated)
* [Commit
5](https://github.com/dfinity/nns-dapp/pull/2705/commits/199081aeee03d6ef6060a302254e5df13432a55e):
For backwards compatibility, include a copy of all .wasm.gz assets with
just the .wasm suffix.
* Non-PR change: For recent releases, I have copied every .wasm in the
release to the same path with suffix .wasm.gz. See the comment in the MR
for details. This is to ensure that tasks such as rollback/rollforward
tests have access to .wasm.gz for existing releases. This may have to be
repeated to cover any more releases created between now and the PR being
merged.

# Tests
* There should be no change in functionality; existing CI should suffice
for canister functionality.
* Here is a test release, to verify that both .wasm.gz and .wasm are
produced:
https://github.com/dfinity/nns-dapp/releases/tag/untagged-29888d6b4b9ee251e9b5
* Here are the last releases by proposal, to verify that .wasm.gz is
available to rollback tests:
  * https://github.com/dfinity/nns-dapp/releases/tag/proposal-123011-agg
  * https://github.com/dfinity/nns-dapp/releases/tag/proposal-123006
  * https://github.com/dfinity/nns-dapp/releases/tag/proposal-122799

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>