Prune nightly (#2735)
# Motivation
We start having a lot of nightly releases. It probably doesn't make
sense to keep these around forever.

Note: I am a bit leery of having a script delete releases automatically.
As long as it works, great, but what if there is a bug? I would feel
more comfortable having a human in the loop for every deletion.

# Changes
* Add a command to delete old nightly builds.
* Note: The command asks for confirmation before deleting a release.
This _can_ be overridden with `--yes` but I personally would recommend
against it.

# Tests
I have used this to delete old nightly builds from March and April.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>