Fix: Screen Title in HW when staking two neurons (#2741)
# Motivation

The user was getting "Send ICP" in the HW screen instead of "Stake
Neuron" if they were staking two neurons in a row without refreshing the
page.

Problem: The agent from `createAgent` is cache. Therefore it was using
the identity from the first staking. Yet, the service to stake a neuron
was getting a new identity and setting the flag to stake neuron in that
identity.

# Changes

* Cache the ledger identity.

# Tests

* Add a test to the ledger service to check that the identity is cached.