One build (#2744)
# Motivation
- Network errors are killing CI.
- The heaviest network usage is in building the canisters, either due to
downloading dependencies or from getting the docker cache.
- So let's build once and reuse.

# Changes
* Move the `docker-build` and `sns_aggregator` workflows into
`build.yaml`, where it can reuse the docker build.
* Add a workflow dispatch to `build.yaml`; this was available on
docker-build.

Note: I have studiously avoided making any "improvements" that are not
strictly required for this PR. Yes, to be sure, any reviewer is likely
to find bits that could be improved. But that is not the point of this
PR.

# TODO
- After this has been merged, remove the empty workflows as merge
requirements and make another PR deleting the workflows altogether.

# Tests
See how the docker build is now reused:

![Screenshot from 2023-06-22
17-43-22](https://github.com/dfinity/nns-dapp/assets/5982633/2176113f-d18b-46af-804b-68fb56ec75fd)

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>