Script to download wasm from CI (#2746)
# Motivation

We want to make the release a less manual process.
One of the steps in the release process is to download the WASM from CI.

# Changes

Add a script to download the nns-dapp WASM from CI.

# Tests

Ran the script, also with wrong commit and wrong filename to exercise
the error code paths.

Added a test which checks that the downloaded file is a gzip file.