GIX-1656: Rename Canister Endpoint (#2785)
# Motivation

Users want to be able to add names to their canisters.

We already enabled canisters to be created with a name, but we were
missing an endpoint to rename canisters.

This PR introduces the nns-dapp canister endpoint to rename canisters.

In another PR: Use this endpoint from the NNS Dapp frontend.

# Changes

* New endpoint `rename_canister` in NNS Dapp canister.
* New method `rename_canister` in AccountsStore to rename a canister.
* Related types.

# Tests

* Test the new method in the AccountsStore.

# Todos

Not yet entry in the changelog until I add all the renaming
functionality.

---------

Co-authored-by: Max <bitdivine@users.noreply.github.com>