Update CI workflow name used by ci-link (#2783)
# Motivation

https://github.com/dfinity/nns-dapp/pull/2744 changed workflow names.
`scripts/nns-dapp/ci-link` expects there to be a workflow named "Docker
build" but it no longer exists.

# Changes

Change `scripts/nns-dapp/ci-link` to use
`scripts/nns-dapp/get-ci-build-run` which is tested.
The only difference is that `ci-link` used to output multiple links and
now only 1.
But the only place where it's used, it was followed by `tail -n 1` so
this fine.

# Tests

None, but now most of the work is done by
`scripts/nns-dapp/get-ci-build-run`, which is used by
`scripts/nns-dapp/download-ci-wasm` and thus tested here:
https://github.com/dfinity/nns-dapp/blob/75936135f738040a3cdfeec73d626cb4e3b9eb46/.github/workflows/checks.yml#L271


# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max <bitdivine@users.noreply.github.com>