Improve reload ckBTC transactions UX (#2819)
# Motivation

When we integrated the observations of the ckBTC balances in PR #2810 we
noticed that in case of updates, the transactions were fully reloaded
which is UX wise suboptimal since those are are also observed.

This PR improves the UX by only reloading the list of transactions when
needed.

# Changes

- imperatively `reloadTransactions` when needed instead of reloading
each time the `account` state changes
- add a four seconds delay in reload to smoothness the UX on mainnet

# Test

I discovered that the test "should reload on close receive modal" was
actually already incorrect on main but, this change lead to make it
fail. I corrected it by asserting that the correct function
`loadCkBTCAccounts` is effectivelly called.

# Screenshots


![demo](https://github.com/dfinity/nns-dapp/assets/16886711/5a9aa770-ba05-47a4-acd9-54b6aeb8c325)