Script to reorganize CHANGELOG entries (#2823)
# Motivation

We now update the "Unreleased" section of `CHANGELOG-Nns-Dapp.md` with
each PR.
We continue doing this after we cut a release and only if the release is
successful do we add a new section to the change log.
At that point we have to split the entries that were added after the
release was cut from the entries that were added before.
This is a bit tedious to do manually so I made a script.

The script *dot not* remove empty sections. This still needs to be done
manually.

# Changes

`scripts/nns-dapp/split-changelog` which does:

1. Take the old changelog until the end of the Unreleased section.
2. Apply the change that have been made since, but add an extra '+' at
the start of new entries.
3. Remove the entries without the extra '+'.
4. Remove the extra '+' from the new entries.
5. Append the old changelog starting at the Unreleased section but
replace the title with 'Proposal XXXXXX'.
6. Instead of 'XXXXXX' we use the proposal number if we can derive it
from the tags on the commit.

# Tests

Added a test which executes the scripts with a specific changelog.

# Todos

- [x] Add entry to changelog (if necessary).