GIX-1679 Make past-changelog-test not required (#2827)
# Motivation

It's currently not possible to merge a PR without adding an entry to the
changelog.
I want to unblock merging PRs ASAP while I fix the test.

# Changes

Make `past-changelog-test` not required in
`.github/workflows/checks.yml`

# Tests

I should be able to merge this PR even though it doesn't change the
change log.

# Todos

Deliberately not.
- [ ] Add entry to changelog (if necessary).