Patch empty record (#2830)
# Motivation
The patch files applied when generating Rust from SNS .did files have
grown quite large and break easily. There are a few changes in the
automatic generation that can make the patch files significantly
smaller. Here is the second:

`didc` generates invalid Rust enum variants of the form: `Name{}` when
what is intended is `Name(EmptyRecord)`. More idiomatic would be `Name`
but that doesn't match the data that is sent on the wire. The patch
files contain entries to make this conversion.

# Changes
* Automate adding `EmptyRecord` in place of empty enum entries.

# Tests
- See CI. Observe that the generated Rust is unchanged but the patch
files have become shorter.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>