Candid principal (#2831)
# Motivation
The patch files applied when generating Rust from SNS .did files have
grown quite large and break easily. There are a few changes in the
automatic generation that can make the patch files significantly
smaller. Here is the third:

The generated rust files namespace `Principal` as `candid::Principal`.
This is done everywhere without exception. The namespacing seems to be
unnecessary, but removing it changes both the rust code and the patch
files simultaneously, which makes it hard to prove correctness. This is
therefore done in two steps. First we namespace automatically, making
the patch files smaller, then remove the namespacing later.

# Changes
* Automate namespacing `Principal` as `candid::Principal`

# Tests
- See CI. Observe that the generated Rust is unchanged but the patch
files have become shorter.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>