GIX-1677: Remove VITE_OWN_CANISTER_URL from e2e tests (#2833)
# Motivation

VITE_OWN_CANISTER_URL should not be used by the frontend code so we want
to remove it.
But it's also used in the Playwright config to determine where to
connect to.
We can use `scripts/dfx-canister-url` to determine where to connect
instead.

# Changes

1. In `frontend/playwright.config.ts` use environment variable
`PLAYWRIGHT_BASE_URL` instead of `VITE_OWN_CANISTER_URL`.
2. Remove `dotenv` from `frontend/playwright.config.ts` because it's no
longer used.
3. In `.github/workflows/build.yml` use `scripts/dfx-canister-url` to
set `PLAYWRIGHT_BASE_URL`.

# Tests

Yes

# Todos

- [x] Add entry to changelog (if necessary).