Use dfx snapshot GitHub action for aggregator_test (#2839)
# Motivation

We have a separate GitHub action which does all the steps to start a dfx
snapshot.
But the `aggregator_test` does all the individual steps.

# Changes

Use the action for the aggregator_test.

# Tests

CI

# Todos

This is in preparation to use the new snapshot which has the aggregator
preloaded.
I'll add an entry when for that PR.

- [ ] Add entry to changelog (if necessary).