Add a manual update trigger for didc and rust updates (#2846)
# Motivation
It would be useful to be able to trigger the didc and rust update
workflows manually rather than either by waiting for the next cron time
or by adding a temporary trigger in a toy PR.

# Changes
* Add a manual trigger, in the form of `workflow_dispatch`, for the rust
and didc update workflows.

# Tests
Workflow dispatch cannot be tested until it is merged in main, as the
button is the button in the UI is defined in main.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>