Patch remnants (#2851)
# Motivation
The patch files applied when generating Rust from SNS .did files have
grown quite large and break easily. There are a few changes that can
make the patch files significantly smaller. Here is the sixth, and
possibly even the last:

* The comment added by didc has changed; we can update it in our rust
code.
* Some of the remaining patch changes are not needed.

Note: One rename is, I suspect, not needed but as it is a semantic
change I have omitted it. The changes in this PR should make no
difference to the behaviour of this code. Just make it more readable.

# Changes
* Update the comment at the top of the generated files to match that
created by didc.
* Remove changes that don't matter, e.g. the order of entries in module
imports.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->

# Todos

- [ ] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>