Set transitions as |local to solve populating destroy (#2865)
# Motivation

Svelte transitions are executed by default in components regardless if
their parent components destroy those or not. This behavior is known as
the `|global` transitions.

This leads NNS dapp to various issues in which components are not
destroyed and therefore, there context remains active.

Until today we solved this by applying a workaround in which we
destroyed the all layout in case of navigation. While it solved the
issue, this had for side effect to be a bit unpleasant for the eyes as
the all page layout disappeared on navigation.

By setting all the animation to `|local` we can solve the issue and make
the navigation more fluid.

# References

- documentation:
https://svelte.dev/docs/v4-migration-guide#transitions-are-local-by-default
- svelte issue: https://github.com/sveltejs/svelte/issues/6686

# PRs

- [x] https://github.com/dfinity/gix-components/pull/246

# Previous issues

- #1741
- #2861

# Changes

- update gix-cmp to set `|local` to any Svelte transitions
- set `|local` to the Svelte animations implemented in NNS dapp
- remove detail page layout navigating guard

# Notes

In the login page we will continue to observe `$navigating` because this
condition was not only used to solve the issue but, also to present a
spinner while the application is loading.