Highlight the wasm size in CI (#2870)
# Motivation
It would be useful to have better visibility of important statistics
such as wasm size and upgrade cycle count.

Let's start with wasm size, as it's easier. There is already a CI check
but the check is effectively invisible until we hit the limit and it
fails.

# Changes
* Print the nns-dapp wasm size as part of CI runs.

# Tests
If you select the build & test workflow in the checks for this PR, you
will see this, at the time of writing:

![Screenshot from 2023-07-12
14-44-54](https://github.com/dfinity/nns-dapp/assets/5982633/a68b62de-2495-487a-bbf3-455233ed5b25)

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>