Print the post-upgrade cycle count (#2872)
# Motivation
We would like better visibility of the cycle count on upgrade.

# Changes
* Add a command to estimate the post_upgrade cycle count based on
mainnet and local stats.
* Display the post-upgrade cycle estimate.
* Make the labeling and logic around the downgrade-upgrade tests cleaner
and more consistent.

# Tests
The build and test workflow output includes this:

![Screenshot from 2023-07-13
12-48-49](https://github.com/dfinity/nns-dapp/assets/5982633/f9f1b731-6242-4eba-8793-a741b153f3ec)

In addition, detailed statistics are printed in the workflow output
itself.

Note: The stats are not as I remember, so I have checked them in detail
and will be doing more profiling in the coming days.

# Todos

- [x] Add entry to changelog (if necessary).

# Future

- Add pre-upgrade cycle counts

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>