Monitor memory usage (#2882)
# Motivation
We would like to have visibility into memory usage.

# Changes
* Add stable and wasm memory to the exported metrics.
* Note: We do not make much use of stable memory yet, except in
upgrades, so this will currently show the stable memory populated by the
last `pre_upgrade` hook. However we plan to use stable memory more in
future.
* Note: Given that memory size thresholds are in GiB, I have provided
this data as MiB so that it is easy to read at a glance without
calculation. The stats API cannot do the same as the stats do not
support f64, as it is does not support Eq.
* Add memory logging to the `pre_upgrade` hook, so we have the very
latest data available if an upgrade goes wrong.
* Add wasm as the target in for `cargo build`, `cargo check` (covered by
`.cargo/config -> build`) and `cargo clippy`, as the wasm memory size
code compiles only when wasm is the target.

# Tests
* A test that all expected metrics are present is included in this PR.
It does not test that the values are plausible or e.g. how much memory
increases with each user.
* The `get_stats` API call is exercised in `estimate-upgrade-cycles`.
This PR also re-enables the stats calls made during the
downgrade-upgrade test. No use is made of the data though, so this just
confirms that the API call works, not that the data is correct.
* This branch has been deployed to an app subnet. The metrics show the
heap size: https://xnjld-hqaaa-aaaal-qb56q-cai.raw.icp0.io/metrics

# Todos

- [x] Add entry to changelog (if necessary).

# Future
- Estimates of memory usage if a given wasm is deployed to production.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>