Fast CI runners for e2e tests (#2895)
# Motivation

End-to-end tests take a lot of time and we are adding more of them.
They are sharded in 2 shards, which can easily be increased but that
also duplicates the setup overhead.

# Changes

Set `runs-on: ubuntu-latest-m` for Playwright e2e jobs on CI.

# Tests

When I tested with a fast runner on 1 of the 2 shards it ran in 4
minutes while the other took 6 minutes.

# Todos

- [x] Add entry to changelog (if necessary).