Remove timeout from waitForAbsent and fix ProjectDetail test (#2903)
# Motivation

During the ProjectDetail test the participation modal only closed after
a timeout.
We put a timeout in waitForAbsent as a work around, but the proper fix
is to advance the fake timers at the right time.
This means we can't call the same `participate` function that's used by
the e2e test because it does all the steps together and we need to
advance time between some of the steps.
We could just always advance time in `waitForAbsent` but advancing time
is only allowed when fake timers are used and there is no way to query
Jest/Vitest whether fake timers are being used.

# Changes

1. In `ProjectDetail.spec.ts`, instead of calling
`projectDetail.participate()`, perform the individual steps and `await
advanceTime()` before expecting the modal to be closed.
2. Add convenience methods to page objects to make doing the individual
steps a bit easier.
3. Remove the timeout from `waitForAbsent` in
`frontend/src/tests/page-objects/jest.page-object.ts`.

# Tests

Only tests.

# Todos

- [ ] Add entry to changelog (if necessary).

I don't think it's warranted.