Rename account identifier in account.api to make its icp nature explicit (#2910)
# Motivation

As we aim to introduce ICRC for ICP we will have to deal with account
identifier that are either a textual representation of ICRC or "old" ICP
account identifier.

The `account.api` communicates with NNS dapp (backend) canister which
uses ICP account identifier.

Renaming the parameters accordingly makes this more obvious.

# Changes

- rename `accountIdentifier` to `icpAccountIdentifier`