Wait for the main account when getting test tokens (#2914)
# Motivation

When getting tokens for testing, a main account needs to be already
loaded or you get an error.
End-to-end tests are fast enough to be burned by this so many of them
need to explicitly wait for the main account to appear before trying to
get tokens.

# Changes

In the `getICPs` function, wait for the main account to be loaded if it
isn't already.

# Tests

Tests still pass even though they no longer wait for the main account to
be present.

# Todos

- [ ] Add entry to changelog (if necessary).

Not note worthy (?)