E2E: Launchpad works when not signed in (#2919)
# Motivation

We want to move away from wdio e2e tests to playwright e2e tests.
`e2e-tests/specs/launchpad.e2e.ts` tests that the launchpad loads both
when logged in and not logged in.
But the only thing it really tests is that images on the page load.
I'm assuming that this is not really about images but just to test
something on the page with minimal investment.
We already have [a
test](https://github.com/dfinity/nns-dapp/blob/main/frontend/src/tests/e2e/images.spec.ts)
that tests that images load on the front page and I'm not convinced
there is much value in testing that on every page.
So what remains is to test that the launchpad works when not logged in.

# Changes

1. Reorder `sns-participation.spec.ts` to only sign in after seeing the
launch pad and project details.
2. Remove `e2e-tests/specs/launchpad.e2e.ts`.

# Tests

yes

# Todos

- [ ] Add entry to changelog (if necessary).

Covered by existing entry:
* Migrate some e2e tests from wdio to playwright.