Use ICRC textual addresses for ICP (#2869)
# Motivation

Add support for Icrc to ICP behind a flag (`ENABLE_ICP_ICRC`). When
turned on, the accounts' identifier will be display with the Icrc
textual representation instead of ICP.

Note that the Nns wallet will support both Icrc and ICP adresses. This
because the utils `findAccount` tries to match an account for both
types.

# PRs

- [x]  #2871
- [x] #2888
- [x] #2889
- [x]  #2897
- [x] #2898
- [x] #2908
- [x] #2909
- [x] #2910
- [x] #2911
- [x]  #2917
- [x] #2921
- [x] #2924

# Changes

- enhance `findAccount` to match Icrc or ICP identifiers
- update few types
- use already reviewed utils `toIcpAccountIdentifier` to load balance,
transactions and rename sub account as NNS dapp backend except an ICP
address