Fix SnsNeuronDetail.spec.ts (#2927)
# Motivation

`SnsNeuronDetail.spec.ts` has 2 identical test:
* "when project is an invalid canister id"
* "when neuron is not found"
Both use an invalid universe ID but the second one was supposed to use a
valid universe ID but a neuron ID for a neuron that can't be found.

# Changes

In "when neuron is not found":
1. Use a correct `universe`.
2. Use an unknown neuron ID.
3. Expect `snsGovernanceApi.getSnsNeuron` to have been called twice
(query + update).

In "when project is an invalid canister id", expect
`snsGovernanceApi.getSnsNeuron` not to have been called to distinguish
it from the other test as both tests expect a redirect to the neurons
overview page.

# Tests

yes

# Todos

- [ ] Add entry to changelog (if necessary).

no