Use Readable type when Writable is not needed (#2939)
# Motivation

Every `Writable` store is also a `Readable` store but not every
`Readable` store is also a `Writable` store.
So if we don't use the `set` or `update` function on a store, it's more
friendly to callers to accept type `Readable<T>`.
This allows functions to also accept queued stores from
https://github.com/dfinity/nns-dapp/blob/main/frontend/src/lib/stores/queued-store.ts

# Changes

Change type `Writable` to `Readable` in some places where a `Writable`
store is not needed.

# Tests

existing tests still pass

# Todos

- [ ] Add entry to changelog (if necessary).

not necessary