Canister status "endless loop" (#2950)
# Motivation

On `mainnet` - not reproducible locally and on testnet - querying the
status of canisters for which the user is not a controller currently
results in an "endless loop" (call finally expires after 5min).

Not sure exactly what was the root cause but, debugging the issue lead
to figure out the issue can be resolved by bumping `agent-js` to last
version `v0.18.1`. We were using an "older" version `v0.15.1` which
received multiple updates.

# Changes

- bump `agent-js`

# Screenshots

Call in endless loop:


![image](https://github.com/dfinity/nns-dapp/assets/16886711/6bacda07-7d0f-4f00-b916-a98ebec2ce0b)

Excepted error (user is not a controller) after 5min only:


![image](https://github.com/dfinity/nns-dapp/assets/16886711/025aa95a-3667-406b-a30e-fee68871af16)


![image](https://github.com/dfinity/nns-dapp/assets/16886711/0a054b41-047c-46d3-abc5-27a5c934982f)