E2E flakiness: Don't vote on proposal created by a different test (#2954)
# Motivation

End-to-end tests run in parallel against the same environment.
Proposals or global objects and so proposals created in one test can be
seen by proposals created in another test.
This causes flakiness when a proposal gets more votes than expected
because another test is voting for it at the same time.
This is likely to happen because we run the same test with Chrome and
Firefox, potentially at the same time.

# Changes

1. In `frontend/src/tests/e2e/neurons.spec.ts` record the proposer ID
and then choose a proposal with that proposer to vote on. This makes
sure that different tests vote on different proposals.
2. In `createDummyProposal` return the neuron ID used to create the
dummy proposals.
3. Page object changes.

# Tests

I was able to reliably reproduce the issue by adding a delay after
creating the proposals.
Then I made sure the issue no longer reproduced with the same delay in
play, before removing the delay again.

# Todos

- [ ] Add entry to changelog (if necessary).

not necessary