Dev build (#2936)
# Motivation
We would like to add API functions to the nns-dapp that are present only
in test deployments. For this, we need a separate test build.

This is not the first case of a wasm having build flavours. For example,
[Internet Identity has production, test and dev
flavours](https://github.com/dfinity/internet-identity/releases/tag/release-2023-07-21).

# Changes
* Add a feature flag called `test` to the nns-dapp crate. The feature
flag will be used to gate API methods available only in test builds.
* Build with `test` in the dockerfile.
* Add the test build to the list of wasms published in releases.
* For consistency, the `noassets` build, used for testnets, is also
published. The `noassets` build could also be named the `tiny` build but
that is outside the scope of this PR.
* Note: This gives us three build flavours: `production`, `test` and
`tiny` aka `noassets`
* Note: For consistency, the wasms are named:
`nns-dapp_${flavour}.wasm.gz`
* Create a separate API method list for each build flavour. The API
method list is used to check that the wasm API is as expected, and we
expect the APIs to diverge.

# Tests
- See a toy release:
https://github.com/dfinity/nns-dapp/releases/tag/dev-build-test-tag

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>