GIX-1777: Communicate when No Voting Power (#2996)
# Motivation

Communicate better when a neuron can't vote.

In this PR: Improve the subtitle of the new neuron details page when
neuron has no voting right and also the value in the voting power
section.

# Changes

* Check whether neuron can vote and show a related message in
NnsNeuronPageHeader and SnsNeuronPageHeading.
* Check whether neuron can vote and show a related text in
NnsNeuronVotingPowerSection and SnsNeuronVotingPowerSection.

# Tests

* Test new cases in the page heading subtitle.
* Test new cases in the voting power section.

# Todos

Covered by changelog entry related to new neuron details UI.