Improve ux by using hyperlinks for the cards (#3005)
# Motivation

Accounts, neurons, proposals, launchpad and canisters cards are button
which is a bit annoying as it makes complicated for the users to work
with multiple tabs. By modifying these cards to become real hyperlinks,
we will allow the user to open more easily information in separate tabs.

# PRs

- [x] https://github.com/dfinity/gix-components/pull/262
- [x] https://github.com/dfinity/gix-components/pull/267
- [x] https://github.com/dfinity/gix-components/pull/266
- [x] https://github.com/dfinity/gix-components/pull/265
- [x] https://github.com/dfinity/gix-components/pull/264
- [x] https://github.com/dfinity/gix-components/pull/263

# Changes

- provide new property `href` instead of using `on:click` to navigate
through the cards from pages to details page
- review `role` of the cards

# Screenshots


![demo](https://github.com/dfinity/nns-dapp/assets/16886711/11def01f-1f25-4f54-9548-b86fd56eb73a)