refactor: Move the stable memory parsing code. (#2993)
# Motivation
We wish to be able to change the schema used to store data. For this,
the post_upgrade hook will need to be able to read data in the format of
the last production canister and in the format of the next canister (to
support rollbacks).

For this we need to determine the schema and branch on that, post
upgrade.

# Changes
* Move the stable memory parsing routines into state.rs.
* Define a function that returns the schema (currently always "none").

# Tests
* CI should verify that state is preserved across downgrades and
upgrades.
* The format is unchanged so cycles and memory usage should be
unchanged.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>