Refactor: Move out sns aggregator types and converters (#3035)
# Motivation

We want to create an SNS aggregator store with the data returned from
the aggregator.

In this PR: I moved the converters and types of the aggregator outside
the api module to be reused.

Following PR: The api layer will return the raw data and the service
will convert it before putting it in the current stores

# Changes

* Move all types to `types/sns-aggregator`.
* Move all converters to `utils/sns-aggregator-converters.utils`.
* Change imports.
* Upgrade ic-js dependencies to use a direct import for
`SnsGetInitResponse`.

# Tests

* Move converter test to `utils/sns-aggregator-converters.utils.spec`.
* Change imports.

# Todos

I don't think moving files is worth a changelog entry.