Fix: SNS Aggregator page 2 doesn't exist (#3037)
# Motivation

Bug in current mainnet. The NNS Dapp sees that the first SNS aggregator
page is full; therefore, it tries to get the second page. Yet, the
second page fails, and the whole process of getting the info from the
aggregator fails and uses the fallback.

In this PR: Ignore errors for pages higher than 0.

# Changes

* Ignore errors for pages higher than 0 in `querySnsAggregator` helper.

# Tests

* I added first a test, it failed, and then I implemented the solution.

# Todos

- [x] Add entry to changelog (if necessary).