Ignore SIGPIPE in past-changelog-test (#3059)
# Motivation

`scripts/past-changelog-test` has been flaky recently, occasionally
failing with exit code 141.
I believe the issue is that `grep -m 1` stops reading input as soon as
it finds a match and then the process before the pipe fails because it
can no longer write to the pipe.
I'm not sure why this happens only sometimes.

Some references:

https://stackoverflow.com/questions/19120263/why-exit-code-141-with-grep-q

https://unix.stackexchange.com/questions/582844/how-to-suppress-sigpipe-in-bash

# Changes

When the line with `grep -m 1` fails, ignore it if it failed with exit
code 141.

# Tests

Difficult to test because I couldn't reproduce the issue (well, only
once) in this script.
But it reproduces quite well with `yes | head -1` so I did some manual
tests with that.

# Todos

- [x] Add entry to changelog (if necessary).