Fix flaky proposals test (#3091)
# Motivation

Fix the flaky proposals test. The e2e test did assumption that the first
Governance proposal it sees is one that was created. But e2e tests run
in parallel and sometime this behaviour throws [an
error](https://github.com/dfinity/nns-dapp/actions/runs/5833943590/job/15822555814).

# Changes

- replace `getFirstProposalCardPoForTopic` with
`getFirstProposalCardPoForProposer`
- move filters test before proposal details (to be sure that filtering
works as expected)
- apply filters before `getFirstProposalCardPoForProposer` to be sure
that just created dummy proposals are on the first page.

# Tests

Yes

# Todos

- [ ] Add entry to changelog (if necessary).