Fix: Nns Followee component test (#3094)
# Motivation

Fix
src/tests/lib/components/neuron-detail/NeuronFollowingCard/Followee.spec.ts
test.

The problem was that the knownNeuronsStore was filled in one test, but
never reset. Yet, most of the tests relied on that neuron to NOT be a
known neuron.

The solution is to reset the knownNeuronsStore before each test.

# Changes

* New method in knownsNeuronsStore "reset".
* Reset the knownNeuronsStore beforeEach for Followee.spec.
* Remove unnecessary `waitFor` in the last test case of Followee.spec.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Already covered by a changelog entry.