Make past-changelog-test work on main branch (#3105)
# Motivation

The change log change in https://github.com/dfinity/nns-dapp/pull/3090
got merged incorrectly.
We have a test for this but it compares the current branch with main so
it doesn't see any changes when run against `main`.

# Changes

When running `scripts/past-changelog-test` against `main`, compare with
`main~1` (the previous commit) instead of with `main`.

# Tests

I'm not sure how to properly test this but the output of the test looks
like it's doing the right thing and I plan to also look at the output of
the test on different commits on the main branch after merging.
Other suggestions welcome.

# Todos

- [x] Add entry to changelog (if necessary).