Fix: HW cannot participate in swap (#3112)
# Motivation

We still don't allow HW to participate in the swap. Yet, there was a bug
where the HW account was selectable in the modal to participate.

In case that the user selected the HW, they wouldn't participate with
the HW. Instead, they would participate with the main account.
Therefore, what the user selected, was ignored.

# Changes

* New variable "filterSourceAccounts" in TransactionForm using the prop
"skipHardwareWallets".
* Pass the "filterSourceAccounts" to "TransactionFromAccount".
* Remove using the prop "skipHardwareWallets" when filtering the
destination accounts. This was an error never seen because the
TransactionForm had never "skipHardwareWallets" with selectable
destination account.

# Tests

* Add methods to related POs to expose the available source accounts.
* Add a test case in ParticipateSwapModal to check that HW accounts are
not available.

# Todos

- [x] Add entry to changelog (if necessary).