SNS aggregator always supplies an partial last page (#3134)
# Motivation
We wish to change the logic for getting all pages:
* From: Get pages indexed 0, 1, 2, ... until there is a 404
* To: Get pages until there is an incomplete page

This requires the caller to know how many SNSs are in a complete page
and also stops prematurely if the cache is still missing some entries,
as the cache is filled from the latest index to the earliest. However
this logic seems to be convenient for the client.

# Changes
* Always provide an incomplete last page.

# Tests
* The test that checks the pagination has been updated and is now also
run in the upgrade/downgrade test.
* Note: I would like to test with no SNSs as well but `dfx stop ; dfx
start --clean` just fails too often.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>