GIX-1823: Use one store for sns derived state (#3144)
# Motivation

We want to use the snsDerivedStateStore in the derived store
`snsSummariesStore`. To do that, we can't use dynamic creationg of the
store. To allow dynamic creation of the store, we would need to also
create dynamically `snsSummariesStore`. This is a bigger refactor than
expected and not sure it's worth the effort now.

Therfore, I implemented the same pattern of storing all project data in
one single store using the root canister id as key in the
snsDerivedStateStore.

# Changes

* `snsDerivedStateStore` doesn't create store dynamically. It's one
store to rule them all.
* Refactor where this store was created.

# Tests

* Test new capabilities and remove old ones.
* Change test after refactor.

# Todos

- [ ] Add entry to changelog (if necessary).
Not worth a changelog entry.