detailed min participation (#3150)
# Motivation

Modclub set `"min_participant_icp_e8s":100000278` which we round in the
UI to 1.00 ICP.
Then when you try to participate with 1 ICP you get an error saying
"Sorry, the amount is too small. You need a minimum of 1.00 ICP to
participate in this swap."

# Changes

Call `formatToken` with `detailed: true`. This renders all significant
digits if amount >= 0.01 (if I understand the code correctly).

# Tests

no

# Todos

- [x] Add entry to changelog (if necessary).