IdentityAgentWrapper to use cached agent with specified identity (#3171)
# Motivation

We cache created agents for performance (according to
https://github.com/dfinity/nns-dapp/pull/1600).
But the agent keeps a reference to the identity used to create the agent
which means the identity is implicitly cached as well.
This has been causing problems with hardware wallet interactions where
we need to keep state on the identity to indicate that a ledger transfer
is for the purpose of staking a neuron.

# Changes

1. Add a wrapper agent which uses a newly specified identity instead of
the identity used to create the agent originally.
2. When returning an agent from the cache, wrap it in the wrapper such
that the current identity is used instead of the previously cached one.

# Tests

1. Method calls are forwarded to the underlying agent.
2. The identity can be invalidated.
3. Method calls on different wrappers result in calls on the same
underlying agent with different identities.

# Todos

- [x] Add entry to changelog (if necessary).