Fix order dependence in sns-proposals.services.spec.ts (#3181)
# Motivation

Some tests were generating errors on the console but only one of the
tests mocked the console.
So if another test logged to the console, it would fail.

# Changes

Explicitly mock `console.error` in the tests that require it and restore
the mock before every test.

# Tests

Pass with random seeds 1 - 20.

# Todos

- [ ] Add entry to changelog (if necessary).
covered