Static loading spinner (#3190)
# Motivation

We have no control on the way the chunks are loaded. As a result, it's
possible that it waits until most chunks are loaded before displaying
the hints - the spinner - that the dapp is booting which is leading to
an unfriendly UX as a blank screen might be displaying for a while.

By rendering a spinner within the static html page and removing it once
most of the JS files are loaded, we can overcome this "race condition".

# Changes

- add (copy) spinner in `app.html`
- when the auth is ready, which indicates most JS resources are loaded,
remove the spinner

# Tests

https://kgqse-raaaa-aaaaa-aactq-cai.nnsdapp.dfinity.network/
https://kgqse-raaaa-aaaaa-aactq-cai.nnsdapp.dfinity.network/accounts

https://xnjld-hqaaa-aaaal-qb56q-cai.icp0.io/

# Screenshots

Mainnet - current main - Fast 3g network => spinner appears for a milli
second when all js is loaded


![mainnet-no-spinner](https://github.com/dfinity/nns-dapp/assets/16886711/6a740ee4-0b83-4b1c-b3b4-137c78df2657)

Mainnet - this PR - Fast 3g network


![mainnet-spinner](https://github.com/dfinity/nns-dapp/assets/16886711/5dd86910-d627-4e3e-b4ab-52e27f3a58dd)

Mainnet this PR - full speed network


![full-speed](https://github.com/dfinity/nns-dapp/assets/16886711/d3efe0de-02f3-442b-89dd-c6364a997949)