GIX-1823: Fix ProjectDetail.spec (#3238)
# Motivation

Final motivation: remove dependency on get_state. Middle step, remove
snsQueryStore.

In this PR: When I set the enable sns aggregator store feature flag to
true, the ProjectDetail.spec fails because some parameters are not set
properly.

It's passing with snsQueryStore because the defaults are aligned with
the test.

# Changes

* Add parameters when setting sns projects in ProjectDetails.spec file

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessay.