GIX-1823: Use ENABLE_SNS_AGGREGATOR_STORE in sns.services (#3241)
# Motivation

Final motivation: set ENABLE_SNS_AGGREGATOR_STORE to `true`.

Do not set the snsQueryStore in the services unless
`ENABLE_SNS_AGGREGATOR_STORE` is false.

Otherwise, the tests need to also fill the `snsQueryStore` and we want
to remove it.

# Changes

* Use ENABLE_SNS_AGGREGATOR_STORE in `loadSnsDerivedState` and
`loadSnsLifecycle`.

# Tests

* Test both `loadSnsDerivedState` and `loadSnsLifecycle` with flag set
`true` and `false`. The difference is only that when it's `false`, we
don't expect the snsQueryStore to be filled.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.