GIX-1759: Remove unnecessary variant prop (#3256)
# Motivation

A "variant" prop was introduced to reused the same buttons when
implementing a new UI for neuron details.

The old UI has already been removed.

In this PR, I remove the "variant" prop which is not needed anymore and
set the buttons to "secondary".

# Changes

* Remove "variant" prop in neuron related buttons.

# Tests

* Remove tests regarding the variant prop.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary. Already covered by the changelog entry specifying
cleaning components.