Import canister IDs from testnet. (#3253)
# Motivation

Releasing to testnet, requires a `canister_ids.json` file.
We used to ask a team member for this but this can be tedious.

# Changes

Add a new `import-from` flag to `scripts/canister_ids` which allows you
to pass a URL.
If the URL does not contain `.raw.` it will be added. Then the URL is
fetched, gunzipped, parsed, and the relevant canister IDs are added to
your `canister_ids.json`.

# Tests

1. Changed the test to start each test with a clean json file.
2. Added a test for `--import-from`

# Todos

- [x] Add entry to changelog (if necessary).