Update example (#3251)
# Motivation
The rustdoc for the AccountsDb trait needed to be updated after
1bb96c0b288f98981d7bb1093044e54f8daf8cc7.

Note: This was not caught in CI because the schema module is still
behind test configuration. It probably make sense to remove the test
flag now, as the code will be used soon and doing so catches errors like
this.

# Changes
* Add the schema module to the production codebase. It is not used yet
so the optimizer will probably remove it. This is fine.
* Update the rustdoc.

# Tests
- See CI.

# Todos

- [ ] Add entry to changelog (if necessary).
  I think this is minor enough not to need a changelog entry.

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>