Add QrWizardModal component (#3260)
# Motivation

Using the QR code scanner requires adding an extra step to a
`WizardModal` and switching back and forth between them.
This is currently done in `TransactionModal`.
I want to reuse the QR code scanner in the `DisburseNnsNeuronModal` so I
want to streamline its usage.

# Changes

1. Add `QrWizardModal` which manages going to and from the QR Code step.
2. Use `QrWizardModal` in `TransactionModal`.
3. Tests for `QrWizardModal`.

# Tests

Tested manually on local that scanning a QR code when making a
transaction still works.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary for a small refactoring