Accounts db (#3240)
# Motivation
We would like to move account storage to stable memory. The plan
involves creating a virtual database that proxies calls to one or more
underlying database implementations.

# Changes
- Place a proxy layer between database calls and the current storage.
Note: Currently accounts are stored on the heap, in a BTreeMap. This is
unchanged. However access to the map has to go through the proxy now.

# Tests
* I think that existing tests should suffice.

# Todos

- [x] Add entry to changelog (if necessary).
- Covered by the existing comment: Add a database abstraction layer,
preparing for migration.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>