Fix empty and missing parameters in canister import (#3286)
# Motivation
The canister import command fails if there is no `canister_ids.json`
file (a common case in a clean new repository). Likewise if `JSON_FILE`
is specified as the empty string (an uncommon case spotted when looking
for the cause of the earlier error).

# Changes
- Provide the variables `CREATE` and `JSON_FILE` with defaults so that
bash does not crash on using an unassigned variable.
- Specify `nargs=0` for the `--create` flag. As it is, it will try to
consume the next argument as the value. This will cause the command to
fail or to parse the line otherwise than intended.

# Tests
## Run in a blank repo
Before:
```
$ ./scripts/canister_ids
./scripts/canister_ids: line 33: CREATE: unbound variable
```
After:
```
$ ./scripts/canister_ids
./canister_ids.json is missing.
```
## Triggering the incorrect nargs
Before:
```
$ ./scripts/canister_ids --import-from-index-html https://df3os-zaaaa-aaaaa-aaddq-cai.nnsdapp.dfinity.network/ --create
/tmp/clap-iQ1AgH.tmp: line 66: $1: unbound variable
```
After:
```
$ ./scripts/canister_ids --import-from-index-html https://df3os-zaaaa-aaaaa-aaddq-cai.nnsdapp.dfinity.network/ --create
$
```
(No response, it just completes successfully.

# Todos

- [ ] Add entry to changelog (if necessary).
  None needed, IMO

---------

Co-authored-by: Max Murphy <max@winning.black>