GIX-1864: Improve ICP logo token selector (#3305)
# Motivation

The logo of ICP in the token selector doesn't have enough contrast when
user hovers or when it's the selected one.

In this PR, I add an outline to the logo to help with the contrast. See
screencast.

# Changes

* Use `noPadding` prop in Card for `SelectUniverseCard` and change the
variable that sets the logo frame color on hover, focus and selected.

# Tests

* Only CSS changes.

# Todos

- [ ] Add entry to changelog (if necessary).