GIX-1911: Split participation in progress bar (#3353)
# Motivation

We want to differentiate between direct and Neurons' Fund participation.

In this PR, the differentiation is done only in the progress bar.

This is not yet supported because it's missing the field in the API
response. Therefore, I used a getter that returns undefined.

I don't think there is any need for a feature flag. As soon as the field
is available, we should support it.

# Changes

* Add a new sns summary getter `getNeuronsFundParticipation` that
returns `undefined`.
* Change prop `value` in CommitmentProgressBar to two props
`directParticipation` and `nfParticipation`.
* Use the new `segments` prop of the ProgressBar instead of the `value`.
* Use the new getter `getNeuronsFundParticipation` in ProjectCommitment
to use new props in CommitmentProgressBar.

# Tests

* Adapt tests of CommitmentProgressBar to new props and add a test case.
* Add a test case in ProjectCommitment.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary yet because there is no change for the user until the
field is supported by the API.