Toy accounts for unit tests (#3380)
# Motivation
We need sample accounts of various sizes for use in low level unit
tests.

There is already code for creating toy accounts as part of an
`AccountStore`, but this is unsuitable when we don't need the
`AccountStore` or its execution environment. We need just the banana,
not the gorilla holding it or the jungle around the gorilla.

# Changes
- Add test code for creating toy accounts.

# Tests
- A test is included that checks that the toy account has the specified
size.

# Todos

- [ ] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>