GIX-1859: New components for the new Canister Details page header (#3386)
# Motivation

Change the canister details header to be more consistent with the new
header in neuron details.

In this PR, I introduce, but not use the components needed to render it.

In another PR, I'll use them in the CanisterDetails and I'll change the
tests in there as well.

# Changes

* New component CanisterHeadingTitle.
* New component CanisterPageHeader.
* New component CanisterPageHeading that uses new component
CanisterHeadingTitle.
* Check whether the "subtitle" slot is present in the PageHeading.
Otherwise, there is an empty space because it uses flex and gap with an
empty element.

# Tests

* Test files and POs for CanisterHeadingTitle, CanisterPageHeader,
CanisterPageHeading

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet. I'll add it with the PR that uses the components.