Fix aggregator changelog (#3400)
# Motivation
The aggregator changelog was invalid with "Proposal 124250" repeated and
multiple empty sections.

It appears that the earlier entry corresponded to the actual [proposal
124250](https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=124250);
that contains some empty sections that should have been purged but
otherwise that section is correct.

The duplicate section is:

> Various minor style improvements: favicon, spacing, status "open"
color and text clamp

This appears to match the rejected [proposal
124535](https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=124535)
although that proposal has a different summary.

The proposal was rejected due to reproducibility issues, not due to any
fundamental rejection by the community of the changes, so the changes
from 124535 were included in the next proposal
([124788](https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=124788)).

# Conclusion
* It looks as if the proposal rejection might have thrown the normal
release procedure out of whack. We probably need to document the
rejection path more thoroughly and provide better tooling.
* Minimal changes should be made to the changelog to make it an
accurate, well formed history.

# Changes
- Remove the empty sections.
- Remove the duplicate proposal header and add the line under that
section to the release that actually included that change.
- Reinstate the section headers that were lost.

# Tests
- See feedback by humans.

# Todos

- [ ] Add entry to changelog (if necessary).
 - Not applicable

---------

Co-authored-by: Max Murphy <max@winning.black>