GIX-1931: Hide commitment split if NF participation is 0 (#3410)
# Motivation

We agreed with design and product to hide the split of direct and NF
commitment until the NF commitment is more than 0.

# Changes

* Change the condition to show the split to the NF commitment greater
than 0.
* Remove unused helper `isFullProjectCommitmentSplit`.

# Tests

I even did TDD in for this task. First I changed the test. It failed.
Then I changed the code and see it pass 🚀
* Change the description and expectation of the test case when NF is
zero.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary yet.