GIX-1859: Implement new header Canister detail page (#3414)
# Motivation

Change the header of the canister detail page to be more consistent with
the new neuron details page header.

In this PR, we use the components introduced in #3386 in the
CanisterDetails.svelte

# Changes

* Use CanisterPageHeader and CanisterPageHeading in
CanisterDetails.svelte
* Manage the spacing between components in the `section` in
CanisterDetails instead of each child setting the spacings.
* Set Unlink button to secondary. Buttons in the header are secondary.
* Change the text when the canister is not the controller to better
explain the "Balance unavailable" title.
* Remove CyclesCard.svelte

# Tests

* Adapt tests in CanisterDetails page to new UI and logic. I didn't move
ALL tests to use POs. Only the ones related to the changes.
* Add some methods to related POs.

# Todos

- [x] Add entry to changelog (if necessary).