GIX-1922: Undo CommitmentProgressBar split commitments (#3433)
# Motivation

The new participation UI doesn't show a progress bar with commitments of
NF and direct as different colors. Instead, if shows only the direct
participation only.

In this PR, I remove the props, logic and tests that allow the
CommitmentProgressBar to show multiple commitments.

I have also adapted the commitment when NF is present, to show only the
direct participation. Yet, the UI still needs some changes that will be
done in another PR.

# Changes

* Change props `directParticipation` and `nfParticipation` for
`participationE8s` in CommitmentProgressBar. Adapt logic inside.
* In ProjectCommitment, use the overall commitment or the direct
commitment for the CommitmentProgressBar. I also moved the label and
value of the NF commitment after the progress bar. Even though this UI
is not available, I didn't want it to be in a confusing state. I also
removed the left spacing of the labels for direct and NF commitments.

# Tests

* Fix, change and remove tests in ProjectCommitment.
* Fix, change and remove tests in CommitmentProgressBar.
* Rename and remove unncessary methods from CommitmentProgressBarPo.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet necessary.