Init vitest (#3444)
# Motivation

We have to migrate from jest to vitest; otherwise, we are stuck with
Svelte v3 forever. I did not find any ways to set up jest for ESM in our
complex dependencies context. ESM which is required by Svelte v4 and the
testing library.

As it is complicated to migrate everything in a single go (see PR
#2475), this PR instantiates and configures
[vitest](https://vitest.dev/). That way, we can develop new tests with
it, and we can iteratively migrate to vitest. Either by migrating
particular test when we develop new features in existing functions or by
providing PRs with batchs of tests migrated.

# Changes

- configure vitest
- setup vitest and jest to be run on `npm run test`
- migrate 2-3 tests to vitest to validate the configuration