Robust traits (#3442)
# Motivation
We will need to import different sets of traits to different files. To
do that, we need to parameterize the set of traits. Additionally, the
current trait insertion is fragile in that changes to didc can break it
easily; that needs to become robust.

# Changes
- Move the inserted traits to the beginning of each `#[derive(...` line,
rather than depending on didc inserting a specific trait.
- Extract the inserted traits as a variable.

# Tests
- Existing tests should suffice, as the rust code is identical apart
from the ordering of the traits.

# Todos

- [ ] Add entry to changelog (if necessary).
  - Covered by existing changelog entries.

---------

Co-authored-by: Max Murphy <max@winning.black>