Specialize did2rs for the aggregator (#3436)
# Motivation
We wish to use `did2rs` to derive rust code for use in proposals
rendering. Currently `did2rs` has parameters hard-wired for use in the
SNS aggregator.

# Changes
- Replace hard-wired constants in `did2rs` with command line arguments.
- Put the constants applicable to the SNS aggregator in
`scripts/sns/aggregator/did2rs`; this can be used as a drop-in
replacement for the former `did2rs.sh`.
- Update other commands to use the aggregator-specific `did2rs`.
- Note: Some of these other files should also be moved into
`scripts/sns/aggregator/` but that is not done in this PR in order to
keep the PR size small.

# Tests
Existing functionality is unchanged, so existing tests should suffice.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>