Chore: Clean up sns services spec from snsQueryStore (#3453)
# Motivation

Final motivation: remove snsQueryStore.

In this PR, sns services spec file was filling up the snsQueryStore and
then checking the values. The services being tested do not use the
snsQueryStore. Therefore, it was a mistake to add those checks and those
filling stores.

# Changes

* Remove filling up and checking the snsQueryStore in sns.services.spec

# Tests

Only changes in test files.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.