feat (proposal rendering): Get sns_wasm types from .did file (#3471)
# Motivation
We would like to free the nns-dapp from the code and libraries pulled in
by the nns-sns-wasm canister.

# Changes
* Remove the dependency on the `sns_wasm` canister from the proposals
rendering crate.
* Derive the `nns-sns-wasm` canister types from `.did`.

Note: The `.didc` is from 2023-09-27 whereas the removed library is from
`2023-08-01`.

# Tests
## Does proposal rendering still work?
There are existing Rust tests for proposal rendering; these should
suffice.
## Is this actually reducing dependencies?
We can get a rough measure of the number of dependencies in the proposal
rendering crate with:
```
cargo tree -p proposals | wc -l
```
For the current release candidate, i.e. a recent commit on main, this
yields 1555 entries. In this branch this has dropped to 1172 and,
critically, no longer contains the timer crate.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>