Chore: Migrate ProjectDetail (#3513)
# Motivation

Migreate tests to vitest.

In this PR, migrate ProjectDetail page to vitest.

# Changes

* Common changes to migrate the test.
* `blockedApiPaths` only for the module used in a fake. I don't
understand why it was working before. It seems wrong.
* Close the ParticipateSwapModal immediately if test environment. This
was the problem for all the participation tests. The modal wasn't
closing, no matter how much time it was advanced. Like this, the modal
closes and the tests pass.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.