GIX-1936: Add "init" to SnsSummary (#3515)
# Motivation

Support the NF enhancements. The new parameters will be added to the
`get_init` endpoint.

In this PR, I introduce a field `init` in the SnsSummary with the data
from `get_init`.

The aggregator already supports it, but the NNS Dapp was not using it
until now.

Even the oldest SNS supports the endpoint:
https://3r4gx-wqaaa-aaaaq-aaaia-cai.icp0.io/v1/sns/root/zxeu2-7aaaa-aaaaq-aaafa-cai/slow.json

# Changes

* Refactor `convertSwapInitParams` to return the type and convert to
nullable if needed.
* Add field in SnsSummary `init`.
* Read the related field from the aggregator data, convert it to sns-js
type and add it when creating SnsSummary.

# Tests

* Test that the SnsSummary contains the new field from
`convertDtoToSnsSummary`.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet necessary.