Chore: Remove unnecessary sns converter (#3537)
# Motivation

Remove unnecessary SNS Aggregator converters and mocks.

Before, we wanted to convert ALL the aggregator data to sns-js types.
But now we can convert only the needed fields on demand.

This will simplify the conversion, mocks and types managed in NNS Dapp,
and hopefully make it more understandable.

# Changes

* Remove the generic converter `convertDtoData`. Instead use the
specific converters in sns.services.
* Remove converters used in `convertDtoData` and nowhere else.
* Remove related type `CachedSns`.

# Tests

* Remove the test for `convertDtoData` and instead test the specific
converters used in sns.services.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.