Don't define mocks directly in describe blocks. (#3545)
# Motivation

Code in `describe` blocks is called before any of the tests are
executed.
For example the `jest.spyOn(` at the top of the `"removeFollowee "`
block is called after the `jest.spyOn(` at the top of the `"addFollowee
"` but before any of the tests in the `"addFollowee "` blocks.

This can cause confusing behavior if the tests depend on code from its
own `describe` block.

Additionally, when we migrate to vitest, defining mock behavior on the
same function again, prevents an earlier spy from spying. This means the
following fails in vitest while the equivalent does not fail in Jest:
```
      const obj = { func: function() { } };
      const spy1 = vi.spyOn(obj, 'func');
      const spy2 = vi.spyOn(obj, 'func').mockImplementation(() => { });
      obj.func();
      expect(spy2).toBeCalled();
      expect(spy1).toBeCalled();
```

# Changes

In `frontend/src/tests/lib/services/sns-neurons.services.spec.ts` move
all `spyOn`s that are directly in a `describe` block to a `beforeEach`
block isntead.

After this change, migrating to vitest should be straightforward.

# Tests

pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary