Fix: Finalizing status in ProjectCard (#3546)
# Motivation

Issue reported by a user:
https://forum.dfinity.org/t/bug-projectpage-sns-launchpad-direct-url/23029/7?u=lmuntaner

Add the "Finalizing" status in the Launchpad, when we show all the
SNSes.

Yet, to do that, we need to fetch the finalizing status of all the
projects.

To not make unnecessary calls, I added a check that the project has
finished within the last week. This way we don't make calls to for open
or old projects.

# Changes

* Changes in `loadSnsFinalizationStatus` to not call the API unless it's
possible that the project is finalizing. Also change the argument to
accept another parameter `forceFetch`.
* Call `loadSnsFinalizationStatus` on mounting the ProjectDetailCard.
* Use `createIsSnsFinalizingStore` on ProjectCardSwap to render
"Finalizing" instead of "Completed" when needed.
* Force fetching autofinalization after a participation. At that moment
the project might not be loaded as Committed locally yet, but it might
be Committed already.
* Fix the `SnsFinalizationStatusStore` that might be `undefined`.
* New sns util `swapEndedMoreThanOneWeekAgo` for convenience.

# Tests

* Add `swapDueTimestampSeconds` parameter when creating an SNS.
* Add `getStatus` to ProjectCardPo.
* New ProjectCardSwapInfoPo.
* Add a test case in ProjectCard that shows the Finalizing status. Also
adapt it so that it has SNS setup. Until now it was not needed because
it was using the props only.
* Add a test case in ProjectCardSwapInfo for the Finalizing status.
* Fix the swap ending time in ProjectDetail so that it triggers the call
to get finalizing data.
* Edit tests in sns-finalization.services.spec after the changes in
`loadSnsFinalizationStatus`.
* Test new sns util `swapEndedMoreThanOneWeekAgo`.

# Todos

- [x] Add entry to changelog (if necessary).