CLI tool for proposal rendering (#3482)
# Motivation
The proposals crate has quite a few test vectors, however I still wanted
some confidence that real proposals would render correctly.

Hence a command line tool, that allows running the proposals renderer
against any proposals payload that has been downloaded from mainnet.

I downloaded hundreds of proposals and rendered them with the current
code and with older versions of the proposals renderer, to check for
differences.

Adding hundreds of proposal payloads to the repository does not make
sense, as many are large. It is quite a useful tool none the less.
Perhaps we can devise a test suite that uses it.

# Changes
- Add a `main.rs` to the proposals crate, to create a proposals
rendering command line tool.
- Add a script that downloads a proposal and renders the payload.

# Tests
None yet. Maybe this could form the basis of a test suite in future,
though.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>