Add --host to scripts/dfx-snapshot-start (#3574)
# Motivation

We are considering using DevEnv instead of testnet to test release
candidates.
To make a local replica accessible remotely, one needs to pass `--host
0.0.0.0:8080` to `dfx`.

# Changes

Add flag `--host` to `dfx-snapshot-start`. If set, the flag is passed on
to `dfx`.

# Tests

I ran `scripts/dfx-snapshot-start -s
~/snapshots/pinned-2023-10-05.tar.xz --host 0.0.0.0:8080` on my DevEnv,
and it is accessible from
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/
If I don't pass the flag, it gives "502 Bad Gateway".

# Todos

- [x] Add entry to changelog (if necessary).