Remove debounce on deriving network from BTC address (#3618)
# Motivation

When sending ckBTC you can enter either a ckBTC address or a BTC
address.
We derive the "network" (Internet Computer/Bitcoin) from the address.
This is debounced so there is a small delay.
The derivation does not require any network requests, and once the
network is derived we don't change it (if you later change to an address
to a different network, we consider it invalid unless you change the
network manually), so this delay does not prevent any jarring
experience.
So I see no reason to have the debounce there.
Additionally it makes unit testing slightly harder so there is a benefit
in removing the debounce.

# Changes

Remove the debounce on deriving the network from the address.

# Tests

Removing it didn't cause any tests to fail so it wasn't tested before.
Manual testing looks good and can be tried here: 

# Todos

- [x] Add entry to changelog (if necessary).