Include exceptional transactions in stats (#3630)
# Motivation
We would like to have visibility of the number of exceptional
transactions and potentially alerts for their number.

# Changes
Provide a metric with a count of the number of stored exceptional
transactions.

Note: If the exceptional transactions log ever becomes full, this metric
will plateau. If that happens before we switch to the index canister we
may need revise this statistic, e.g. with a dedicated counter.

# Tests
Awaiting #3629 to write an e2e test.

# Todos

- [ ] Add entry to changelog (if necessary).
  - Covered by an existing changelog entry

Co-authored-by: Max Murphy <max@winning.black>