Fix: Dissolve Delay Max (#3614)
# Motivation

There is a bug raised in the
[forum](https://forum.dfinity.org/t/changing-future-max-lock-up-for-sns-1-dkp/23576/34).

This screen has already been fixed with other bugs. The source of the
bugs is the duality of seconds and days and synchronizing the slider
with the input.

In this PR, we remove the slider and show a read-only progress bar
instead.

We use days to be user friendly, but the code uses the values in
seconds.

# Changes

* New component RangeDissolveDelay instead of the InputRange.
* Days are not managed in SetDissolveDelay, instead in the child
components DayInput and RangeDissolveDelay.
* New props and functionality in DayInput, like managing Min and Max
buttons.
* Change delayInSeconds of ConfirmDissolveDelay to `bigint` to avoid
issues with decimals.
* Remove rounding in `secondsToDays`.

# Tests

* SetDissolveDelay.spec. Remove tests regarding slider and check
progress bar instead.
* DayInput.spec. Add new test cases for new functionality.
* IncreaseDissolveDelayModal.spec. Use the input instead of range to set
delay.
* NnsStakeNeuronModal.spec. Use the input instead of range to set delay.
* IncreaseSnsDissolveDelayModal.spec. Use the input instead of range to
set delay. Remove test for slider.
* New PO RangeDissolveDelayPo
* Edit SetDissolveDelayPo accordingly

# Todos

- [x] Add entry to changelog (if necessary).