mapTransaction prop on IcrcTransactionsList (#3701)
# Motivation

For ckBTC withdrawal transactions we want to render the BTC address as
the "To:" address.
This address is encoded in the memo of burn transactions and we want to
place it in the `to:` of the transaction rendered in the
`TransactionCard`.
This can be done in `mapIcrcTransaction` but the logic is specific to
ckBTC so shouldn't be applied to all tokens.
My solution is to be able to pass a different `mapTransaction` function
per token.

This PR only makes it possible to pass a `mapTransaction` to
`IcrcTransactionsList` but does not use it for any custom mapping for
ckBTC.

# Changes

1. Add `mapTransaction` prop to `IcrcTransactionsList` and pass it on to
`IcrcTransactionCard` which will call the passed function instead of
`mapIcrcTransaction`.
2. Pass `mapIcrcTransaction` from `CkBTCTransactionsList` and
`SnsTransactionsList`. This is for now. In a future PR we'll pass a
custom function from `CkBTCTransactionsList`.

# Tests

Add tests for `IcrcTransactionsList` and `IcrcTransactionCard` that the
passed in mapping function is used.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary