Remove irrelevant TODOs from update-snsdemo PR description (#3714)
# Motivation

The update-snsdemo workflow used to also update candid files and rust
types.
Now it only updates the snsdemo release that we use for test environment
scripts and snapshot.
The TODOs in the automated PR description no longer apply to this
workflow.

# Changes

Remove the TODOs that no longer apply to the update-snsdemo workflow.

# Tests

Not tested. Fingers crossed that the workflow will work correctly next
week.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary.