Stop relying on duplicate transactions from index (#3734)
# Motivation

When you make a transaction where the "from" and "to" are the same
(sub)account we render this as 2 transactions: 1 sent and 1 received.
To do this we rely on the index canister returning 2 copies of this
transaction as well.
But this is a bug in the index canister that might be fixed at some
point and also it makes our code fragile.
For example, in the transactions works, we [dedupe
transaction](https://github.com/dfinity/nns-dapp/blob/d4a4d206db95027b58653ccd7ab0da6014497939/frontend/src/lib/worker-services/icrc-transactions.worker-services.ts#L63)
which means that actually most of the time we fail to render 2
transactions.

# Changes

1. Add a utility function `getUniqueTransactions` to dedupe
transactions.
2. In `icrcTransactionsStore` always dedupe all transactions instead of
only removing existing transaction from new transactions.
3. In `getSortedTransactionsFromStore`, duplicate self-transactions
instead of depending on them to already be duplicated and just marking
them.
4. Add an optional `id` param to `createIcrcTransactionWithId` for
convenience.
5. Removed template parameter from `mapToSelfTransaction` as it is no
longer used for ICRC transactions.

# Tests

Added unit tests for:
1. `getUniqueTransactions`
2. `icrcTransactionsStore`
3. `getSortedTransactionsFromStore`
4. `CkBTCTransactionsList`

Tested manually at:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).