GIX-1979: Render table visitors (#3739)
# Motivation

Show the tokens table without balance to the non logged in user.

In this PR, render the table in the SignInTokens using the derived store
with base data.

In the upcoming PRs I will improve the UX. I will add a "goToDetail"
action and logic that will redirect the user to login on row click.

# Changes

* Render `DesktopTokensTable` in `SignInTokens`.
* Expect `userTokensData` in `SignInTokens`.
* Pass `tokensListBaseStore` to `SignInTokens` in the route.

# Tests

* Add methods to get the name of the project of the row.
* Add a test in the route component to check that there are tokens when
the user is not logged in.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.