GIX-1980: Setup icp accounts with tokens table (#3775)
# Motivation

When we migrate to the tokens page, we also want a similar view for the
ICP accounts, subaccounts and hardware wallet accounts.

In this PR, I introduce the TokensTable for NNS Accounts.

# Changes

* Render `TokensTable` in `NnsAccounts` depending on feature flag.
* Expect an optional list of `UserTokenData` in `NnsAccounts`.
* Hard code a list of `UserTokenData` in the Accounts route and pass it
to `NnsAccounts`.

# Tests

* Add test in NnsAccounts.spec that the tokens table is there if feature
flag is enabled. I added myself some TODOs comments to remember
improving the test before adding more features here.
* Add methods to NnsAccountsPo.
* Add test in Accounts route to check that tokens table is rendered if
flag is enabled.

# Todos

- [ ] Add entry to changelog (if necessary).