Pass a single prop to TransactionCard (#3780)
# Motivation

We currently pass `mapTransaction` from `CkbtcTransactionsList` to
`IcrcTransactionsList` to `IcrcTransactionCard` and then we apply the
mapping and pass a lot of props to `TransactionCard`.
My goal is to simplify this and give the higher level component more
direct control over how the transactions are rendered.

In this PR I introduce the `UiTransaction` type which contains all
information necessary to render the `TransactionCard` component. This
allows us to pass a single prop to `TransactionCard`.
My goal is to eventually return `UiTransaction` from `mapTransaction`
and possibly retire the intermediate `Transaction` type.
Also, I want to map all the transactions in the higher level component
to avoid sending down all these different props.

This is the first step.

# Changes

1. Introduce a new type `UiTransaction` which has the information needed
to render a transaction in the transaction list.
2. Change `TransactionCard` to accept 1 prop of type `UiTransaction` and
remove the logic that is made obsolete by that.
3. Add `toUiTransaction` which converts `Transaction`, plus some extra
params, to a `UiTransaction`.
4. Change `IcrcTransactionCard` and `NnsTransactionCard` to pass a
`UiTransaction` to `TransactionCard` instead of a collection of props.

# Tests

Unit tests updated and pass.
Manually at:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary