Rm static testnets (#3785)
# Motivation
Static testnets are dead.  ~Long live static testnets.~

# Changes
Delete or rename references to the `nnsdapp` testnet in many, but not
all places. In particular, changes to the release SOP require special
attention so they deserve to be be in a separate PR.

Note: Each set of changes is in a separate commit, so you can check the
changes as atomic units. Any that turn out to be complicated can be spun
out into separate PRs.

# Tests
See CI

# Todos

- [x] Add entry to changelog (if necessary).