GIX-2058: Derived store ICP Tokens visitor (#3807)
# Motivation

The user might arrive directly to the accounts page of the ICP without
going through tokens.

The route is the current `/accounts` page. Although for now, it will be
available only when `u=OWN_CANISTER`.

In case a non-logged in user lands there, they should see a table like
in `/tokens` but with only ICP main account.

In this PR, I introduce the derived store for those visitors.

# Changes

* New derived store `icpTokensListVisitors`.

# Tests

* Test new derived store.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.