Remove unnecessary showTransactionFee (#3808)
# Motivation

The logic in `showTransactionFee` is unnecessary.
Burn and Mint transactions don't have a fee so even if we tried to add
it, there's nothing to add.
If `toSelfTransaction` is true, then `isReceive` is also true, so
there's no need to check `toSelfTransaction` separately.

# Changes

1. Remove `showTransactionFee` and just use `!isReceive` instead.
2. Remove test for `showTransactionFee`.

# Tests

Existing unit tests pass.
Transaction list, including Burn, Mint, Approve and TransferFrom, looks
correct.
I also changed `transactionDisplayAmount` to add a fee when it shouldn't
or not add a fee when it should and in both cases there was a failing
test so that confirms there is test coverage.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary