Map directly to UiTransaction (#3813)
# Motivation

In `IcrcTransactionsList` we map `IcrcTransactionData`s to
`Transaction`s and then to `UiTransaction`s.
This is not necessary and by doing it in 1 step we can give the custom
`mapTransaction` function more control over the rendering of the
transactions.
This will make it easier to implement ckBTC UI improvements.

# Changes

1. Change `mapIcrcTransaction` and `mapCkbtcTransaction` to return
`UiTransaction` instead of `Transaction`.
2. In `IcrcTransactionsList` map transactions with the new single
mapping function. Then filter out `undefined` transactions.
3. 

# Tests

1. Accept optional `timestamp` and `memo` params to
`createIcrcTransactionWithId`.
2. Change unit tests for `mapTransaction` to expect the full result
instead of just a few properties. Define default values so we still only
have to specify the properties relevant to the test.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary