GIX-2085: Use TokensTable in SignInAccounts (#3810)
# Motivation

The accounts page for non-logged in users should be similar to the
logged-in page.

Therefore, we use the TokensTable in SignInAccounts when flag is
enabled.

# Changes

* Move the `MainWrapper` to the layout of the route accounts and apply
it when flag is enabled.
* Add logic in the `SignInAccounts` to render different components
depending on flag value.
* Use the `icpTokensListVisitors` when rendering `SignInAccounts` in the
accounts route pages.

# Tests

* New SignInAccountsPo.
* Add methods to PO
* Add a test in the routes layout spec to check that token navigation is
present or not depending on the flag.
* Add tests in the routes page to check that either tokens table or
empty cards are present depending on the flag.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.