GIX-1983: Add token and fee to UserTokenData (#3820)
# Motivation

To render the action modals like to make an SNS transaction, the token
and the fee are mandatory props.

To make it more convenient, I added those two fields in the
UserTokenData type.

The data already lives within the `balance`, because the `TokenAmount`
has the token. But there are two reasons I don't like using it.

* It seems ugly to get the token from the balance.
* The token in the balance doesn't have the `fee` field.

# Changes

* Add `fee` and `token` to `UserTokenData`.
* Populate those fields in `tokensListBaseStore`.
* Populate those fields in `icpTokensListUser`.
* Populate those fields in `icpTokensListVisitors`.

# Tests

* New mock helper `createIcpUserToken` to create an ICP token data.
* Moved some reused mocks to a mock file instead of duplicating them in
test files.
* Add new fields to mock data.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary