Remove the fallbackDescription slot from TransactionCard (#3821)
# Motivation

We use the `fallbackDescription` on `UiTransaction` only to render
"From: BTC Network" on ckBTC mint transactions.
We can accomplish the same thing by setting `otherParty` to "BTC
Network".

# Changes

1. Remove the `fallbackDescription` field from `UiTransaction`.
2. Remove the description slot from `TransactionCard`.
3. Stop passing `$i18n.ckbtc_transaction_names` from
`CkBTCTransactionList` to `IcrcTransactionsList` and to
`mapTransaction`.
4. Pass the full `i18n` object to `mapTransaction` instead of just
`transaction_names`. This gives it additional access to
`ckbtc.btc_network` to use as `otherParty`.
5. In `ColumnRow` use `justify-content: space-between` instead of
`@include display.space-between` on larger screens. [@include
display.space-between](https://github.com/dfinity/gix-components/blob/ff7f9beb27b15f91fdee2bbe640d6c7dba28d1e2/src/lib/styles/mixins/_display.scss#L1)
comes additionally with `display: inline-flex` and `width: 100%`. We
don't need these extra styles and the `display: inline-flex` has the
strange effect that the transaction row containing the To/From and date
gets taller when the To/From slot isn't there. You can see it in the
video in [this
comment](https://github.com/dfinity/nns-dapp/pull/3645#discussion_r1378616357).
I spent several hours debugging this and I still don't understand why
this happens. But removing the include seems cleaner than rendering a
zero-width space when there is no To/From. `ColumnRow` is only use in
`TransactionCard` and in `CardInfo`. `CardInfo` is used in many places
but as far as I can tell, the `ColumnRow` is only relevant to it when it
has an `end` slot, which it only has in `NeuronCardContainer` and in
`SkeletonCard`. But `NeuronCardContainer` itself isn't used anywhere
with an `end` slot. I've looked at `SkeletonCard` and it still looks
fine to me.
6. Remove now unused `"ckbtc_transaction_names"` from `i18n`.
7. Set `uiTransaction.otherParty` to "BTC Network" in
`mapCkbtcTransaction` for Mint and Burn transactions.
8. Remove `getDescription` from the page object and only check
`getIdentifier` in tests.

# Tests

Unit tests updated and passing.
Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary