Pass UiTransaction[] to IcrcTransactionList (#3828)
# Motivation

In CkBTCTransactions list I want to render multiple transactions (such
as Approve + transfer) as a single transaction.
This is easier if the conversion from `IcrcTransactionData` to
`UiTransaction` happens closer to the CkBTCTransactionsList.

In this PR I move the conversion from `IcrcTransactionData[]` to
`UiTransaction[]` from `IcrcTransactionsList` to `CkBTCTransactionsList`
and `SnsTransactionsList`.

# Changes

1. In `CkBTCTransactionsList` convert `IcrcTransactionData[]` to
`UiTransaction[]` and pass that to `IcrcTransactionsList`. No longer
pass `account` and `token` which were only passed to do the conversion.
2. In `SnsTransactionsList` convert `IcrcTransactionData[]` to
`UiTransaction[]` and pass that to `IcrcTransactionsList`. No longer
pass `governanceCanisterId`, `account` and `token` which were only
passed to do the conversion.
3. Remove the conversion step an unused props from
`IcrcTransactionsList`.
4. Drive-by: Remove `fallbackDescription` from `UiTransaction`, which I
thought I did in a previous PR.

# Tests

1. Removed irrelevant tests from
`frontend/src/tests/lib/components/accounts/IcrcTransactionsList.spec.ts`.
2. Added `createMockUiTransaction` for convenience.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary