Draft Migration test (#3831)
# Motivation
Testing migration is very similar to the downgrade-upgrade test. The
downgrade-upgrade test can almost certainly be expanded to make it
appropriate for this use case. However the schema change tests are
likely to need a lot of modification over the next few weeks, so I would
prefer to have a separate test for schema changes that can "move fast
and experiment" and then migrate useful code back into the original,
eventually deleting the experimental code.

Why not keep the test in a branch? Most development is on Rust code.
Importing tests into every rust development branch has been a real pain.

# Changes
* Make a copy of the downgrade upgrade test dedicated to schema changes:
  * Copy the test script
  * Mark the new test as unstable.
  * Split the test into small chunks, as the file is huge!
  * Add argument linters & backtraces to help with debugging.
* Adapt the command line arguments to make them suitable for schema
migration tests:
* Previously the downgrade-upgrade test was always between the default
current build and prod. Added flags to allow the downgrade-upgrade test
to work against any pair of wasms and argfiles. This way, we can e.g.
upgrade to self, just changing the schema in the schema.
* Add flags to make it easy to change the schema without writing a new
argfile. We will need this a lot! :-)

# Tests
* Run locally. Marked the most urgent missing unit tests with TODOs in
the code.

# Todos

- [x] Add entry to changelog (if necessary).