Launch ENABLE_CKBTC_ICRC2 (#3832)
# Motivation

The ICRC-2 flow for ckBTC has been implemented for a while.
I've been holding off enabling it because I wasn't sure if it's OK to
show the "Approve transfer" as a separate transaction.
In a meeting it was just decided that this is fine.

# Changes

1. Set `ENABLE_CKBTC_ICRC2` to true by default.
2. Update config test with `scripts/nns-dapp/test-config --update`

# Tests

Tested manually many times and also with ckTESTBTC on mainnet.
Tested once on mainnet with BTC.
https://drive.google.com/file/d/1_5p7JADy7Tg_FwnQksFuSr2j7WJjA_aq/view

# Todos

- [x] Add entry to changelog (if necessary).