Add pending UTXOs store (#3837)
# Motivation

We want to show pending incoming BTC deposit transactions.
This PR just adds a store to store these transactions after we receive
them from the minter so that we can display them in the wallet.

# Changes

1. Adds a store for pending UTXOs. There is a separate list per
universeId (ckBTC vs. ckTESTBTC). UTXOs are sorted by number of
confirmations so that they will be listed with newer transactions above
older transactions.
2. Unit tests for the store.

# Tests

Unit tests pass.
Tested manually in a branch which uses the store.

# Todos

- [ ] Add entry to changelog (if necessary).
will add when it's used