GIX-2117: Load ckETH on Accounts page (#3868)
# Motivation

Load ckETH and ckETHSepolia (if needed) in Accounts page.

# Changes

* Use services `loadCkETHCanisters` and `loadIcrcAccounts` in Accounts
page.

# Tests

* Test that the store is filled with proper data in Accounts route.

I added this test because the Accounts page doesn't render the token
selector sidebar. Therefore, I can't test that the token is present in
the UI from that component.

This will change with the new Tokens page. Therefore, there is no need
to improve it for now.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary