GIX-2122: Send CkETH Tokens (#3897)
# Motivation

User can send ICRC Tokens

It's a fairly big PR, but the functionality was better tested once all
was in.

# Changes

* Add new event name `nnsIcrcTokenModal` to HTMLAttributes.
* Implement the button to open the Send modal.
* New component IcrcTokenAccountsModals to render the Icrc Token modals.
For now only `IcrcTokenTransactionModal`. The type of the data needed
differes from ckBTC and SNS; therefore, I couldn't easily reuse any. I
bet we can further improve this, but in another PR.
* New IcrcTokenTransactionModal that renders a TransactionModal and
makes the ICRC Token transfer.
* Render IcrcTokenAccountsModals in Accounts
* New icrc accounts service `icrcTransferTokens`. There is some
duplication with ckBTC. We can improve this further.
* New types for icrc-accounts modals.
* New modal util `openIcrcTokenModal`.

# Tests

* Test new IcrcTokenTransactionModal
* Test that users can perform transactions from Accounts.
* Test new ICRC service `icrcTransferTokens`.
* Add new PO IcrcTokenTransactionModalPo
* Add necessary methods to POs

# Todos

- [ ] Add entry to changelog (if necessary).