Map the list of transactions instead of individual transactions (#3899)
# Motivation

We want the ckBTC wallet component to have full control over the
transactions list in order to:
1. Include pending transactions mapped from `PendingUtxo`s
2. Merge some transactions into a single transaction, such as
Approve+Transfer or Failed+Reimbursed.

This control was accomplished with
https://github.com/dfinity/nns-dapp/pull/3828 by passing a
`UITransaction[]` from `CkBTCTransactionsList` to
`IcrcTransactionsList`.

Unfortunately, before we could make use of this, the effort was reversed
in https://github.com/dfinity/nns-dapp/pull/3884 by passing only
`mapCkbtcTransaction` (which maps individual transactions) from
`CkBTCTransactionsList` to `IcrcWalletTransactionsList`.

In this PR we, once again, give `CkBTCTransactionsList` control over the
full list by passing a function that maps a list of transactions instead
of passing a function that maps individual transactions.

This is a pure refactoring. The PR does not yet do anything other than
mapping each individual transaction.

# Changes

1. In `IcrcWalletTransactionsList` remove the `mapTransaction`
(singular) prop and add the `mapTransactions` (plural) prop. It gets a
default value which does the mapping that was already the default in
that component.
2. From `CkBTCTransactionsList`, pass a function which maps all
transactions through `mapCkbtcTransaction`.

# Tests

Existing tests pass.
Added a unit test for `IcrcWalletTransactionsList` that tests that the
custom mapping function is used. It previously did not have such a test
for the singular mapping function.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary