feat: review load Icrc accounts and extract common Icrc accounts page (#3891)
# Motivation

This PR addresses two observations:

1.  When the query+update call is activated:

When a user access the accounts, the balance of the universes are
fetched with uncertified calls for performance reason.

However, the selected universe is fetched with query+update for security
reason.

Moreover, if a universe is selected, we try to avoid to perform both -
i.e. not query+query+update but only query+update.

This PR implements those features for the new Icrc/ETH accounts.

2. Similar to WalletPage we can extract a single common page for all
Icrc token accounts related page, that way the logic is reused and
ultimately we can refactor Snses

# Notes

The loading of the accounts was adapted but, components wise, no
functional changes. The logic and display of the component is extracted
to a new component.

# Changes

- use `uncertifiedLoadAccountsBalance` to load balance in token selector
and perform same logic to load the information only when needed
- extract `IcrcAccountsPage` from `CkBTCAccounts`
- use this new component in `CkBTCAccounts`
- create `IcrcTokenAccounts` component also use and inherit similar code
and logic