Feat: Add mainnet cketh canister ids (#3901)
# Motivation

Change the canister ids for ckETH related canisters. It was using the
Sepolia ones.

# Changes

* Change canister ids in `cketh-canister-ids.constants`.

# Tests

* Enable mock with both ckETH universes.
* Enable tests that had ckETH and ckETHSEPOLIA together.
* Set `ENABLE_CKTESTBTC` to false in Accounts to avoid the unwanted
calls regarding `ckETHSEPOLIA`. This was not needed before because it
was overriden by the normal ckETH universe id.
* There was an error in cketh-canisters services spec that appeared when
the ckETH and ckETHSEPOLIA stopped being the same universe id.
* Fix test in Wallet.spec which was relyig on ENABLE_CKTESTBTC set to
true and ckETHSEPOLIA having the same universe id as ckETH.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary yet.