Add pending UTXOs to store after updateBalance (#3909)
# Motivation

In order to get a BTC deposit credited as ckBTC, we need to call
`update_balace` on the minter.
If (and only if!) there are no new deposits with enough confirmations,
the minter returns (as an error) any deposits that do not yet have
enough confirmations.
This means that if there are new completed deposits, we need to call
`update_balace` again to know if there are also any pending deposits.

# Changes

1. In `frontend/src/lib/services/ckbtc-minter.services.ts` call
`update_balance` until we get pending transactions (if there are any)
and then put them in the store.
2. Pass `universeId` to `updateBalance` and to components that call
`updateBalance` because it's used as a key in the store.

# Tests

Unit tests are added.
Test manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/accounts/?u=oq4ec-3maaa-aaaaa-qabeq-cai

# Todos

- [x] Add entry to changelog (if necessary).