tweak: Finer backtrace for migration test (#3934)
# Motivation
The migration test provides backtraces on error. This is useful, however
if we break the long main test function into smaller functions, the
backtrace tells us at which stage the test failed. The backtrace does
provide line numbers but typically just knowing the stage is enough, so
providing the name avoids having to look in the code for the line
number.

# Changes
- Split the migration test into smaller functions.

# Tests
Sample backtrace before:
```
ERROR: Schema is not as expected.
Expected: AccountsInStableMemory
Actual:   Map
=============================================
   at: ./scripts/nns-dapp/migration-test.canister:1: assert_schema_is
   at: ./scripts/nns-dapp/migration-test:104: test_upgrade_downgrade
   at: ./scripts/nns-dapp/migration-test:117: main
```
Sample backtrace after:
```
ERROR: Schema is not as expected.
Expected: AccountsInStableMemory
Actual:   Map
=============================================
   at: ./scripts/nns-dapp/migration-test.canister:1: assert_schema_is
   at: ./scripts/nns-dapp/migration-test:94: install_wasm1
   at: ./scripts/nns-dapp/migration-test:124: test_upgrade_downgrade
   at: ./scripts/nns-dapp/migration-test:131: main
```

# Todos

- [ ] Add entry to changelog (if necessary).
  - Too minor.