Remove ICRC-1 neuron staking logic (#3946)
# Motivation

ICRC-1 transfers use a different memo.
We use the memo when staking an NNS neuron to make sure we can complete
the transaction.
So currently, using ICRC-1 for neuron staking wouldn't be safe.
We should move directly to ICRC-2.

# Changes

Remove the logic for using ICRC-1 when staking a neuron.

The feature flag itself will be clean up, together with some other
unused feature flags, in a separate PR.

# Tests

Corresponding tests have been removed.

# Todos

- [x] Add entry to changelog (if necessary).