fix: CI (#3988)
# Motivation
CI is broken

## Version update test failing
The node version is not specified in one of the tests, so fails now that
GitHub is running a node version that uses an NPM version that is out
the supported range:

```
npm ERR! notsup Required: {"npm":">=9.0.0 <10.0.0","node":">=18.0.0 <19.0.0"}
npm ERR! notsup Actual:   {"npm":"10.2.3","node":"v18.19.0"}
```

## Spellcheck build failing
It looks as if the spellcheck build should work. Perhaps some
dependencies are being updated though as `--locked` is not passed to
`cargo install`.

# Changes
- Install the specified node version for the version update test.
- Lock the spellcheck version.
- Note: There have been no new versions recently but better safe than
sorry.
- Require dependencies to be used as specified in the spellcheck
Cargo.lock file.

# Tests
- See CI

# Todos

- [x] Add entry to changelog (if necessary).