GIX-2057: Add ckBTC logic to Tokens page (#4007)
# Motivation

Users might not understand that they need to update the balance after
converting BTC to ckBTC. Therefore, the NNS Dapp makes the call to the
minter in the background.

In this PR, I add that functionality currently present in the ckBTC
Accounts to the Tokens page.

# Changes

* Call service ckBTC Minter `updateBalance` in the Tokens page.

# Tests

* Add a test that simulates the api returning completed UTXOs and
checking that the balance is updated.

I also tested manually that after getting some test bitcoins, the
balance was never updated even after refreshing. When I introduced the
call, the balance was refreshed when the user visited the page.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.