Fix: Proposal payload rendering (#4021)
# Motivation

A proposal failed to render the payload:
https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=126392

The problem was a [breaking change in the IC for the payload
types](https://github.com/dfinity/ic/blob/master/rs/nervous_system/root/src/change_canister.rs#L137).
One field was removed from two types and also renamed.

Yet, updating the IC depedency is becoming a huge pain. Therefore, to
solve the issue in the short term, I hardcoded the types in the PR
instead of importing them.

# Changes

* Do not import AddNnsCanisterProposal but hardcode it.
* Do not import ChangeNnsCanisterProposal but hardcode it.
* Remove the field `authz_changes` from both `XxxTrimmed` type and
transform functions.

# Tests

* I downloaded the proposal. It was failing withuot this changes, but
with this changes it was working.

I'm trying to make such a proposal in a local environment and test it
there.

# Todos

- [ ] Add entry to changelog (if necessary).