fix: local OWN_CANISTER_ID (#4027)
# Motivation
The nns-dapp configuration is wrong for local deployments: The local
nns-dapp canister ID is overridden.

# Changes
- Use the term `OWN_CANISTER_ID` consistently to make this easier to
debug. Previously sometimes `CANISTER_ID` was used and that gives many
false hits when searching.
- Make the canister IDs provided by dfx override those in configuration
files.

# Tests
- I added a test for teh output of config.sh for local, with two
different deployments of the nns-dapp.
- Note: There is an existing test that covers mainnet and the app
subnet, but these both have fixed canister IDs. The test needed to be
adapted a little to cope with a variable canister ID.
- I read through all uses of these variables I could find and am
reasonably confident on that basis that these changes are correct, not
just that the tests pass. Additional vigilance will also be needed in
the next release to make sure that the arguments are correct there.

# Todos

- [x] Add entry to changelog (if necessary).