A store for BTC retrieval statuses (#4069)
# Motivation

We want to render failed and pending "Sending BTC" transactions as such.
These statuses are not part of the transactions themselves but come from
the ckBTC minter.
We should store them somewhere after fetching them so we can later use
them for rendering.

# Changes

Add a store to put the `RetrieveBtcStatusV2WithId` objects that will be
fetched by the minter in a separate PR.

# Tests

Added unit tests for the store.
Manually tested in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary