GIX-2181: Build wrappers cache before aggregator data (#4054)
# Motivation

More robust NNS Dapp and launchpad.

In this PR, build the SNS wrappers cache before setting the aggregator
data.

This avoids that a component is rendered with the aggreagator data and
triggers a call to an SNS canister before the wrappers' cache is set.
That would mean that also a call to list_sns_canisters of the root is
triggered.

# Changes

* Move the line `snsAggregatorStore.setData(aggregatorData);` after
`buildAndStoreWrapper` in `loadSnsProjects`.

# Tests

* Add a test for `loadSnsProjects` that when `wrapper` is called in the
subscribe of the snsAggregatorStore, it doesn't trigger a call to
`listSnses`.

I run the tests without the change in loadSnsProjects and they failed.

# Todos

- [x] Add entry to changelog (if necessary).