GIX-2180: Move IcpTransactionModal (#4071)
# Motivation

Users can open the Send ICP modal from the tokens table in the ICP
tokens page.

In this PR, I move the IcpTransactionModal from the NnsAccountsFooter to
the AccountsModals.

This will make it easy to then open it when clicking in the icon of the
tokens table.

# Changes

* Add `"nns-send"` to AccountsModalsType.
* Remove IcpTransactionModal from NnsAccountsFooter and dispatch an
event instead.
* Add IcpTransactionModal to AccountsModals.

# Tests

* Added a test in the Accounts route that the user can send ICP by
clicking in the button of the footer.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.