GIX-2183: Unnecessary error toast (#4087)
# Motivation

In mainnet, when the user visits twice the Tokens page a toast error
might appear because of the background task to call update_balance on
the ckBTC minter.

In this PR, do not show the toast error if `uiIndicators` is set to
`false`.

# Changes

* In `updateBalance` ckbtc minter service, show error toast only when
`uiIndicators` is `true`.

# Tests

* Add a test for the new functionality.
* Tested manually that when the api endpoint raises this error the toast
is not shown.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.