GIX-2184: Add accountsTitleStore (#4090)
# Motivation

Change the title of the accounts page to have the name of the token on
it.

This is a change requested regarding the new tokens page, but I believe
that it's a good change even for the current accounts page. Therefore, I
didn't put it behind the feature flag.

In this PR, I introduce the store that I want to use in the layout to
set the title. But I don't use it yet.

# Changes

* New store `accountsTitleStore`.
* New i18n key

# Tests

* Test new store `accountsTitleStore`.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.