GIX-2184: Fix padding icp tokens page (#4089)
# Motivation

There was a visual bug in the ICP Tokens page because the outer paddings
weren't the same as in the Tokens page. This bug was visible in mobile
devices.

The problem came from an extra padding added by a `main`'s default
padding inside the Accounts route. This `main` will be removed when the
flag is removed, but for now, I fixed it by passing a class and setting
the padding to 0 when the flag is enabled.

# Changes

* Add a new class `noPadding` to the `main` in Accounts route and set it
when ENABLE_MY_TOKENS is enabled.

# Tests

Only UI changes.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.