Load BTC retrieval statuses in ckBTC wallet (#4111)
# Motivation

We want to render failed and pending "Sending BTC" transactions as such.
These statuses are not part of the transactions themselves but come from
the ckBTC minter.
In this PR we call the service method that loads the statuses and puts
them in the store.
This is the final PR that ties it all together.

# Changes

1. Call `loadRetrieveBtcStatuses` when opening the `CkBTCWallet`
component.
2. Also call `loadRetrieveBtcStatuses` after making a transaction.

# Tests

2 unit tests added.

# Todos

- [x] Add entry to changelog (if necessary).