Use concise bigint notation in tests (#4123)
# Motivation

Similar to https://github.com/dfinity/nns-dapp/pull/4114 but for test
files.

`bigint`s can be written as `123n` but in many places we write
`BigInt(123)` or `BigInt("123").
The former is more concise and also more accurate if the number actually
requires a bigint for full accuracy.

# Changes

1. Replace `BigInt(???)` with `???n`.
2. Add underscore separators for numbers with more than 3 digits if they
weren't there already.

# Tests

Still pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary