Fix parsing the proposal ID in scripts/nns-dapp/release (#4145)
# Motivation

`scripts/nns-dapp/release` has a flag to write the created proposal ID
to a file.
This functionality depended on `ic-admin` writing only a single line to
stdout.
But newer versions of `ic-admin` write more to stdout, breaking the
proposal ID parsing.

# Changes

`grep` for the specific line containing the proposal ID and ignore
everything else.

# Tests

Tested manually on local with
```
$ PATH=$PATH:../../snsdemo/tree2/bin scripts/nns-dapp/release --save-proposal-id-to-file tmp-id --network local --neuron "$(cat ~/.config/dfx/identity/snsdemo8/neurons/local)"

$ cat tmp-id
```

# Todos

- [x] Add entry to changelog (if necessary).