Remove double scroll container from My Tokens table (#4194)
# Motivation

The `main` element of the My Tokens table is a scrolling container, but
it exists in another scroll container.
This results in a strange cut-off at the top and bottom and strange
placement of the scroll bar.
See https://github.com/dfinity/nns-dapp/pull/4182 for details.

# Changes

1. Remove `height` and `overflow-y` from the `main` element.
2. Remove custom padding that was necessary because of the double scroll
container.

# Tests

Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary