Refactor: GetTokensModal (#4188)
# Motivation

With the new tokens view, we don't have a universe anymore in the tokens
view. Therefore, the button "Get Xxxx" is always "Get ICP". I want to
add a dropdown of all the available tokens within the modal.

Before doing that and to avoid checking the balances of all the test
accounts always, I refactored the Modal into another component so that
then, the balances of the tokens is checked only when the modal is open.

In this PR, I'm only moving the modal to another file. The main changes
will come in a subsequent PRs.

# Changes

* Create a new GetTokensModal and use it in GetTokens instead of
rendering the Modal directly.

# Tests

This is a dev feature only. There are no tests.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.