Fix BTC address QR code spinner in dark theme (#4197)
# Motivation

When using the `Spinner` component, it renders in the current color.
By default, this is dark in the light theme and light in the dark theme.
But the QR code for the BTC deposit address is always rendered on a
white background, so it should always have a dark spinner.

# Changes

Set `color` to `black` in the `qr-code` element.

# Tests

Tested manually:

<img width="788" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/f6d1b2f3-aae7-457e-9b40-735902ea3827">

# Todos

- [x] Add entry to changelog (if necessary).