GIX-2187: Make ckBTC wallet renderable when not signed in. (#4198)
# Motivation

We want to be able to show wallet pages when the user is not logged in.
This PR makes it possible to render the `CkBTCWallet.svelte` page when
the user is not logged in.
This is not a user visible change because `wallet/+page.svelte` still
doesn't render any wallet when the user is not logged in.

# Changes

1. Hide Send/Receive buttons.
2. Don't load data.
3. Don't require `nonNullish($selectedAccountStore.account)` to show the
info card.
4. Show transaction list placeholder.

The rest of the functionality comes from reusing
`IcrcWalletPage.svelte`.

# Tests

Unit tests added.
Also added missing tests for Send/Receive button for `IcrcWalletPage`
that should have been in the [previous
PR](https://github.com/dfinity/nns-dapp/pull/4196).
Test manually in a separate branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet