GIX-2187: Render wallet page when not signed in. (#4208)
# Motivation

People should be able to see the wallet page while not being signed in.
In particular, this will allow us to deep link to the ckBTC wallet so
that people can read info about ckBTC and then convert after signing in.

# Changes

In `frontend/src/routes/(app)/(u)/(detail)/wallet/+page.svelte` always
render `Wallet` and never `SignInAccounts`.

# Tests

The existing unit test only checked that a sign-in button was rendered.
It still passed but because the wallet also renders a sign-in page.
The current component doesn't have any logic anymore, except for passing
on the account identifier.
So I just added a unit test to check that the account identifier makes
it into the wallet component.

Also tested manually
[here](https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/).
You can navigate to a wallet and then log out, or use the deep links for
[ICP](https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/wallet/?u=qsgjb-riaaa-aaaaa-aaaga-cai),
[SNS](https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/wallet/?u=be2us-64aaa-aaaaa-qaabq-cai),
[ckETH](https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/wallet/?u=omy6t-mmaaa-aaaaa-qabgq-cai)
and
[ckBTC](https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/wallet/?u=oz7p6-neaaa-aaaaa-qabfa-cai)
(all different implementations).

# Todos

- [x] Add entry to changelog (if necessary).