Console feature to temporarily add ICRC-1 token (#4222)
# Motivation

Sometimes people accidentally send tokens to their NNS dapp address
while those tokens are not supported by NNS dapp.
It would be nice if there was a way to recover those tokens.

The experience could definitely be improved but I don't want to spend
more time on this now, so please review based on whether this is an
improvement or not.

# Changes

1. Provide a hard coded logo for unknown tokens.
2. Add a method to the `window` object to add a ledger and index
canister to the store.

# Tests

1. Adjusted an existing unit test for an unknown token.
2. Tested manually adding an unknown ledger.

Currently it happens to work with and without index canister but there
are no guarantees:
```
__experimentalAddIcrc1Token('m7h5t-euaaa-aaaaa-qabia-cai');
__experimentalAddIcrc1Token('m7h5t-euaaa-aaaaa-qabia-cai', 'myg3h-jmaaa-aaaaa-qabiq-cai');
```
<img width="991" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/7c4a2981-e8a6-4093-87ce-ecaa682eeb2c">


# Todos

- [x] Add entry to changelog (if necessary).