GIX-2063: Add link to tokens table for not logged in (#4235)
# Motivation

User can navigate between tokens, wallet and accounts when not logged
in.

In this PR, the rows of the tokens page redirect to different pages when
the user is not logged in.

# Changes

* Add `rowHref` in `tokensListVisitorsStore `.
* Remove the part that forced a sign in from the tokens route when a
tokens action was triggered but the user was not logged in. No further
handling is needed because the navigation uses a normal anchor tag.

# Tests

* Add the new `rowHref` in the test file for `tokensListVisitorsStore`.
* Change test `"should click on a row should trigger a login"` to
`"should render an anchor tag with href"` for Tokens page when user is
not logged in.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.