Change wording about staking neurons to staking tokens (#4244)
# Motivation

We use the phrasing "Stake neuron" in the UI. But you actually stake
tokens to create a neuron.
We want to change the text int he UI to reflect this.

# Changes

1. Change the text on "Stake Neurons" button to "Stake ICP" for NNS or
"Stake $token" for SNS.
2. Change the modal title from "Stake Neuron" to "Stake ICP" and from
"Stake $token Neuron" to "Stake $token".
3. Change error message about "needing a minimum" to "needing to *stake*
a minimum"
4. Remove `i18n.auth.stake` and `i18n.error__sns.sns_neuron_account` and
`i18n.error__sns.sns_amount_not_enough_stake_neuron` which are unused.
5. Change the transaction title from "Stake Neuron" to "Staked" to be
consistent with "Sent" and "Received" which are also perfect tense
verbs.

# Tests

Added tests that the stake button and modal contains the token name.
The other changes are text changes without logic.

# Todos

- [x] Add entry to changelog (if necessary).