GIX-2150: Use TokenAmount for fee in SnsNeuronDetail (#4272)
# Motivation

Use TokenAmountV2 instead of TokenAmount.

In this PR, change the type of the fee in SnsNeuronDetail to TokenAmount
from bigint. This will allow us to then migrate to TokenAmountV2.

The following changes were triggered in [this
comment]([url](https://github.com/dfinity/nns-dapp/pull/4215#discussion_r1455261670)).

# Changes

* Set local variable `transactionFee` type to `TokenAmount` instead of
`E8s`.
* Fix the necessary props in the way.

# Tests

* Fix the tests with the new prop type and name.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.