preparation: Implement state traits manually (#4278)
# Motivation
We would like to restrict changes to stable memory, so that such changes
are made only as part of updates to the state structure. We will do this
by adding a reference to memory in state and making changes only via
that pointer. However we will need to implement traits such as Debug and
Eq manually.

# Changes
- Implement Debug manually.
- Alternative considered: Implement the Debug trait for stable memory.
Rejected as the Debug printout is already much too verbose. It seems
better to limit the debug printout to what is actually useful.
- Implement `Default` manually, so that the default can later include
the stable memory.
- Implement `Eq` manually, as I would like to ignore changes to the
stable memory pointer when comparing state e.g. pre and post upgrade. I
would like to restrict access to memory but carry on not caring how the
state is represented, as long as the meaningful content of the state is
the same.

# Tests
- Tests are unchanged.

# Todos

- [x] Add entry to changelog (if necessary).