Feat: Add fee to sendICP and stakeNeuron (#4362)
# Motivation

Remove proto dependency.

When the fee of a transaction is not passed, the LedgerCanister will get
the fee with a request because the fee is mandatory. The hardware wallet
doesn't support the action of getting the ledger fee; therefore, that
call fails and also the whole transaction call.

In this PR, I added the `fee` as mandatory in both `sendIcp` and
`stakeNeuron` api functions.

# Changes

* Add `fee` in canister api functions that call `sendICP`.
* Add `fee` to `stakeNeuron` governance api function.
* Add `fee` to `sendICP` icp-ledger api function.
* Add `fee` in canister services.
* Add `fee` in icp-account service `transferICP`.
* Add `fee` in call to `sendICP` inside the governance service
`stakeNeuron`.
* Add `fee` in call to `sendICP` inside the sns-sale service
`participateInSnsSale `.

# Tests

* Add `fee` to the api functions that needed it and the expects.
* Add `fee` to the expect data in Accounts.spec
* Add `fee` to the expect in the services specs.

# Todos

- [x] Add entry to changelog (if necessary).