GIX-2189: Add token in Send transactions (#4365)
# Motivation

With the new tokens page the modals are not so clearly defined within a
universe. To help the user be aware of the transactions, design proposed
to add the token symbol in the Send and Receive modals' title.

In this PR, add the token symbol i the title of Send transactions,
except ckBTC.

# Changes

* New i18n key "core.send_with_token".
* Use new i18n for the title of the first step in IcpTransactionModal.
* Use new i18n for the title of the first step in
IcrcTokenTransactionModal.
* Use new i18n for the title of the first step in SnsTransactionModal.

# Tests

* Add a test to check the modals's title in IcpTransactionModal.spec.
* Add a test to check the modals's title in
IcrcTokenTransactionModal.spec.
* Add a test to check the modals's title in SnsTransactionModal.spec.

# Todos

- [x] Add entry to changelog (if necessary).