GIX-2189: Add token label in ckBTC send modal (#4376)
# Motivation

Help users understand the context better by adding the token symbol in
the title of the send modals.

In this PR, add the token label in the ckBTC transaction modal.

# Changes

* New helper `getTitleForm` inside CkBTCTransactionModal to select the
title in the first step depending on the network and the universe id.

# Tests

* Add test id to TransactionFormItemNetwork and create its page object.
* Add two tests to CkBTCTransactionModal.spec. One to check the ckBTC
label and the other one when the network is BTC.


# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary. There is already a changelog entry for this.