Use topic (sub)title i18n messages directly (#4391)
# Motivation

I want to set up a test that makes sure there are no unused i18n
message.
For this I'm trying to make it easy to find if i18n messages are used.
For this I'm trying to reduce the dynamic usages of i18n messages.

In this PR, I change the dynamic usage of
`i18n.follow_neurons["topic_${topic}_title"]` to direct usage of
specific messages.

# Changes

1. Add `getTopicTitle` and `getTopicSubtitle` which each have a map from
`Topic` to specific messages.
2. Use these functions in `Followee.svelte` and
`FollowNnsTopicSection.svelte`

# Tests

Unit tests added

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary