Use ledgerCanisterId instead of universeId (#4399)
# Motivation

For SNS universes we use the rootCanisterId as the universeId.
For most other universes, we use the ICRC ledger canister ID as the
universeId.
To reduce code duplication we want to use the ICRC wallet components for
the SNS wallet.
The ICRC wallet components assume the universe ID is the ledger canister
ID and so the use both terms to mean the same.
But if we reuse those components for SNS there is a clear difference
between universe ID and ledger canister ID.

So in this PR I'm renaming `universeId` to `ledgerCanisterId` in places
where it's relevant for reuse by SNS.

The PR is large but it's split up into many commits which are each
self-contained.
So it might be better to review per commit.

# Changes

Rename parameters and props to refer ledger canister IDs instead of
universe IDs.

# Tests

Unit tests updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary