Make universeId prop optional on ReceiveModal (#4404)
# Motivation

The `ReceiveModal` component has a prop for the `universeId`.
This is only used to list the available accounts when `canSelectAccount`
is `true`.
But in most cases `canSelectAccount` is not set to true and `universeId`
is ignored.

When we merge the ICRC and SNS wallet components, it can be confusing
whether to use the `ledgerCanisterId` or the `rootCanisterId` where
previously a `universeId` was required. So to avoid confusion, I'm
removing the `universeId` on the `ReceiveModal` when it's not necessary.

# Changes

1. Make `universeId` optional on the `ReceiveModal` and related
components, events and functions.
2. Only render the accounts picker when `canSelectAccount = true` and
`universeId` is present.

Drive-by:
1. Add curly brackets on a multi-line function for convenience.
2. Fix the name of `getSelectAccountDropdownPo` to correctly reflect
what it returns.

# Tests

Added a unit test with `universeId` absent

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary