Fix: check SNS neuron balances if force query (#4415)
# Motivation

There is a bug in mainnet.

If a user starts staking an SNS neuron, but loses connection after the
transfer happens but before the claiming happens, the neuron isn't
claimed when the user returns.

The reason was because we were checking the balances only on update
calls. We lost that when we forced only query calls with
`FORCE_CALL_STRATEGY`.

# Changes

* Check SNS neuron balances also if `FORCE_CALL_STRATEGY` forces only a
`"query"` strategy.

# Tests

* I added the same test to claim a neuron forcing the strategy to
`"query"`, the test failed. As expected with the bug.
* Then I added the fix to also check when `FORCE_CALL_STRATEGY` is
`"query"` and it passed.

# Todos

- [x] Add entry to changelog (if necessary).