Fix missing ballots issue (#4420)
# Motivation

Fix the issue when all user neurons are shown as ineligible to vote.

**HowTo reproduce:**
1. goto sns proposal list page
2. log-in
3. refresh the page
4. wait for 10 seconds (otherwise there could be no certified proposals
in the `snsProposalsStore` store)
5. click on any proposal that you are eligible to vote on
6. check the ineligible neurons block

The problem is the calculation of the `reloadForBallots` flag that
prevents the dapp from redundant proposals fetching.
The proposal call is skipped when:
- `reloadForBallots` is false (when no neurons available)
- there is a `certified` version of the proposal data in the
`snsProposalsStore`

If this logic is applied before the neurons being loaded, but after the
`snsProposalsStore` being updated, the wrong neurons state is shown on
the proposal page, because no ballots were loaded.

# Changes

- before loading full sns proposal details

# Tests

Not in this pr.

# Todos

- [x] Add entry to changelog (if necessary).