GIX-1985: Remove my tokens flag NnsAccounts page (#4422)
# Motivation

Remove `ENABLE_MY_TOKENS` flag.

In this PR, remove the flag from NnsAccounts.

# Changes

* Remove the branch where `ENABLE_MY_TOKENS` is false in NnsAccounts.

# Tests

* Remove the tests of the branch were `ENABLE_MY_TOKENS` is false except
for one moved to when `ENABLE_MY_TOKENS` is true. But no changes were
needed.
* Change setup needed for the test. I changed resetting the accounts
store by mocking a return of the NNS Dapp API. The NNS Accounts expects
specific props that for testing make more sense to be passed. The logic
of manipulating the data is in the parent Accounts route and therefore,
it should also have the logic of fetching the data.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.