Add selectedTokenStore (#4449)
# Motivation

SNS tokens are in the `tokensStore` keyed by `rootCanisterId` and now
also by `ledgerCanisterId`.
We want them only keyed by `ledgerCanisterId` so we need to change all
the places that access by `rootCanisterId`.

In some places we use
`$tokensStore[$selectedUniverseIdStore.toText()].token` to get the token
for the current universe.

In this PR we provide a convenient `selectedTokenStore` which gives the
token for the selected universe.
This can then, in the next PR, be changed to use the SNS
`ledgerCanisterId` instead of `rootCanisterId`.

# Changes

1. Add `selectedTokensStore` which holds the token for the selected
universe.
2. Use `selectedTokensStore` instead of
`$tokensStore[$selectedUniverseIdStore.toText()].token` in
`SnsNeuronPageHeading.svelte` and `SnsActiveDisbursementsModal.svelte`.

# Tests

Unit tests added for 

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary