NNS1-2849: Add derived token stores (#4461)
# Motivation

`tokensStore` used to be keyed by universe ID but in order to reuse the
IcrcWalletPage for the SnsWallet, it needed to be keyed by ledger
canister ID. But there are other context where it is still more
convenient to access the token by universe ID or root canister ID. It is
currently temporarily keyed by both, which means it has the same tokens
multiple times. But converting root canister ID to ledger canister ID
every time would be tedious.
Additionally, the SNS tokens data already exists in the aggregator data,
which means we now have several sources of truth for the tokens data.

The clean solution is to have 1 source of truth, and derived stores for
everything else.

So with this PR:
* `tokensStore` only holds non-SNS tokens
* `snsAggregatorStore` becomes the only source of truth for SNS tokens
* several derived tokens stores are added to be used to actually get a
token by universe ID, root canister ID or ledger canister ID.

# Changes

1. Stop adding SNS tokens to `tokensStore` in both
`lib/services/$public/sns.services.ts` (for production) and
`tests/utils/sns.test-utils.ts` for testing.
2. Add derived stores `snsTokensByRootCanisterIdStore`,
`snsTokensByLedgerCanisterIdStore`, `tokensByUniverseIdStore` and
`tokensByLedgerCanisterIdStore`.
3. Use the derived stores as appropriate in places where the
`tokensStore` was used directly before.
4. In `loadIcrcToken` and `loadToken`, don't load the token if it's a
known SNS token as it should already be available in the appropriate
derived stores.
5. Support setting a `fee` in the aggregator test token data.
6. Make the `lifecycle` optional in the aggregator test data.

# Tests

1. Unit test for the added derived stores.
2. Unit test that `loadIcrcToken` doesn't load SNS tokens.
3. Existing unit tests updated to rely on token data from the aggregator
data.

# Todos

- [x] Add entry to changelog (if necessary).