Chore: Remove unnecessary busy screens (#4468)
# Motivation

Clean up unused code.

Looking at unused busy screen initiators a bug was found because the
busy screen started was not the same being stopped in the FilterModal.

The bug was never encountered because the filters were always present
when rendering the FilterModal.

Yet, in this PR we use a different approach and remove the unused
initiators. Instead of showing a whole busy screen that won't allow the
user to do anything, we show a spinner within a modal.

# Changes

* Remove using busy screens in FilterModal and instead render a modal
component with a spinner if the filters are not present.

# Tests

* Add a test when the filters are not present.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.