Test that we show the project logo on the token wallet (#4476)
# Motivation

The SNS token logo in the ledger canister metadata can be different from
the SNS project logo.
Currently in the wallet, we show the project logo but you could argue
that we should show the token logo instead because the wallet is about
tokens and also the wallet should probably act the same for SNS tokens
as for non-SNS tokens.

But before we fix this, we want to make sure that SNSes can change their
token logo if they want to.

So in this PR we add a test to make sure that we don't accidentally fix
this until we should.

# Changes

1. Add a test to make sure the logo displayed on the SnsWallet is the
project logo and not the token logo.
2. Support setting the token logo in `setSnsProjects`.
3. Support getting the token logo URL in page object.

# Tests

Tested by setting `logo: summary.token.logo` instead of `logo:
summary.metadata.logo` in `frontend/src/lib/utils/universe.utils.ts` (as
can be seen in the first commit) and seeing the test fail.

# Todos

- [x] Add entry to changelog (if necessary).