Add snsLedgerCanisterIdsStore (#4480)
# Motivation

In order to deprecate `snsAccountsStore` we'll need to know the ledger
canister ID from the root canister ID in many places.

The `snsLedgerCanisterIdsStore` was introduced in
https://github.com/dfinity/nns-dapp/pull/4477 but that PR is blocked
because we still need to load SNS accounts after getting the tokens.

So I split this PR from that PR.

# Changes

Add a derived store which maps root canister IDs to ledger canister IDs.
This is not used yet.

# Tests

Add a unit test for the derived store.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary