Simplify universesAccountsBalance derived store (#4482)
# Motivation

Readers of `universesAccountsBalance` never look at whether the data is
`certified` and if we don't need to include that field, the store can
trivially be derived from `universesAccountsStore`.

# Changes

1. Change `UniversesAccountsBalanceReadableStore` to just hold the
balance and not whether it's `certified`.
2. Change the logic from assembling the data from 3 different accounts
store to simply derive it from `universesAccountsStore`.
3. Change readers of `universesAccountsBalance` to use the value
directly instead of having to get the `balanceUlps` field from it.
4. Remove `sumNnsAccounts` which became unused.

# Tests

updated

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary