Delete getSnsTokens (#4477)
# Motivation

We have a test feature to get free tokens.
There is a function `getSnsTokens` to support this for SNS tokens.
But SNS tokens are just ICRC tokens so we can use `getIcrcTokens`
instead.

# Changes

1. Add a derived store which maps root canister IDs to ledger canister
IDs. (This will probably be used in more places to unify icrc accounts
with sns accounts.)
2. Use `getIcrcTokens` instead of `getSnsTokens`.
3. Delete `getSnsTokens` and `acquireSnsTokens` which are now unused.

# Tests

Added a unit test for `snsLedgerCanisterIdsStore`.

Tested getting tokens (ckETH and SNS) manually.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary