Fix: Bug with neuron appearing (#4484)
# Motivation

A bug was raised in the forum where an SNS neuron was appearing even
though the user transferred ownership of such a neuron.

The issue came from the background task that checks neuron subaccounts.
The background task didn't consider whether the user has still some
permission with the neuron and refreshes it and adds it to the store.

The fix is to try to query the neuron from the canister directly and
then check the permissions before continuing with refresh or not
refresh.

# Changes

* Add code in the main `checkNeuronsSubaccounts` to fetch the neuron
when it wasn't passed and compare the permissions with the user's
identity.

# Tests

* Add two tests regarding the bug. One simulating the bug and another
one with an edge case to confirm that if a neuron without permissions is
found, it doesn't break the loop of claiming neurons.
* Fix the tests in sns-neurons that were broken due to expecting the
user to be the controller of the neuron.
* Add a missing test when a neuron that wasn't created in NNS Dapp
needed to be refreshed.

# Todos

- [x] Add entry to changelog (if necessary).