feat: Functions to save and restore the heap to managed memory (#4423)
# Motivation
On upgrade, we will need to be able to save to and restore from managed
memory, much as we save to and restore from raw memory now.

# Changes
- Add functions to save to and restore from managed memory.
- Add code that calls the appropriate save or restore function,
depending on schema.

# Tests
- A unit test is included for managed memory.
- Raw memory using the old API can run only inside a canister, so that
is excluded from the unit test. There is an end to end test that
exercises upgrades and downgrades, though, and that checks that accounts
are preserved.

# Todos

- [ ] Add entry to changelog (if necessary).