Add tests for env-vars (#4515)
# Motivation

`lib/utils/env-vars.utils.ts` was mocked in `vitest.setup.ts` and
because of that it couldn't be tested and thus it had no unit tests.
But using `vi.unmock` we can unmock the module just for one test.

# Changes

Add unit tests for `lib/utils/env-vars.utils.ts`.

# Tests

only tests

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary