NNS1-2868: Load transactions in NnsWallet (#4544)
# Motivation

Fetch the ICP transactions from ICP Index canister not the NNS dapp
canister.

In this PR, load the transactions in NnsWallet from ICP. Only loads the
first page.

In upcoming PRs: loading and paging functionality.

# Changes

* Load the transactions from ICP Index in NnsWallet as well as the
neurons.
* Create a list of UiTransction from the Index type and pass it to
UiTransactionsList in NnsWallet.
* New derived store `neuronAccountsStore`.

# Tests

* Add a test that the transactions are rendered with the flag enabled.
* Add a test that a "stake" transaction is rendered with the flag
enabled. This is to test that NnsWallet also loads the neurons.
* Add a test that a "Decentralization Swap" transaction is rendered with
the flag enabled. This is to test that NnsWallet checks the aggregator
data to calculate transaction types.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.