NNS1-2857: Add icpAccountBalancesStore (#4562)
# Motivation

The current icpAccountStore holds a combination of accounts and balances
in a single store. This means all the data has to be loaded together and
if we just want to change a balances, we have to manipulate the data in
the store, which can interfere with loading the accounts.

The cleaner way is to have a separate store for accounts and a separate
store for balances and a derived store which merged both into accounts
with balances. That way when we update a balance in the balances store,
it will just automatically recreate the accounts with balances.

In this PR, I'm only adding a new store for account balances (separate
from accounts). In a later PR this will be used as described above.

# Changes

Add `icpAccountBalancesStore` which holds a record of `balanceE8s` for
each ICP account identifier of the user.

# Tests

Unit tests added for `icpAccountBalancesStore`.
Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet