NNS1-2857: Add derived icpAccountsStore (#4563)
# Motivation

The current icpAccountStore holds a combination of accounts and balances
in a single store. This means all the data has to be loaded together and
if we just want to change a balances, we have to manipulate the data in
the store, which can interfere with loading the accounts.

The cleaner way is to have a separate store for accounts and a separate
store for balances and a derived store which merged both into accounts
with balances. That way when we update a balance in the balances store,
it will just automatically recreate the accounts with balances.

In this PR, I'm adding the derived `icpAccountsStore` which is derived
from the `icpAccountDetailsStore` and the `icpAccountBalancesStore`. It
is designed to be a drop-in replacement for the existing non-derived
`icpAccountsStore` so it has the same initial value (except for the
`certified` field).

The code for `mapAccount` is mostly copied from the `mapAccount`
function in `icp-accounts.services.ts`
[here](https://github.com/dfinity/nns-dapp/blob/67bedd56e88351825a50c5d03578d5b634067cbf/frontend/src/lib/services/icp-accounts.services.ts#L118-L140).

The store will be used an another PR when the stores that it is derived
from are actually populated.

# Changes

Add derived `icpAccountsStore`.

# Tests

Add unit tests for `icpAccountsStore`.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet

---------

Co-authored-by: Carly Gundy <carlene.gundy@dfinity.org>