NNS1-2857: Utility to (re)set accounts for testing (#4567)
# Motivation

We want to switch from using the direct `icpAccountsStore` from using
the derived `icpAccountsStore`. This means we have to change how the
data is set up in a lot of tests.

In this PR I'm adding a utility to set up the accounts for testing. If
we use that utility in every test, then we only have to change code in 1
place when we change how the tests need to be set up.

# Changes

1. Add test utility functions to manipulate `icpAccountsStore` in
`frontend/src/tests/utils/accounts.test-utils.ts`.
2. Use it in every test that sets up the `icpAccountsStore`.

# Tests

Still pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet