NNS1-2857: Use NNS subaccount in accounts-list.derived.spec.ts (#4576)
# Motivation

This test fails with a refactoring that I'm doing.
But it's using an SNS account as an NNS account so this shouldn't work
in the first place.

# Changes

Use `mockSubAccount` instead of `mockSnsMainAccount` as the NNS
subaccount used in
`frontend/src/tests/lib/derived/accounts-list.derived.spec.ts`.

# Tests

Still passes and like this also passes with the refactoring I have in
another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary