Remove duplicate syncAccounts test (#4579)
# Motivation

These 2 tests are identical, except that one has the setup inside the
test and the other has the setup in the `beforeEach` before it:

https://github.com/dfinity/nns-dapp/blob/c2d47f25488e68194e9945c4873402937f0ae7d5/frontend/src/tests/lib/services/icp-accounts.services.spec.ts#L713

https://github.com/dfinity/nns-dapp/blob/c2d47f25488e68194e9945c4873402937f0ae7d5/frontend/src/tests/lib/services/icp-accounts.services.spec.ts#L398

The newest one was added in [this
PR](https://github.com/dfinity/nns-dapp/pull/1969).
I checked with @lmuntaner and he confirmed that he probably wanted to
remove the one in the `describe("services"`.

# Changes

Remove the duplicate test.

# Tests

removed

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary