NNS1-2857: Reload balance when opening NNS wallet (#4584)
# Motivation

If a transaction happens from a different tab/browser and you reopen the
NNS wallet, you see the new transaction but not the new balance.

# Changes

1. When mounting the `NnsWallet` component, reload the balance of the
account. Only do this if there was already a balance, otherwise
`pollAccounts` will already be loading the balance and we want to avoid
doing it twice unnecessarily.
2. Don't reload transactions every time the accounts store is updated
because it also triggers when the value doesn't change. Instead,
generate a string and only reload transactions when the string changes,
because this does not trigger when the store is set but the string
remains the same. There is a corner case here where the transactions
won't be reloaded if there were multiple transactions that together add
up to the balance not having changed. The right solution to this is to
derive the balance from the transactions store but we can only do that
after we fully migrate to the index canister.

# Tests

Unit tests added and updated.
Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).