Remove ENABLE_ICP_ICRC functionality (#4587)
# Motivation

Feature flag `ENABLE_ICP_ICRC` was added to display ICRC-1 account
identifiers for ICP accounts instead of ICP account identifiers.
But we never figured out a good way to still also access the ICP account
identifier, which would be necessary because transactions to/from are
ICP account identifiers.

Since we haven't done anything with this for 8 months, it's better to
just remove it as it's slowing down development.

# Changes

1. Remove logic conditional on `ENABLE_ICP_ICRC` (added in
https://github.com/dfinity/nns-dapp/pull/2917)
2. Remove type `IcpAccount` which was added for this purpose in the same
PR.

# Tests

Removed and updated.

# Todos

- [x] Add entry to changelog (if necessary).