feat: Async migration (#4611)
# Motivation
If there are several sequential large accounts, migration may fail to
move that batch of accounts to stable memory in one call and the
migration step will be rolled back. The migration will keep trying and
keep failing. As transactions are also processed by the periodic runner,
transactions would also be rolled back and be stalled permanently until
we upgrade. While fairly unlikely given the conservatively small number
of accounts migrated in every step, it is still an unnecessary risk.

# Changes
- Run migration separately from transactions, in a way that does not
cause transactions to fail if migration fails.
- This is done by adding an API method to perform a migration step. That
method is called on the canister itself. That way, if the migration step
panics the periodic task can handle the error.
- Note: The migration method is left exposed with a capped step size. I
think it is not harmful and could be interesting.
- If the normal migration step fails, single step.

# Tests
- The existing migration e2e test still passes.
- If I add a panic to `step_migration()` if the step size is greater
than 1, the migration still succeeds, albeit slowly.

# Todos

- [ ] Add entry to changelog (if necessary).
  Please let me know if this is wanted.  NNS team's rules apply now.