Load ckBTC info in CkBTCTransactionModal (#4627)
# Motivation

When sending BTC directly from the "My Tokens" table, the amount input
field shows an error (unless you've previously visited the ckBTC wallet
page):
<img width="578" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/241c3961-1ac9-4b15-88b3-a53a55922e1c">

This is because the `CkBTCTransactionModal` assumes that the minter info
is already loaded, but this only happens on the ckBTC wallet page.

# Changes

Load the minter info in the CkBTCTransactionModal if it isn't already
loaded.

# Tests

1. Extend the existing unit test that tests that the amount can't be
lower than the minimum amount to verify that the minimum amount
corresponds to what's already in the minter info store.
2. Add an additional unit test that tests that the minter info is loaded
from the API if not present in the store.

# Todos

- [x] Add entry to changelog (if necessary).