NNS1-2918: Add HideZeroBalancesToggle component (#4634)
# Motivation

We want to add a setting to hide tokens with a zero balance.
We're adding a settings button in the top right of the tokens table.
This button will open a popup with a toggle.
This PR adds the component that will be used as the content inside this
popup.

# Changes

Add `HideZeroBalancesToggle` component to be used inside the settings
popup.

# Tests

Added unit test for the component.
Here's what the component will look like in context:

<img width="381" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/54df1e25-5b03-4139-b001-e5777e95fc04">

# Todos

- [ ] Add entry to changelog (if necessary).
not yet