Remove unused GoToTokenDetail action type (#4654)
# Motivation

When the user clicks on a token row, we dispatch an event with type
`ActionType.GoToTokenDetail`.
But there is nothing that responds to this event.
This is not a problem because the token row is a link and the intended
navigation happens through the link `href` without the need to dispatch
an event.

# Changes

1. Remove the click handler and code to dispatch this event.
2. Remove the event type.

# Tests

1. Remove the corresponding test.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary