NNS1-2987: Display block timestamp in transaction history (#4665)
# Motivation

In order to have timestamps consistent with the dashboard, we need to
use
[this](https://github.com/dfinity/ic/blob/03de843e2a01f57a8fb5c2ef928e91574396f6c5/rs/rosetta-api/icp_ledger/index/index.did#L61)
instead of
[this](https://github.com/dfinity/ic/blob/03de843e2a01f57a8fb5c2ef928e91574396f6c5/rs/rosetta-api/icp_ledger/index/index.did#L60).

# Changes

Use the `timestamp` field if available and fall back on the
`created_at_time` field otherwise.

# Tests

Unit test added and tested manually.
Also tested that it doesn't break with an older version of the index
canister that doesn't have the `timestamp` field.

# Todos

- [x] Add entry to changelog (if necessary).