NNS1-2887: Bump: ic-js (proto dependency removal) (#4671)
# Motivation

We removed the requirement to depend on protocol buffers for hardware
wallet transactions from ic-js.

The actual removal of the nns-proto dependency will be done in a
separate PR to keep this PR clean.
But the effect is present with this PR as it reduces the compressed WASM
size by ~80KB.

# Changes

Run `npm upgrade:next`

# Tests

Manually tested that the following still work with hardware wallet:
* transfer
* listNeurons
* increaseDissolveDelay
* startDissolving
* stopDissolving
* disburse
* spawnNeuron
* addHotkey
* removeHotkey

The following don't seem to be available in the NNS dapp UI so I didn't
test them manually:
* joinCommunityFund
* mergeMaturity

# Todos

- [x] Add entry to changelog (if necessary).