Don't sign read-state separately for HW neuron staking (#4673)
# Motivation

In the past, when using a Ledger hardware wallet, you always had approve
twice because the read_state call had to be signed separately.
Then functionality was added such that both requests could be signed
with a single approval.
But there was a bug in the neuron staking flow so just neuron staking
still required 2 approvals.
This was fixed in the Ledger app at some point.
So now that we always require version 2.4.9 of higher, we don't need to
ask for double approval anymore and also don't need to have this depend
on the Ledger app version.

# Changes

Remove the special case for neuron staking and just always sign the
read_state request together with the call request.

# Tests

Unit test was adjusted.
Also tested manually.

# Todos

- [x] Add entry to changelog (if necessary).