NNS1-2995: Redirect from wallet to accounts after sign-in with wrong account identifier (#4687)
# Motivation

If you are on the wallet page for an account that doesn't exist or
doesn't belong to you, you should be redirected to the accounts page.
But the code that did this, was only triggered when the loaded account
object changes.
If you log out, the loaded account object becomes `undefined` and if you
then sign in with another user, the account object remains `undefined`
so it doesn't trigger the redirect code.
This results in an infinite loading spinner.

# Changes

Trigger the redirect from another function which is run any time the
account identifier or the list of all your account (which is never empty
because you always have a main account) is loaded.

# Tests

Tested manually and added a unit test.

Drive-by change: Fixed the description of another test which was wrong
and confusing.

# Todos

- [x] Add entry to changelog (if necessary).