Rename HashPo.getText (#4692)
# Motivation

The `Hash` component is a UI component for displaying an opaque string
of characters by replacing the middle section with `...` and having a
tooltip with the full text.
The corresponding page object `HashPo` has a `getText()` method for
getting the full text.
But all page objects have a `getText()` method from their common base
class for getting the text content of the HTML element or specific
children.
The `getText()` on `HashPo` prevents us from using the `getText()` from
the base class.

# Changes

Rename `getText()` in `HashPo` to `getFullText()`.

# Tests

Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary