Use unique tooltip IDs in IdentifierHash (#4696)
# Motivation

Whenever we have a tooltip, the tooltip element must have a unique `id`
attribute and the tooltip target must have a matching `aria-describedby`
attribute to help screen readers associate the tooltip to the target.
Currently, the `IdentifierHash` component uses `identifier` as the `id`
which is used for the corresponding tooltip.
When multiple `IdentifierHash` components are on the page (for example,
the account identifier and the user principal), this `id` is not unique.

# Changes

Attach a sequence number to the `id` in `IdentifierHash` to make it
unique.

# Tests

1. Add a unit test which instantiates `IdentifierHash` twice and checks
that the tooltip IDs are different.
2. Don't reuse the same container in `render` so that we can have
multiple components instances and corresponding page objects in a single
test.

# Todos

- [x] Add entry to changelog (if necessary).