Let AgeBonusTextPo and DissolveDelayBonusTextPo extend TooltipPo (#4698)
# Motivation


[AgeBonusText](https://github.com/dfinity/nns-dapp/blob/446117ce1b3aa6f33fe9139a98a0d1dcbae9f408/frontend/src/lib/components/neuron-detail/AgeBonusText.svelte#L4)
and
[DissolveDelayBonusText](https://github.com/dfinity/nns-dapp/blob/446117ce1b3aa6f33fe9139a98a0d1dcbae9f408/frontend/src/lib/components/neuron-detail/DissolveDelayBonusText.svelte#L4)
consist of only a `Tooltip` component.

But instead of relying on `TooltipPo` to get the tooltip text, they have
their page objects have their own implementation of `getTooltipText`.
The `Tooltip` implementation in gix-components has changed, which
requires changes in `TooltipPo`.
To avoid also having to change `AgeBonusTextPo` and
`DissolveDelayBonusTextPo`, they should reuse `TooltipPo` by extending
it.

Unfortulately, private fields in TypeScript are not fully hidden from
outside the class, so we need to rename `TID` in both page objects, in
order not to collide with `TID` in `TooltipPo`.

# Changes

1. Extend `AgeBonusTextPo` and `DissolveDelayBonusTextPo` from
`TooltipPo`.
2. Rename `TID` in both to avoid collision.
3. Remove `getTooltipText` from both because it's now provided by their
base class implementation.

# Tests

Still pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary