security: Make it hard to create partitions without a schema label (#4691)
# Motivation
Partitions SHOULD always have the `schema_label` populated in virtual
stable memory. `Partitions::from(..)` does not write such a label,
relying instead on later code to populate it. This is an unnecessary
risk.

## Options
- If `Partiitions::from(..)` creates a memory manager, it should also
populate a schema. Presumably the default schema. Given that neither of
the two places that call `Partitions::from()` would benefit from this
code, this seems not the best way to proceed.

- Given that `Partitions::from()` is a tiny function, a simple
two-liner, and used only internally in two places it seems best to
simply delete it. Furthermore the two lines are likely to be reduced to
one line once we move to the next version of the stable structures
library.

# Changes
- Delete `Partitions::from(memory)`

# Tests
- None.  Existing tests should suffice.

# Todos

- [ ] Add entry to changelog (if necessary).
- When loading existing partitions from stable memory, verify that the
schema label is defined. Four unit tests that use over-simplistic toy
memory will have to be updated. This can be done in a later PR.