test: Pocketic migration tests (#4670)
# Motivation
We would like to increase the variety of events that occur in migration
tests. For example:
- The e2e migration tests do not add accounts during the migration.
- The e2e tests do not cover redeploying in the middle of a migration.
This has been tested manually but not in CI.

# Changes
- Separate running unit tests from running integration tests in CI.
- Note: Historically we had only unit tests. The new tests are
integration tests. If we do not separate them, CI becomes significantly
slower.
- Add a `pocket-ic` test environment with a stable reference `nns-dapp`
and a second `nns-dapp` to which we apply migrations.
- Note: The two canisters should contain the same data regardless of
what migrations, rollbacks or account changes we apply.
- Use the `pocket-ic` test environemnt to verify that accounts are not
lost during a normal migration.
- Use the `pocket-ic` test environment to verify that accounts are not
lost if we need to roll back.
- Use `proptest` to apply many random operations to the test environment
and verify that, in doing so, no accounts are lost.

# Tests
- These new tests are exercised in CI.

# Todos

- [ ] Add entry to changelog (if necessary).