Component to display voting power with tooltip (#4720)
# Motivation

In 2 different places we display rounded voting power with a tooltip for
the exact amount.
We want to do this in even more places.
So in this PR I factored out a component for this to reuse.

# Changes

1. Add `VotingPowerDisplay` component which displays rounded voting
power and has a tooltip with exact voting power.
2. Use this in `VotableNeuronList` and `VotingNeuronListItem` instead of
duplicating the code.

# Tests

Added unit tests for `VotingPowerDisplay` and adapted existing unit test
for the new structure.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary