NNS1-2905: Remove get_transactions method from nns-dapp canister (#4723)
# Motivation

The nns-dapp frontend now reads ICP transactions from the ICP index
canister.
It no longer needs to read transactions from the nns-dapp canister and
we want to clean up the canister code.

In this PR we only remove the possibility to get transactions from the
nns-dapp canister. We don't remove the data yet so that it's easy to
revert if for whatever reason this causes problems.

# Changes

Remove the `get_transactions` method from the nns-dapp canister.

# Tests

Existing unit tests, test the method on the account store, which still
exists.

# Todos

- [x] Add entry to changelog (if necessary).