NNS1-2905: Remove transactions from histogram (#4756)
# Motivation

We no longer read transactions from the nns-dapp canister, but from the
index canister instead.
We want to remove the transactions from nns-dapp canister so we need to
remove all the code that is still using the transactions.

https://github.com/dfinity/nns-dapp/pull/3273 added a method to get
histogram stats about the accounts stored in nns-dapp. Some of the
histograms are about transactions. When we remove the transactions, we
no longer can or need to include transaction stats in the histogram.

# Changes

Remove transaction stats from the histogram.

# Tests

Removed tests specific to transactions in histograms.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary