NNS1-2905: Stop detecting StakeNeuronNotification transactions (#4757)
# Motivation

We no longer read transactions from the nns-dapp canister, but from the
index canister instead.
We want to remove the transactions from nns-dapp canister so we need to
remove all the code that is still using the transactions.

Staking a neuron used to involve 2 ledger transactions:
1. Transferring the stake to the neuron account.
2. Transferring 0 ICP to the neuron account with a memo pointing to the
previous transaction.

The second transaction would prompt the nns-dapp to claim the neuron.
But this is no longer how neurons are staked.
The second transaction would have transaction type
`StakeNeuronNotification`.
Detecting this transaction type requires looking at previous
transactions.
Because we want to remove the transactions from the nns-dapp canister,
we can no longer do this.

# Changes

1. Treat every transaction to an existing neuron account as a top-up.
2. Remove `is_stake_neuron_notification` and `get_transaction_index`
which are no longer necessary.

# Tests

Remove the test for detecting `StakeNeuronNotification` transactions.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary