NNS1-2905: Stop fixing old transactions (#4758)
# Motivation

We no longer read transactions from the nns-dapp canister, but from the
index canister instead.
We want to remove the transactions from nns-dapp canister so we need to
remove all the code that is still using the transactions.

At some point in the past, transactions didn't have a transaction type.
New transactions were assigned a type and there was also code in place
to fix old transactions for old users once they logged in.
We want to stop storing the transactions so we no longer need to care
about fixing the type of a few very old transaction.s

# Changes

1. Stop calling `fix_transactions_for_early_user` and remove resulting
unused code.
2. Mark `get_transaction` which is no longer used, but is still used in
a test (that will be deleted in another PR) as test-only.

# Tests

Still pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary