Update IC revs in backend crate, to match proposals crate (#4751)
# Motivation

The nns-dapp canister consists of 2 crates in the nns-dapp repo:
`backend` and `proposals`.
These crates have dependencies on the IC repo, which are both old and
also different (12 months old and 8 months old).
We want to bring both of them up to the current state of the IC repo but
as a first step we want to bring the `backend` crate to the same point
as the `proposals` crate.

# Changes

1. In `rs/backend/Cargo.toml` change all the IC dependencies to `rev =
"release-2023-08-01_23-01"` which is the same as in
[rs/proposals/Cargo.toml](https://github.com/dfinity/nns-dapp/blob/main/rs/proposals/Cargo.toml).
2. Change from `ic_crypto_sha` to `ic_crypto_sha2` similar to when we
updated the `proposals` crate dependencies in
https://github.com/dfinity/nns-dapp/pull/3415
3. Remove some `unreachable!()` statements which the compiler started
warning about having become actually unreachable for any `rev` after
(including) [this
MR](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/13347).
Since that MR also removes `unreachable!()` statements, this seems like
the right approach. I only removed ones which actually caused warnings.

# Tests

CI passes.

# Todos

- [x] Add entry to changelog (if necessary).
Wrote a generic line which should cover future related PRs as well.