Use same IC deps for sns_aggregator as for proposals and backend (#4779)
# Motivation

Since https://github.com/dfinity/nns-dapp/pull/4751 the `backend` and
`proposals` crates use the same `rev` for the IC repo dependencies.
But `sns_aggregator` still uses an older `rev`.
Using the same `rev` for all crates makes it less likely that
dependencies will interfere.
And `sns_aggregator` builds fine with the same `rev` as `backend` and
`proposals`, except for 1 issue.
The issue is that `sns_aggregator` uses
`ic_nervous_system_common::get_canister_status` which was removed in
https://gitlab.com/dfinity-lab/public/ic/-/commit/bac55cb48c342112f8f20756e67e813eb3f2be2d
The function is small so I inlined it in `sns_aggregator` instead.

Defining the dependencies in the workspace makes it easier to keep them
in sync.

# Changes

1. Define IC repo dependencies in `/Cargo.toml`.
2. Use workspace dependencies in `backend`, `proposals` and
`sns_aggregator` crates.
3. Replace calling `ic_nervous_system_common::get_canister_status`
(which no longer exists) with the code that used to be inside that
function.

Also note that `Cargo.lock` got updated and has a lot more red than
green, which I assume is a good thing.

# Tests

Still pass.

# Todos

- [ ] Add entry to changelog (if necessary).
covered by existing entry `Update IC dependencies in nns-dapp crates.`