upgrade-downgrade test script (#4787)
# Motivation

We test that we can upgrade and downgrade between the prod canister and
the next version of the canister.
This test first downgrades from the new version to the prod version and
then upgrades from the prod version to the new version. This was done
because we assumed the new version was already running but this is no
longer the case.

So to be more realistic, let's upgrade first and then downgrade.

This test is performed on CI. To make it more convenient, I've added a
script to do it and call the script from CI.

# Changes

1. Add a `scripts/nns-dapp/upgrade-downgrade-test` script to perform the
upgrade-downgrade test.
2. Call the script from `.github/workflows/build.yml` instead of the
original steps.
3. Swap the `--wasm1`/`--args1` and `--wasm2`/`--args2` parameters to
make the test more realistic.
4. Rename the job from `test-downgrade-upgrade` to
`test-upgrade-downgrade` to reflect the swap.

# Tests

I changed the name of a struct that's stored in stable memory and made
sure it was caught by the test.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: sa-github-api <138766536+sa-github-api@users.noreply.github.com>
Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>
Co-authored-by: Max Strasinsky <98811342+mstrasinskis@users.noreply.github.com>