Do not rely on Docker build for SNS aggregator release template (#4815)
# Motivation

Our Docker build runs out of memory on M1 Mac.
It would be nice to be able to do most of the SNS aggregator release
process on a Mac, even if we need to run the Docker build separately on
a DevEnv.

# Changes

To get the SNS aggregator WASM to release, download it from CI instead
of building it with Docker.

# Tests

Used for this week's SNS aggregator release.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary