NNS1-2905: Remove transactions fields from accounts (step 1) (#4800)
# Motivation

We no longer store transactions in the nns-dapp canister because the
frontend now reads transactions from the index canister.
However we still store transaction indexes on individual accounts. These
indexes used to point into the vector of transactions which we no longer
have and are now completely unused. So we want to clean these up.
Because these accounts are stored in stable memory we have to be careful
about removing fields from them or we might get errors when decoding
stable memory.
In particular, we should be able to revert to a previous release of
nns-dapp in case of an emergency.
So we have to stop reading the fields at least 1 release before we stop
writing the fields.
For this purpose we introduce "Old" types for each of the account types
which still have the removed fields and use these types when writing to
stable memory, while using types with the unused fields removed when
reading.

Once we have a release with this code, we can remove the "Old" types and
read and write without the unused fields.
There will still be accounts in stable memory with the unused data but
it will be ignored on reading and removed on writing.

# Changes

1. Add `OldAccount` which is a copy of `Account` (as before this PR).
2. Add `OldNamedSubAccount` which is a copy of `NamedSubAccount` (as
before this PR).
3. Add `OldNamedHardwareWalletAccount` which is a copy of
`NamedHardwareWalletAccount` (as before this PR).
4. Remove `default_account_transactions` from `Account`.
5. Remove `transactions` from `NamedSubAccount` and from
`NamedHardwareWalletAccount`.
6. In `Storable::to_bytes` for `Account`, convert to `OldAccount` before
encoding.
7. Stop populating transactions fields (which no longer exist) when
creating toy accounts for testing.

# Tests

1. Unit tests added to convert old to new and new to old accounts and to
decode an encoded new account to a new account.
2. `upgrade-downgrade-test` passes.
3. Tested upgrade and downgrade manually.
4. Prepared [branch with future
clean](https://github.com/dfinity/nns-dapp/pull/4836) up and tested
(manually and with `upgrade-downgrade-test`) upgrade and downgrade
between this branch and that one.
5. Also tested upgrading directly to the final branch and noticed that
it had the expected problems.

# Todos

- [x] Add entry to changelog (if necessary).