Text changes for API Boundary nodes proposal (#4851)
# Motivation

[This
MR](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/19097)
changed the proposal type to add boundary nodes to allow adding multiple
boundary nodes in a single proposal. This is a breaking change but it's
fine because no such proposals have ever been created and will not be
created for a few more weeks.

We'll have to make changes to the payload parsing in the NNS dapp but
this PR only has the corresponding text changes.

# Changes

Change proposal description to be about multiple API boundary nodes.

# Tests

Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/proposal/?u=qsgjb-riaaa-aaaaa-aaaga-cai&proposal=17

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary