Separate candid for proposals from candid for sns_aggregator (#4861)
# Motivation

We generate Rust types from candid.
We do this both for the `proposals` crate and the `sns_aggregator`.
We have 3 automatic workflows for this:
1. Update candid files from IC repo
2. Create Rust types for `proposals`
3. Create Rust types for `sns_aggregator`

This has some downsides:
1. We need to review the same changes twice: once in Candid and once in
Rust.
2. After step 1 but before steps 2 and 3, our Candid and Rust are out of
sync.
3. If step 2 or 3 fails, changes keep piling up in the Candid file and
we either need to close those PRs or they get more and more out of sync.
4. When changing the scripts for steps 2 and 3, it's hard to test when
Candid and Rust are already out of sync.

So I propose we change the process to the following workflows:
1. Update candid files and Rust types for `proposals`.
2. Update candid files and Rust types for `sns_aggregator`.

One thing that makes this difficult is that the candid files for
`sns_wasm` are used by both. So in order to update the Candid files for
both separately, we need 2 copies of the `sns_wasm.did` file, which are
potentially at different versions.

In this PR, I just separate the Candid files for `proposals` and
`sns_aggregator` into different directories and duplicate the
`sns_wasm.did`. We still have the original 3 workflows, but this step
makes it easier to change that.

# Changes

1. Move `declarations/nns_governance` to
`declarations/used_by_proposals/nns_governance`.
2. Move `declarations/nns_registry` to
`declarations/used_by_proposals/nns_registry`.
3. Copy `declarations/sns_wasm` to
`declarations/used_by_proposals/sns_wasm`.
4. Move `declarations/sns_*` to
`declarations/used_by_sns_aggregator/sns_*`.
5. Update scripts to point to the correct directory in `declarations/`.

Drive-by: Remove a comment about wasms in `scripts/update_ic_commit`
because that code only deals with .did files and no wasms.
# Tests

1. I ran `scripts/update_ic_commit` and checked that it updates the
correct Candid files.
2. I ran `scripts/sns/aggregator/mk_nns_types.sh` and checked that it
updates the correct Rust files.
3. I ran `scripts/proposals/did2rs` and checked that it updates the
correct Rust files.

# Todos

- [ ] Add entry to changelog (if necessary).
not necssary