Update Candid for proposals and sns_aggregator independently (#4875)
# Motivation

See https://github.com/dfinity/nns-dapp/pull/4861 for motivation.
This PR is to make it possible to update candid for `proposals` and
`sns_aggregator` independently.
In a future PR we'll use that possibility to make the proposed changes
to the workflows.

# Changes

1. Add a `--crate` flag to `scripts/update_ic_commit` to specify for
which crate we want to update the IC commit and Candid files.
2. Allow `all` (which is default) to keep the original behavior of
updating all Candid files.
3. Keep separate fields in `dfx.json` instead of the single `IC_COMMIT`
fields.

# Tests

1. Manually tested with `--crate proposals`, with `--crate
sns_aggregator` and without `--crate`.
2. Updated the workflow test.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary