Actionable query param support (#4880)
# Motivation

To enable linking to an actionable tab and allow navigation using
browser history, a new query parameter - `actionable` has been added. It
will serve as a flag indicating that the actionable tab should be
displayed. The plan is to set this additional parameter when the user:

	1.	Clicks on “Voting” in the main menu.
2. Navigates to the proposals page without a specified universe query
parameter (`u`).

# Changes

- New const `ACTIONABLE_PARAM`.
- Make layout to provide `actionable`  as a bool value.
- Make `actionable` available through the `pageStore`.
- New util `buildActionableProposalsUrl`

# Tests

- Update layout tests mock (expect actionable param).
- Test `buildActionableProposalsUrl`.
- Manually verified that the `pageStore` reflects the `actionable`
parameter in the URL. Currently, there appears to be no straightforward
method for unit testing this.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.