NNS1-2905: Remove transactions fields from accounts (step 2) (#4836)
This PR should only be merged after we have a release on mainnet
containing https://github.com/dfinity/nns-dapp/pull/4800.

The commit currently on mainnet:
```
$ dfx canister metadata nns-dapp git_commit_id --network mainnet
581eb325ce38d98f3e6e9c13ab2e490701ca91c3
```
That
[commit](https://github.com/dfinity/nns-dapp/commit/581eb325ce38d98f3e6e9c13ab2e490701ca91c3)
is from May 15th, well after PR #4800, which was merged May 10th.

# Motivation

This is the follow-up to https://github.com/dfinity/nns-dapp/pull/4800.
In the previous PR we removed the transactions fields but still encoded
to stable memory including the fields.
In this PR we stop encoding the old fields completely.
There will still be accounts in stable memory with the fields but they
will be ignored when reading and removed when writing.
We have not yet decided if we also want to do a migration round to
remove all the fields from stable memory.

# Changes

1. Stop converting `Account` to `OldAccount` before encoding it to
stable memory.

# Tests

1. `upgrade-downgrade-test` passed.
2. Manually tested upgrading and downgrading while creating new
subaccounts before, in between and after.

# Todos

- [x] Add entry to changelog (if necessary).