Snses without actionables banner (#4902)
# Motivation

Because not all projects support actionable proposals (return neuron
ballots), not all actionable proposals will be shown on the actionable
proposals page. To make it clear that there could be more votable
proposals, we display a banner with a list of projects that do not have
actionable support. In this PR, we create a banner component that will
be shown on the actionable proposals page.

# Changes

- Add new component `ActionableProposalsNotSupportedSnses`

# Tests

- PO for `ActionableProposalsNotSupportedSnses`
- Tests for `ActionableProposalsNotSupportedSnses`

# Todos

- [ ] Add entry to changelog (if necessary).
not yet.

# Screenshot

<img width="1325" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/98811342/ab1fbcf6-0f6e-4a8f-ac69-cb79fe990808">