Extract cell components from TokensTableRow (#4909)
# Motivation

We want to try to reuse the tokens table structure and style for the
neurons table.
The `TokensTable` component is currently very specific to the tokens
table.
By extracting tokens specific subcomponents, I'm hoping to separate the
generic table structure from the tokens specific rendering.
And this should make it easier to subsequently create a generic table
component and use that for both the tokens table and the neurons table.

# Changes

1. Create `TokenTitleCell`, `TokenBalanceCell` and `TokenActionsCell`
component.
2. Use those components in `TokensTableRow.svelte`

# Tests

All tokens table functionality continues to be covered by
`src/tests/lib/components/tokens/TokensTable.spec.ts`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary