Remove data-title from TokensTableRow (#4911)
# Motivation

Having an attribute on the `TokensTableRow` which is specific to the
tokens data makes it more difficult to make the table generic so it can
be reused for the neurons table.
This attribute was only used for testing.

# Changes

1. Remove `data-title` attribute.
2. Instead of using the `data-title` attribute to find a specific row,
iterate over the rows and return the one that matches.
3. Use `await` where necessary because the method now has to be async.

# Tests

Pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary