Rename userTokenData prop to rowData (#4917)
# Motivation

We want to evolve `ResponsiveTable`, which is currently specific to
tokens to a fully generic table.
So its props should not refer to tokens in the name.

The token table cell components will remain specific to the tokens
table, but because they will have to interact with the generic table,
their props need to be generic as well.

# Changes

1. In `frontend/src/lib/components/ui/ResponsiveTable.svelte` change
`userTokensData` to `tableData` and `userTokenData` to `rowData`.
2. In `frontend/src/lib/components/ui/ResponsiveTableRow.svelte` and in
all the cell components, change `userTokenData` to `rowData`.

Note that each cell gets the data for the full row because it's the
cell's responsibility to know which data from the row it needs to
render.

# Tests

Existing tests continue to pass.
Tested manually as well.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary