New actionable proposals loaded store (#4918)
# Motivation

We need to determine when the actionable proposals are fully loaded to
display a loader. Here, we add a new store that returns true only when
everything is preloaded. We can’t display a loader per SNS because we
only know if the Sns needs to be displayed after fetching, based on the
`includeBallotsByCaller` flag.

# Changes

- New derived store `actionableProposalsLoadedStore`

# Tests

- tests for the new store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.