Use a generic type for ResponsiveTable data (#4921)
# Motivation

We want to use the `ResponsiveTable` with different types of data.

# Changes

Declare the `tableData` in `ResponsiveTable` and `rowData` in
`ResponsiveTableRow` to be of a generic type.
I found this syntax
[here](https://github.com/ciscoheat/sveltekit-superforms/issues/245#issuecomment-1698928562)
and it seems to satisfy both the linter and the compiler.

# Tests

Existing tests pass.
I've also tested this with a preliminary neurons table in another
branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary