Add test for ResponsiveTable (#4944)
# Motivation

A generic `ResponsiveTable` component was extracted from `TokensTable`.
So far we relied only on `TokensTable.spec.ts` for testing but the
`ResponsiveTable` is now generic enough that it makes sense for it to
have its own test.
I'll continue making it more generic but now it will also be possible to
test separately.

# Changes

1. Add `testId` prop to `ResponsiveTable`.
2. Change `data-tid` on `ResponsiveTableRow`. This doesn't need to be a
prop since once table only has one kind of row so there can never be
confusion.
3. Add some cell components for testing.
4. Add page objects.
5. Add tests.

# Tests

Old and new tests pass.
Also did a quick manual check.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary