Enable ENABLE_CKUSDC feature flag (#4948)
# Motivation

We already created the [custom
proposal](https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=130093)
to enable the `ENABLE_CKUSDC` feature flag.
But we also need to make sure that the feature flag is enabled in future
release.

# Changes

1. Set `ENABLE_CKUSDC` to `true` in `dfx.json`.
2. Ran `scripts/nns-dapp/test-config --update`.
3. Add `ckUSDC` to the list of non-SNS tokens in
`TokensTable.page-object.ts` to make the e2e test pass.

# Tests

The feature flag has been tested a lot. But not separately for this PR.

# Todos

- [x] Add entry to changelog (if necessary).