Update actionable page gaps (#4955)
# Motivation

Different types of components will be rendered on the actionable page,
and it will be easier to control the gaps between them from the page
component itself. This PR changes how the gaps are set, switching from
`margin-bottom` in child components to flex `gap` in the parent
component.

# Changes

- Use flex `gap` to set the spacing between the direct children of the
page component..
- Remove bottom margin from `UniverseWithActionableProposals`, but the
same gap was added to it's parent component `ActionableSnses`.
- Switch from `TestIdWrapper` to `div` in skeleton component to avoid
container gaps.

# Tests

- Pass.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.