Display unsupported snses banner (#4961)
# Motivation

Some Snses don’t support actionable proposals yet, meaning there could
be additional proposals to vote on that are not displayed on the
actionable proposals page. To make this clear, we display a banner with
a list of Snses that don’t support actionable proposals.

Note: This banner should only be displayed when some actionable
proposals are available. Otherwise, Snses without support will be
mentioned in the `ActionableProposalsEmpty` banner.

# Changes

- Display `ActionableProposalsNotSupportedSnses` on actionable page.

# Tests

- Move some test variable to the root describe block, to make them
reachable by other tests.
- Test `ActionableProposalsNotSupportedSnses` visibility.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.