Display unsupported snses in all-voted banner (#4960)
# Motivation

In cases where the user has voted on all available actionable proposals,
we currently display the `ActionableProposalsEmpty` banner with the text
"You’re all caught up." However, if some SNSes still don’t support
actionable proposals, we cannot be sure that there are no votable
proposals for these SNSes. To make this clear, the
`ActionableProposalsEmpty` banner should display more detailed
information. In this PR, we update the description to display the SNSes
that may contain votable proposals.

# Changes

- New `joinWithOr` util.
- Display unsupported Snses in `ActionableProposalsEmpty` banner
description.
- Not related refactoring: removed redundant array creation
[here](https://github.com/dfinity/nns-dapp/pull/4960/files#diff-bdabfefb322354658061b1e40e4e4b5a1fc2feab676ffcd43e37d91b9b7d4894R12).

# Tests

- Test for `joinWithOr`.
- Update `PageBannerPo`.`under` arguments to support it's creation from
the root component.
- Test for `ActionableProposalsEmpty` component (now it makes sense,
since it includes a dynamic text).

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.