Render voting power for min dissolve delay (#4963)
# Motivation

When setting the dissolve delay to the minimum on an NNS neuron (182
days and 15 hours), on the details page we say the neuron has no voting
power. This is because we incorrectly used a strict inequality `>`
instead of a weak inequality `>=` when comparing the dissolve delay.
Coincidentally we also use language suggesting a strict inequality.

# Changes

1. Use the correct inequality in `NnsNeuronPageHeading.svelte` and
`NnsNeuronVotingPowerSection.svelte`.
2. Change the language from saying "greater than" to "at least".

# Tests

1. Use strict bounds in the tests.
2. Update strings.

# Todos

- [x] Add entry to changelog (if necessary).