NNS1-3092: Show spinner while neurons table is loading (#4967)
# Motivation

While the neurons page is loading, we currently show skeleton cards.
This doesn't make sense if after loading a table appears instead of
cards.

Also, if there are zero neurons, we should not show anything except the
empty message.
With cards this is automatic because we just show zero cards.
But we don't want to show a table header with zero rows if there are
zero neurons.

# Changes

1. If neurons are loading, show a spinner.
2. If neurons are loaded but there are no neurons, hide the table.

# Tests

Unit tests added for spinner, skeleton cards and empty message with and
without feature flag.

Drive-by: Update `isContentLoaded` to take into account that we might
have a spinner instead of skeleton cards. This is not used yet but we're
going to need it eventually.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet