Remove "Schedule silence alerts" from release-sop (#4991)
# Motivation

After migrating to stable structures and dropping the storage of some
user data in the `nns-dapp` canister, the performance of `nns-dapp`
upgrades significantly improved. The last canister upgrade was completed
without silencing any alerts and that worked out well. Therefore, we no
longer need this step in our SOP script, and here we remove it.

# Changes

- Remove "Schedule silence alerts" step from the script.

# Tests

- Update related tests.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.