NNS1-3092: Support custom column templates in responsive table (#5000)
# Motivation

We want to align the neuron ID copy icons across different rows in the
neuron table.
For this we nee a subgrid within the neuron ID cell, with an extra grid
line aligning the icons.
The first step for this is to allow customizing the template columns.
In the next step we'll allow multiple grid columns/rows per table
column/row so the `templateColumns` field on the `ResponsiveTableColumn`
is already an array, even though for now, arrays with a size unequal to
1 don't work yet.

Currently we assume that the template columns are always `1fr` for the
first column and `max-content` for every other column. In this PR we
make that explicit. And in a future PR we'll actually customize it.

# Changes

1. Add `TemplateItem` type and `templateColumns` field in
`ResponsiveTableColumn`.
2. Specify the `templateColumns` for each column in the column spec of
the `TokensTable` and `NeuronsTable`.
3. Generate `desktopGridTemplateColumns` based on the column spec
instead of always assuming `1fr max-content max-content ...`.
4. Use the template from the first and last column in the mobile grid
template columns.

# Tests

1. Unit tests added/updated.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not yet