NNS1-3092: Align neuron ID copy icons in table (#5006)
# Motivation

Because we don't use a fixed-width font for the neuron IDs, they don't
all have exactly the same width.
This means the copy icons right after them are not aligned, which looks
off.

To solve this we define extra grid lines in the table and use them to
align the copy icons.

# Changes

1. Allow specifying multiple grid columns per table columns and multiple
grid rows per table row.
2. Specify 2 grid rows per table row and 3 grid columns for the first
table column of the neurons table.
3. Use the extra grid columns for the copy icon and the empty space and
the extra row for the neuron tags.
4. Change the `NeuronIdCell` component to use separate elements for the
ID and the copy icon so they go into separate columns.

# Tests

1. Unit tests updated.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/


# Todos

- [ ] Add entry to changelog (if necessary).
not yet