Fix row gap in mobile view of responsive table (#5025)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/5013 I removed a `gap` style
because I thought it was redundant.
Because the style was defined on `[role="row"]`, I assumed its purpose
was to define a gap between columns, which was already define somewhere
else. But I had forgotten for a moment, that in mobile view, each row
consists of multiple rows, and it's these rows for which the gap was
being defined here.

# Changes

1. Restore a `row-gap` rather than `gap` to make it clearer what the
purpose is.
2. Add comments and group styles, also to make it clearer why each of
the styles are there.

# Tests

1. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/
2. Adding a screenshot test in
https://github.com/dfinity/nns-dapp/pull/5019

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary