NNS1-3092: Unalign copy icons (#5045)
# Motivation

The design has changed and we no longer want to align the copy icons.
We also want to put the neuron tags on the same row as the neuron ID if
there is space.

# Changes

1. Remove grid layout from `NeuronIdCell`.
2. Use the copy button functionality of the `Hash` component instead of
separating it, which was only necessary for the alignment.
3. Make the tags inline.
4. Use template column `minmax(min-content, max-content)` to use the
space for tags if possible but not use less space than needed to with
the `Hash` component`.
5. Go back to using a single grid row per table row. Double grid rows
was only necessary to make the icon alignment work with tags on the
second row of the cell.

# Tests

1. Unit test updated.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not yet