Prevent clicking on tooltip icons to follow links (#5073)
# Motivation

We want to put tooltip icons in the neurons table rows.
On mobile, you have to touch the tooltip target for the tooltip to
appear.
The table rows are actually links so if you touch anything in it, it
opens the neuron details view.

# Changes

1. Add a `preventDefault` click handler on the tooltip icon component.

# Tests

Can't really be unit tests because of limitations in the testing
framework.
Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary