Reset proposals scroll position (#5072)
# Motivation

When switching between universes, the user would find themselves
somewhere in the middle of the list. This issue went unnoticed before
due to paginated loading, which resulted in the list being empty
immediately after switching. However, with actionables now being
preloaded, this has become a significant UX issue.
In this PR, we reset the scroll position on the Proposals page after
switching projects.

# Changes

- Added the `resetScrollPositionAfterNavigation` prop to the
`UniverseSplitContent` component to avoid resetting the scroll on the
neurons/ page. Currently, the `UniverseSplitContent` component is used
only on the neurons/ and proposals/ pages.
- Call the `resetScrollPosition` method of the proposals/ page split
container after every navigation (there is no other navigation on this
page besides the page opening and the universe switch).

# Tests

- Manually tested.

# Todos

- [x] Add entry to changelog (if necessary).