Rename TooltipIconPo.getText() to getTooltipText() (#5077)
# Motivation

`TooltipIconPo.getText()` is masking `getText()` on the base class.
I want to add the possibility to pass the tooltip content as a slot to
the `TooltipIcon` component and I want to add a test which calls
`TooltipIconPo.getText()` (from the base class) to make sure the slot
doesn't end up in the tooltip target as opposed to the tooltip content.

# Changes

Rename TooltipIconPo.getText() to getTooltipText()

# Tests

Unit tests updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary