Remove unnecessary step from did2rs.sh (#5087)
# Motivation

`did2rs.sh` uses `didc` to create Rust from Candid files.
It adds a step to make types and fields public.
But newer versions of `didc` already do this.
So the extra step is not necessary anymore.
And in a beta version I tried this unnecessary step was causing
problems.

# Changes

Remove the step to add `pub` to types and fields.

# Tests

CI has
[tests](https://github.com/dfinity/nns-dapp/blob/ca8d8c26be5d64d9319e3ccfce8a2b7e4e3a1a1d/.github/workflows/checks.yml#L146-L153)
to make sure generated code is the same.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary