NNS1-3134: Add tooltip icon for available/staked maturity (#5090)
# Motivation

So available/staked maturity can be seen without opening neuron details.

# Changes

1. Add a tooltip icon in the maturity column which shows maturity split
out to available and staked maturity.
2. Make the `text` prop on `TooltipIcon` optional because the slot can
be used instead.
3. Change the Playwright click function to click in the corner of
elements. The maturity tooltip icon happens to be right in the middle of
the row and it is not clickable to allow triggering the tooltip on
mobile without opening the neuron details. This was preventing clicks on
the neuron row because they would happen right in the middle where the
tooltip icon is.

# Tests

1. Update page objects.
2. Add unit tests.
4. Manually:
<img width="683" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/cf8e1f48-7213-48f7-9432-36cc35bab9c1">

# Todos

- [ ] Add entry to changelog (if necessary).
covered