Move generic sorting methods from neurons-table.utils to responsive-table.utils (#5101)
# Motivation

We want to allow custom sorting in the neurons table.
It should potentially work on a places where the `ResponsiveTable` is
used.
We have so sorting utility functions that are not specific to the
neurons table that will be used by the responsive table.

# Changes

Move the generic table sorting utility functions from
`neurons-table.utils.ts` to `responsive-table.utils.ts`.

# Tests

Unit tests were added.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary