NNS1-3135: Add util for table resorting logic (#5107)
# Motivation

When we make the neurons table sortable we want to use the following
logic:
1. If you click on a column header, the rows should be sorted by that
column.
2. If you click again on the same column header, the rows should be
sorted by the reverse of that column.
3. When changing sort order, ties in the chosen column should be broken
based on the previous order.

So if you sort by stake and then by dissolve delay, neurons with the
same dissolve delay should be sorted by stake because that was the
previous order.

This PR does not make any UI changes but adds the utility that
implements the above logic.

# Changes

1. Add optional `reversed` field to the elements of
`ResponsiveTableOrder`.
2. Add `selectPrimaryOrder` function which creates a new order following
the above logic.
3. Support the `reversed` field in `sortTableData`.

# Tests

1. Unit tests added for `selectPrimaryOrder`.
2. Unit test extended for `sortTableData` with `reversed: true` items.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary