NNS1-3135: Neurons table order store (#5110)
# Motivation

When changing the sorting order on the neurons table, we want to be able
to navigate without losing the order.
(Resetting the order on page reload is OK, at least for now.)

This PR adds a Svelte store to keep the ordering across navigation but
does not use it yet.

# Changes

Add `neuronsTableOrderStore`.

# Tests

1. Unit test added.
2. Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary