NNS1-3135: Refactor column header hiding logic (#5111)
# Motivation

We want to make table columns sortable by clicking on them, if there is
a sort order defined for that column.
In mobile view we only show the first column, but sorting is not done by
clicking on the column headers.
So we might have to show 2 versions of the first column header.

To make this easier, this PR only changes the way by which columns are
hidden by using an explicit class instead of always hiding everything
but the first and last column headers.

# Changes

1. Add a `desktop-only` class to column headers that should be hidden on
mobile.
2. Change the CSS to use this class instead of `:first-child` and
`:last-child` selectors.

# Tests

Add separate unit tests for mobile and desktop column headers.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary