NNS1-3135: Sort neurons table based on store order (#5112)
# Motivation

Once we support custom sorting, we want to keep the sort order between
navigations.
In this PR we just hook up the sort order to the store. There is no
visible difference because there is no way yet for the user to change
the contents of the store.

# Changes

Sort the neurons table based on the order in the
`neuronsTableOrderStore` instead of a constant in the component.

# Tests

Unit test added where the store contents is changed.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary