Exclude empty neurons with listNeurons (#5123)
# Motivation

NNS dapp does not display empty neurons and just filters them out.
The governance canister is going to have a new feature where empty
neurons can be excluded.
The API was already added so we can start passing the relevant parameter
to benefit from it as soon as it's available.

# Changes

Pass `includeEmptyNeurons: false` to `listNeurons`.

# Tests

1. Unit test updated.
2. Tested manually that nothing breaks, but the functionality is not
implemented yet so couldn't be tested.

# Todos

- [ ] Add entry to changelog (if necessary).