NNS1-3094: Add TableProject type (#5150)
# Motivation

We want to use the `ResponsiveTable` component for the staking projects
table.
In this PR we define a `TableProject` type to use as the row data for
the table and add a simple first version of a function to create the row
data.

# Changes

1. Add `frontend/src/lib/types/staking.ts` with the `TableProject` type.
2. Add `frontend/src/lib/utils/staking.utils.ts` with `getTableProjects`
to create the row data for the projects table (which will be added in
another PR).

# Tests

1. Unit test added.
2. Tested manually in another branch which already has a simple projects
table:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [ ] Add entry to changelog (if necessary).
not yet