Refresh SNS neuron from details page if needed (#5169)
# Motivation

Topping up a neuron is a 2 step process.
1. Transfer tokens to the neuron account.
2. Notifying the governance canister to refresh the neuron.

If the process is interrupted between step 1 and 2, your tokens are now
missing.
Because of this possibility, we compare the neuron account balance of
SNS neurons with the neuron stake and refresh the neuron if there is a
difference.

Currently we do this for all neurons of an SNS every time we load the
neurons of that SNS, which is when you navigate to the neurons table for
that SNS. This is overkill since this is really only a fallback
mechanism. It also resulted in loading neurons in a different way for
actionable proposals to avoid doing the check even more often.

It should be enough to only check a neuron when opening its details page
and at most once per session for the same neuron.

This PR adds the check on the neuron details page but does not yet
remove any existing checks.

# Changes

1. Add a store to keep track of neurons which have already been checked.
2. Refactor `checkNeurons` to extract `checkNeuron` to check a single
neuron.
3. Add `refreshNeuronIfNeeded` which uses the store to make sure the
same neuron is only checked once and then refreshes the neuron if
needed.
4. Call `refreshNeuronIfNeeded` from `SnsNeuronDetails` and reload the
neuron if it was refreshed.

# Tests

1. Unit tests added.
2. Tested in a branch where the other check has already been removed, by
manually transferring tokens to the neuron account and seeing the stake
update after opening the details page.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary