Test CMC notify fallback (#5166)
# Motivation

Creating or topping up a canister is a 2-step process:

1. Sending ICP to a subaccount of the CMC.
2. Notifying the CMC about the transaction.

Both steps are done in the frontend, but if the process is interrupted
and
the second step is not performed, the ICP could go missing.
So the nns-dapp canister monitors the ledger for such transactions and
also
notifies the CMC about them.

Because this is fallback behavior, it's easy not to notice it when this
is not working.

And indeed this recently got broken on the test environment because in
order to fix a vulnerability in the CMC, the nns-dapp got exclusively
whitelisted to do this notifying on behalf or other principals. But on
the test environment, nns-dapp has a different canister ID so it wasn't
whitelisted.

This is now fixed and I want to put a test in place to make sure it
doesn't break again without us noticing.

# Changes

1. A bash script which transfer ICP to the correct subaccount of the CMC
and checks that a canister is created or topped up.
2. Added the test to CI.

# Tests

Ran the test on Mac and on CI (Linux).

# Todos

- [x] Add entry to changelog (if necessary).