Remove old CMC notify mechanism (#5171)
# Motivation

When an NNS dapp user creates or tops up a canister, the FE make a
transaction to the CMC. The nns-dapp canister sees this transaction
(because it looks at every ICP ledger transaction) and then notifies the
CMC to finish the creation/topup.

More than 2 years ago we had a different mechanism where the FE makes a
transaction to the nns-dapp. Then nns-dapp sees that transaction of
forwards the ICP to the CMC and notifies it.

The code for this second mechanism still exists in the nns-dapp canister
but hasn't been used by the frontend for over 2 years.

Having the extra code is confusing and a maintenance burden so we should
remove it.

# Changes

Remove the canister code for the old canister creation/topup flow.

# Tests

The test added in https://github.com/dfinity/nns-dapp/pull/5166 still
passes.

# Todos

- [x] Add entry to changelog (if necessary).