Add claimNextNeuronIfNeeded function (#5175)
# Motivation

Staking an SNS neuron is a 2-step process:

1. Transfer tokens to the neuron account.
2. Notify the governance canister to claim the neuron.

If the process is interrupted between steps 1 and 2, the tokens are now
missing.
To be able to recover from this, we create neuron accounts in a
deterministic sequence.

Currently every time you go to the neurons table view we check all the
neuron accounts of all the existing neurons and the non-existing neurons
until the first account without balance. This is all done in a single
function.

Because this is just a fall-back mechanism, this is overkill.
We want to only check neuron accounts once per neuron per session.
1. For existing neurons we do this only when you visit their neuron
details page. This is done in
https://github.com/dfinity/nns-dapp/pull/5169
2. For new neurons we want to do this when you visit the neurons table
page.

Currently `checkSnsNeuronBalances` does both (1) and (2) so in order to
replace it, this PR adds a function to do just (2).
The next PRs will call the new function from the neurons table view and
stop calling the existing function.

This PR adds a function to 

# Changes

Add `claimNextNeuronIfNeeded` which
1. For an SNS finds the first neuron subaccount for which there is no
neuron
2. Check if that subaccount is already checked
3. And if not, checks the balance
4. And if it's at least the minimum stake, claim and loads the neuron.

# Tests

1. Unit tests added.
2. Tested manually in a branch which replaces the existing function with
the new function.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary