Remove refresh from getSnsNeuron (#5179)
# Motivation

We refresh a neuron if needed once per session per neuron if you visit
the neuron details page.
This is a fallback mechanism only needed in case a neuron top-up gets
interrupted so we don't need to do it very often as long as there is
some way to recover.

I found that `getSnsNeuron` actually also does this every time.
This is overkill.

# Changes

Remove the neuron refresh code from `getSnsNeuron`.

# Tests

1. Unit tests adapted and removed.
2. Tested manually that transferring tokens to the neuron account still
results in a top-up after visiting the neuron details page.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary