Omit large fields for proposals (#5187)
# Motivation

For certain proposal filter selections (e.g., "Topic: SNS and Neurons’
Funds"), the NNS-Dapp may respond with the error: `The current proposals
response is too large. Please adjust proposal filters to get fewer
results.` As a result, no proposals are displayed. This issue occurs due
to the large size of the SNS logos.

Error sample from browser console:
```
index.7d348019.js:16 Error: Call failed:
  Canister: rrkah-fqaaa-aaaaa-aaaaq-cai
  Method: list_proposals (query)
  "Status": "rejected"
  "Code": "CanisterError"
  "Message": "IC0504: Error from Canister rrkah-fqaaa-aaaaa-aaaaq-cai: Canister violated contract: ic0.msg_reply_data_append: application payload size (3661753) cannot be larger than 3145728."
    at o (vendor.19eafb27.js:47:3669)
    at async Ef.listProposals (vendor.19eafb27.js:71:74092)
    at async kd (index.7d348019.js:14:24857)
    at async In (index.7d348019.js:14:28975)
    at async v (index.7d348019.js:22:573350)
    at async index.7d348019.js:22:573569
```

# Changes

- Apply the `omit_large_fields` flag to the proposal request on the
proposals page. Since logos are not shown on that page, this change will
not affect the display. The proposal will be re-fetched using a separate
call on the proposal detail page, ensuring the logo is still displayed
in the payload section.

# Tests

- Tested manually against `mainnet` that it fails w/o the flag and works
as expected with.

# Todos

- [x] Add entry to changelog (if necessary).