Get summary lifecycle through SnsSummaryWrapper (#5192)
# Motivation

We want to stop using the `summary.swap` field because it's deprecated
and everything on it is available in other ways.
One of the things we get from the `swap` field is the `lifecycle`.
In this PR we introduce a getter on `SnsSummaryWrapper` to determine in
a single place how we get the lifecycle.
This will make it easy to switch over in another PR.

# Changes

1. Add `getLifecycle()` to `SnsSummaryWrapper` which returns the
lifecycle from an `SnsSummary`.
2. Use `SnsSummaryWrapper` and `getLifecycle()` instead of getting
`summary.swap.lifecycle` directly.
3. Add `override()` to `SnsSummaryWrapper` because `SnsSummaryWrapper`
can't be spread into a new `SnsSummary` object.

# Tests

1. Unit tests added for `SnsSummaryWrapper`.
2. Existing unt tests updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary