Stop getting lifecycle from swap state (#5207)
# Motivation

We want to stop using the `summary.swap` field because it's deprecated
and everything on it is available in other ways.
In this PR we stop getting the `lifecycle` from the `summary.swap`.

# Changes

1. Update`ProjectStatus.svelte` which (I believe) was the last place
where weren't getting the lifecycle via
`SnsSummaryWrapper.getLifecycle`.
2. Change `SnsSummaryWrapper.getLifecycle` to get the lifecycle from the
lifecycle field instead of the swap fields.

# Tests

Existing tests pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary