Do not render error messages as HTML by default (#5209)
# Motivation

Some error messages include URL params, for example if the URL contains
a canister ID or account ID that's invalid.
The invalid ID is that repeated in the error message and rendered as
HTML instead of text.
This allows HTML injection.
For examples (require being signed in):
*
https://nns.ic0.app/wallet/?u=qoctq-giaaa-aaaaa-aaaea-cai&account=%3Ch1%3Ehello%3C/h1%3E
* https://nns.ic0.app/canister/?canister=%3Ch1%3Ehello%3C/h1%3E

One error message needs to be rendered as HTML because it includes a
link.
This message was added in https://github.com/dfinity/nns-dapp/pull/4373
To facilitate this, the PR changes error rendering to HTML by default,
which is not safe.

We should render text by default and only render HTML when we know the
error does not contain user input.

# Changes

1. Instead of always rendering toast errors as HTML, accept a parameter
which is not set by default.
2. Add a `renderAsHtml` parameter to `LedgerErrorKey` which can be used
to indicate that the error should be rendered as HTML.
3. Set this to true for `app_version_not_supported` which includes a
link to a forum post.
4. When converting an error to a toast, check if the error has the
`renderAsHtml` field set to true, and only then render the toast as
HTML.

# Tests

1. When expecting calls to throw `LedgerErrorKey`, instead of `toThrow`
(which only compares the message) use `toMatchObject` to check
`renderAsHtml`.
2. Test that `app_version_not_supported` is rendered as HTML.
3. Test that other errors are not rendered as HTML.

# Todos

- [x] Add entry to changelog (if necessary).