NNS1-3094: Make projects without neurons clickable again (#5216)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/5202 I made projects without
neurons not clickable.
But this makes it not possible to stake those tokens.
We are still working on a way to stake tokens from the projects table
but this doesn't need to be part of the MVP.
So we want to make all projects clickable again for now.

# Changes

Give every `TableProject` a `rowHref` regardless of neurons.

# Tests

1. As we may go back again on this once there is a way to stake from the
projects table, I kept the structure of the tests the same but just
changed the expectations.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary