NNS1-3094: Projects table column styling (#5232)
# Motivation

Currently the "Nervous Systems", "Stake" and "Neurons" columns each get
an equal amount of space.
Because the first one is left aligned and the other 2 are right aligned
but at the same time the project name tends to be longer, it looks like
the gaps between the columns are divided equally.
But if one project has a really long name, it might have to wrap while
there is actually still enough space if the other columns would give up
some.

In the neurons table, we switched to the content columns getting the
width they need and adding additional columns in between to take up the
additional space.

# Changes

1. Add empty columns between the "Nervous Systems", "Stake" and
"Neurons" columns.
2. Give the empty columns style `1fr` to take up additional space.
3. Give the "Nervous Systems" column style `minmax(min-content,
max-content)` so that it doesn't wrap unnecessarily but can wrap if
necessary.
4. Give the "Stake" and "Neurons" columns style `max-content` so they
get the space they need.
5. Drive-by fix: Remove unused `class="title-logo-wrapper"` from
`ProjectNeuronsCell.svelte` which was accidentally copied from another
file.

# Tests

1. Unit tests updated.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/
3. Tested manually with mainnet project names.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet