Activate warning dialog for beta deployment (#5238)
# Motivation

We will release code that hasn't been voted on by the community to
beta.nns.ic0.app.
To make sure people are aware they are using a test deployment, we have
a warning dialog on sign-in.
This was already implemented a while ago in
https://github.com/dfinity/nns-dapp/pull/2570 but never enabled.

# Changes

Enable `IS_TEST_MAINNET` when the `DFX_NETWORK` in the canister args is
set to `"app"`, which is the network we currently use to deploy the
beta.

# Tests

1. The warning dialog itself was tested in
https://github.com/dfinity/nns-dapp/pull/2570
2. Enabling it was tested manually by deploying to
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/
with different values in the canister args.

When you sign in, you see this every time:
<img width="530" alt="image"
src="https://github.com/user-attachments/assets/e7e9b007-3baa-4397-be47-b37e4dd26b9d">


# Todos

- [x] Add entry to changelog (if necessary).