Remove custom added PartialEq from SNS aggregator types (#5275)
# Motivation

We have custom code that adds the `PartialEq` trait to some sns
aggregator Rust types derived from Candid.
Sometimes this breaks the build because a new type must receive the
`PartialEq` trait as well for the old types to be able to keep the
`PartialEq` trait.

But removing the `PartialEq` trait and the custom code that adds it
doesn't seem to break anything.

# Changes

Remove custom code to add `PartialEq` to types that don't need it.

# Tests

CI passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary