Enable following on the new topics (#5277)
# Motivation

Previously, following on the new topics cannot be set on the NNS Dapp,
because NNS Governance did not support it yet. Now NNS Governance has
supported it.

# Changes

Remove a filtering of topic enums that prevented the new topics from
appearing on the followee edit modal. Note that the mechanism to disable
following by `TOPICS_WITH_FOLLOWING_DISABLED` won't be needed in the
future because of https://github.com/dfinity/nns-dapp/pull/5213.

# Tests

* Unit tests
* Manual test to set followees on the new topics:
![Screenshot 2024-08-05 at 3 06
27 PM](https://github.com/user-attachments/assets/d2da3385-f3a9-46e2-a624-375acf9474f3)


# Todos

- [x] Add entry to changelog (if necessary).