NNS1-3139: Add maturity column in projects table (#5287)
# Motivation

We want to make it easy to see all the maturity on all projects by
adding a column to the staking projects table.

# Changes

1. Add `ProjectMaturityCell` component which renders either `-/-` of
there is no data, the maturity if there is data, or nothing if the user
doesn't have neurons for that project.
2. Use the cell component to add a column in the `ProjectsTable`.

# Tests

1. Page objects and unit tests added.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [ ] Add entry to changelog (if necessary).
We haven't done a release yet since the projects table was enabled so
that entry covers this as well.