Round maturity to 2 decimals in neurons tables (#5295)
# Motivation

To make numbers more aligned in the neurons/projects tables we agreed to
always round maturity to 2 decimals.
The tooltip will still show the detailed amount for available and staked
maturity.

# Changes

1. Add an optional parameter `extraDetailForSmallAmount` to
`formatToken*` functions, to configure the detault behavior of showing
more decimals for numbers < 0.01.
2. In the `MaturityWithTooltip` component, which is used in the neurons
table and projects table, set `extraDetailForSmallAmount: false` and
also render 0.00 as 0.

# Tests

1. Unit tests for `formatToken`.
2. Unit tests for `MaturityWithTooltip`.
3. Drive-by: Remove an old TODO.

# Todos

- [x] Add entry to changelog (if necessary).