Import token modal (#5292)
# Motivation

To allow users to import custom tokens, we need to provide a modal with
a form for entering the ledger and optional index canister IDs. This PR
focuses on creating the form to input these IDs, without including
validation, reviewing or additional functionality at this stage.

Demo is available at [beta](https://beta.nns.ic0.app/).

# Changes

- New ImportTokenForm component.
- New ImportTokenModal component.
- Bind canister IDs to be editable in the form and accessible in the
modal.
- Show modal on "Import token" button click. The button is behind the
feature flag.

**Ancillary changes**

- Move the `import_token` label under the `import_token`, because same
wordings we user for the modal title.
- Expose `testId` and `require` props for the `PrincipaInput` component.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.