NNS1-3250: Dispatch event for click on non-link table row (#5333)
# Motivation

In the staking projects table, when you click on a project for which you
don't have any neurons, instead of navigating to the (empty) neurons
table for that project, we want to open the staking modal.
This means that those rows should not be links (which would cause
navigation) but should clicks should instead dispatch an event which can
be used to open the modal.

This PR only implements dispatching the event for non-link rows and does
not yet do anything with it.

# Changes

When a non-link row is clicked in the `ResponsiveTable`, dispatch an
`nnsAction` event with the `rowData` for that row as event payload.

# Tests

1. Unit test added.
2. Tested manually in another branch which already has the staking modal
implemented:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary