Use isNullish instead of ! to check if fee is present (#5337)
# Motivation

The `stakeNeuron` function in `sns-neurons.services.ts` checks if the
fee is loaded before staking.
But it does this by checking if the fee is truthy, which fails if the
fee is zero.
There currently aren't any SNSes with zero transaction fees but
[mockToken](https://github.com/dfinity/nns-dapp/blob/b20eb25d72f8e93a450784adbf1f2ae0cc6dbf1a/frontend/src/tests/mocks/sns-projects.mock.ts#L206)
happens to have a fee of 0 so I wasted some time debugging because of
this.

# Changes

Use `isNullish` instead of `!` to check if the fee is present.

# Tests

1. There is an existing unit test with an absent fee.
2. I added a unit test with a zero fee.

# Todos

- [x] Add entry to changelog (if necessary).