NNS1-3250: Stake button for projects without neurons (#5339)
# Motivation

When a user has 0 neurons for a project, instead of showing "0" in the
table, we want to show a button to stake tokens for that project.
Clicking the button (or anywhere else on the row) will open the staking
modal, and if the user does stake a neuron they will navigate to the
neurons table afterwards.

# Changes

1. Show a "Stake XYZ" button in the neurons cell in the projects table,
if the user has zero neuron.
2. When the user clicks on a non-link row in the projects table,
dispatch a `nnsStakeTokens` event.
3. In the `Staking` component, when an `nnsStakeTokens` event is
received, open the corresponding staking modal.
4. When the staking modal closes, check if there are now a positive
number of neurons and if so, navigate to the neurons table.
5. Don't provide a `rowHref` for projects without neurons, to prevent
navigating to the neurons table.

# Tests

1. Unit test for dispatching the `nnsStakeTokens` event.
2. Unit test for project rows without neurons not being a link.
3. Unit tests for staking modals.
4. Support setting a custom governance canister ID on mock SNSes.
5. Page object support.
6. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [x] Add entry to changelog (if necessary).