chore: introduce decoding quota to http_request in SNS aggregator canister (#5004)
This PR introduces the decoding quota to the `http_request` method of
the SNS aggregator canister. Setting a decoding quota of 10000
(corresponding to roughly 10KB of decoded data incl. overhead) prevents
the SNS aggregator canister from wasting instructions on parsing large
HTTP requests which are not expected to be used (the SNS aggregator
supports GET requests for its assets and thus no HTTP requests should be
larger than a few KB). I've tested the MR manually using dfx
(successfully retrieving the candid interface and the landing page and
failing to make a large HTTP request which is what this PR addresses).