Nns wallet popup (#5311)
# Motivation

With the introduction of the import token feature, a new token popup
with additional actions was added. Each token will have two entries:
"View in Dashboard" and "Token Details". Imported tokens will have an
extra entry: "Remove".

In this pr, we’ve added the popup with the "View in Dashboard" button.
The same popup for the IcrcWallet page will be added in a separate pr.


[Demo](https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/wallet/?u=qsgjb-riaaa-aaaaa-aaaga-cai)

# Changes

- New `LinkToDashboardCanister` component.
- Add "actions" slot to the `WalletPageHeader` to display 3 dots button
next to the token title.
- Add a popup to the `Nns` wallet page with the "View in Dashboard"
button.

# Tests

- Unit tests for LinkToDashboardCanister.
- Added tests for Nns wallet page.
- Tested locally:
<img width="508" alt="image"
src="https://github.com/user-attachments/assets/ace70445-2e59-4b6a-8f46-9117bf2b2509">


# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.