Imported tokens utils (#5351)
# Motivation

For validating the addition of imported tokens, we need to ensure that
the entered ledger canister hasn’t already been imported and that it
isn’t an SNS ledger canister ID. To facilitate this, we’ve added a pair
of utility functions that will simplify the validation process.

# Changes

- Add `isImportedToken` util.
- Add `isSnsLedgerCanisterId` util.

# Tests

- Added unit tests for new utils.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.