Match ledger index pair (#5350)
# Motivation

For validating the ledger and index canister IDs entered by the user, we
will use the `ledgerId` API of the index canister. When the entered
ledger canister ID matches the ledgerId response from the entered index
canister ID, we can confirm that this pair of IDs is suitable for each
other. Here, we add a service that validates the ID matching.

# Changes

- New `matchLedgerIndexPair` service.

# Tests

- Unit test for it.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.