Stop rendering BlessReplicaVersion and RetireReplicaVersion proposals (#5360)
# Motivation

The payload types for the `BlessReplicaVersion` and
`RetireReplicaVersion` proposals has been removed from the
`registry.did` file in https://github.com/dfinity/ic/pull/867 and
https://github.com/dfinity/ic/pull/880.

This means that when we update the candid from which we generate the
Rust types for proposals rendering, we can no longer refer to those
types in our code.

This resulted in build failures in our automated Candid update PR:
https://github.com/dfinity/nns-dapp/pull/5355

Fortunately those proposal types are old and have been purged from the
governance canister so we no longer need to be able to render them.

# Changes

Remove code to render NNS function 5 and 36.

# Tests

Was not tested.

# Todos

- [x] Add entry to changelog (if necessary).