Revert "FOLLOW-1337: Disable FORCE_CALL_STRATEGY (#5376)" (#5393)
This reverts commit
[3e064af8b3988c818d7422e7cdb69a7378aa0d87](https://github.com/dfinity/nns-dapp/commit/3e064af8b3988c818d7422e7cdb69a7378aa0d87).

# Motivation

In https://github.com/dfinity/nns-dapp/pull/5376 we disabled
`FORCE_CALL_STRATEGY`.
This has resulted in flakiness in the end-to-end test
`proposals.spec.ts` caused by a response to an update call for an old
request coming back after a query call for a newer request, resulting in
old data being display in the UI.
We've had similar issues in the past. They can be solved using the
`queuedStore` but until we make the change for loading proposals, we
don't want the test to be so flaky.

Additionally, there have been some concerns regarding extra load caused
by enabling `queryAndUpdate` and I want to make sure we are all on the
same page before we release the change.

For both these reasons, I've decided to revert the change for now.

# Changes

Ran `git revert 3e064af8b3988c818d7422e7cdb69a7378aa0d87`.

# Tests

Pass

# Todos

- [x] Add entry to changelog (if necessary).