NNS1-3281: Add get_tvl to nns-dapp (#5429)
# Motivation

The NNS dapp displays the USD value of the total amount of ICP locked in
neurons.
It gets this information from the TVL canister but we want to move this
functionality to the nns-dapp canister and remove the TVL canister.

The functionality to fetch the amount of locked ICP and the exchange
rate and to get the TVL was already implemented in the `tvl` module.

This PR exposes the `get_tvl` method (which just calls `tvl::get_tvl`)
on the canister and initializes the timers to fetch the necessary data.

# Changes

1. Initialize the timers to update the amount of locked ICP and exchange
rate periodically, when installing or upgrading the canister.
2. Add `get_tvl` canister method.
3. Updated the candid file to match the canister interface.
4. Ran `scripts/nns-dapp/test-exports --wasm nns-dapp_test.wasm.gz
--update-golden --flavour test` and `scripts/nns-dapp/test-exports
--wasm nns-dapp_production.wasm.gz --update-golden --flavour production`
to update golden exports files.

# Tests

1. Added a script to test getting the TVL value from the canister.
2. Test manually end-to-end in another branch.

# Todos

- [x] Add entry to changelog (if necessary).