NNS1-3281: Encode TVL state to stable memory (#5431)
# Motivation

The nns-dapp loads the input to the TVL calculation right after upgrade.
But if this is slow or it fails, it might not be able to provide a TVL
value. So we also want to store the TVL state in stable memory during
the upgrade.

We can't read the TVL state from stable memory unless it's first written
to stable memory.
So we implement the encoding first and only implement the decoding once
we've had 1 release that has written the state to stable memory.

# Changes

1. Implemented both encoding and decoding of the TVL state, but
commented out the decoding and added TODOs.

# Tests

1. Added a test that encoding is done but decoding is not.
2. Added a commented-out assert that decoding is done.
3. Existing upgrade-downgrade test still passes:
https://github.com/dfinity/nns-dapp/actions/runs/10792784452/job/29933422210?pr=5431

# Todos

- [ ] Add entry to changelog (if necessary).
covered under existing entry