Remove $ from directory name (#5433)
# Motivation

It's confusing to have a `$` in the name of a file path.
When using it on the command line, you need to escape it or `$public`
gets replaced with the empty string.
Even more confusing: the resulting path also exists.

# Changes

1. Move `$public` to `public`.
2. Move the corresponding test directories from `_public` to `public` so
the tests are now in a place predictable from the code path.
3. Make `.eslintignore` more specific to avoid ignoring the moved files.

# Tests

Pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary