NNS1-3158: Failed imported tokens store (#5428)
# Motivation

There is no guarantee that the imported token ledger canister’s API
won’t be updated in a way that makes it incompatible, or the canister
became unreachable. As a result, any call to this canister could
potentially fail. To provide appropriate user feedback, the NNS-dapp
should display all such imported tokens as failed. As the first step
toward this, we will collect the failed imported token ledger IDs and
display only the remaining valid ones.

# Changes

- Add `failedImportedTokenLedgerIdsStore` to store failed ledger
canister IDs.
- Update `failedImportedTokenLedgerIdsStore`. Currently,
`IcrcWalletPage` calls
[syncAccounts](https://github.com/dfinity/nns-dapp/blob/c80f3f202d8a90c329a3c283b00f9daf027e74ac/frontend/src/lib/services/icrc-accounts.services.ts#L220),
which calls `loadAccounts` and `loadIcrcToken` in parallel. Each call
can fail, so on any error, we mark the imported token as failed by
adding a record to the failedImportedTokens store. And skip displaying a
toast with the error, since this is not directly an NNS-dapp issue.
- Add `loadedImportedTokens` store, a derived store that returns
imported tokens that are not in `failedImportedTokens`.
- Switch `icrcCanistersStore` (source of the displayed tokens) from
`importedTokensStore` to `loadedImportedTokens`.

# Tests

- Unit tests updated.
- Tested locally.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.