Disable import token validation feature flag (#5436)
# Motivation

To test failed imported tokens in any environment, we’re introducing a
new feature flag that skips the import token validation. Since this
feature flag is intended for testing purposes only, we’re using it in
the same PR.

# Changes

- Add new feature flag.
- Run `scripts/nns-dapp/test-config --update` to update `arg.did` and
`env`.
- Skip the validation when the flag is true.

# Tests

- Pass.
- Tested locally: when the flag is set, we can import any canister ID,
and when it’s not set, the validation prevents this.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.