Add convertNervousSystemParameters function (#5455)
# Motivation

Instead of loading nervous system parameters from the SNS governance
canister, we want to use the parameters present in the SNS aggregator
data that is already loaded. We'll do this by making
`snsParametersStore` a derived store so that we don't have to change how
the parameters are accessed.

However, the type of the parameters in the aggregator data is different
from the type used in the store. So we'll have to convert the data.

In this PR we just add the conversion function.

# Changes

1. Fix the type of `CachedDefaultFolloweesDto` which was wrong before
but wasn't caught until used.
2. Add `convertNervousSystemParameters` to convert the parameters from
`CachedNervousSystemParametersDto` to `SnsNervousSystemParameters`.

# Tests

Added unit tests for full and empty values.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary