Move more logic of remove imported into service (#5466)
# Motivation

Currently the imported token can be removed only from the wallet page.
With the upcoming functionality to remove failed imported tokens from
the tokens table, we need to apply the same logic in another place. To
avoid code duplication, the removal logic has been moved to the service
function.

# Changes

- Simplify the `removeImportedTokens` arguments.
- Move the busy screen handling into the service.

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.