Fix confusing use of 'multiplier' (#5468)
# Motivation

We apply several bonuses to the voting power, such as dissolve delay
bonus and age bonus.
This is implemented by multiplying the voting power with a "multiplier".
For example, if you have a bonus percentage of 25, this means the voting
power gets multiplied by 1.25.
This number 1.25 is the bonus multiplier.

The code was using the name `multiplier` for things that were not
multipliers, which made the code very confusing.

# Changes

1. Fix inappropriate usages of the term "multiplier".
2. Add a comment on the `bonusMultiplier` function clarifying the
parameters.

# Tests

Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary