Make snsParametersStore a derived store (#5471)
We were loading nervous system parameters from the SNS governance
canisters.
This is unnecessary because we already have this data in the aggregator
data.
It was also risky because it means that data might not be loaded when we
need/expect it.

**Note:** I changed the imports from `stores` to `derived` in the first
commit so you can see a cleaner diff by looking at just the other
commits.

1. Make `snsParametersStore` derived from `snsAggregatorStore` instead
of a stand-alone store.
2. Remove all calls to `loadSnsParameters` as they no longer need to be
loaded separately.
3. Remove `loadSnsParameters` itself.
4. Remove `i18n` message `load_parameters` which became unused.

1. Unit tests for new derived store.
2. Updated tests to not expect calls to `loadSnsParameters`.
3. Removed code that fills `snsParametersStore` from
`frontend/src/tests/utils/sns.test-utils.ts`
4. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

- [x] Add entry to changelog (if necessary).