Extract snsAggregatorDerived function (#5494)
# Motivation

We have a number of stores derived from `snsAggregatorStore` which all
map the aggregator data to some record with the root canister ID as the
key.
These all have very similar code.

# Changes

Extracted the duplicate code into a function and used it for
`snsFunctionsStore`, `snsParametersStore` and
`snsTotalTokenSupplyStore`.

# Tests

1. Added a unit test for the new function.
2. The changed derived stores were already covered by tests that are
still passing.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary