FOLLOW-1337: Stop making unnecessary requests for SNS proposal (#5495)
# Motivation

When there is an open SNS swap on the project detail page, we make a
request every 10 seconds for the number of participants and commitment,
as these might change during an open SNS.

However, this also caused the proposal to be refetched each time. This
happened because the proposal fetch logic was derived from the summary
store, which is updated each time the derived state is fetched. This is
useless because the SNS proposal won't change if the swap is already
open.

By adding an intermediate `proposalId` variable, we make sure that we
don't fetch the proposal again when the summary changes, but the
proposal ID doesn't.

# Changes

Change `ProjectProposal.svelte` to only load the proposal when the
proposal ID changes.

# Tests

1. I first created the test to see the proposal being fetched multiple
times. After the changes the test passed.
2. I also tested it manually with a log statement inside
`queryAndUpdate`.

# Todos

- [x] Add entry to changelog (if necessary).