Clean up magic numbers from SNS neuron tests (#5497)
# Motivation

These 2 tests had some unexplained numbers in them.
I tried to make the numbers in the setup more explicit so it's clearer
where the numbers in the tests come from.

# Changes

1. Declare constants with specific value.
2. Explain how the expected results derive from the inputs.
3. Merge `"should render voting power"` into `"should render voting
power and description if neuron can vote"` because the second
expectation explains the value of the first expectation.

# Tests

Only tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary