Fix canister IDs import script (#5517)
# Motivation

The replica with dfx 0.23.0 does not appear to serve `.raw.` URLs
correctly.
Before the service worker was removed, we could only use `curl` to fetch
raw URLs.
So when importing canister IDs from a URL, we would insert `.raw` into
the URL before fetching it with `curl`.
This no longer works, but is also no longer necessary.

# Changes

Remove the code that inserts `.raw` into the URL to fetch canister IDs
from.

# Tests

1. Tested manually.
2. Added golden tests to CI that use the script on the local replica and
on mainnet.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary