FOLLOW-1337: Expose race condition in NnsProposals (#5522)
# Motivation

We want to re-enable the update calls that are disabled via
`FORCE_CALL_STRATEGY`.
But this currently results in the following race condition:
1. Load proposals, a certified and an uncertified request are made.
2. The uncertified response data is displayed.
3. Change filters, 2 new requests are made.
4. The second uncertified request returns but then the first certified
request returns.
5. This results in displayed proposals going back and forth between
different responses.

This PR just adds a test that demonstrates the problem.
The issue will be fixed in a follow-up PR.

# Changes

Add a test the goes through the above scenario and expects the wrong
result.

Drive-by: Remove unused `overrideFeatureFlagsStore` from the test.

# Tests

Test only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary