Clean up public/proposals.services.spec.ts (#5525)
# Motivation

Make the test easier to maintain.

Using stores is cheap. We shouldn't care how the store got its content.
What matters is that the contents is what it should be.
So instead of mocking store methods, we should expect the store contents
itself.

# Changes

1. Use `resetIdentity` and `setNoIdentity` instead of mocking the
`authStore`.
2. Check the contents of the `toastsStore` instead of mocking the calls
to `toastsShow`.
1. Also check the actual error message in the toast rather than only
that there is a toast.
3. Check the contents of `proposalsStore` instead of mocking its
methods.
4. Check the contents of `proposalPayloadsStore` instead of mocking its
methods.
5. On the remaining mocks, replace `mockImplementation` with simpler
methods that directly specify return/resolve/reject values.
1. In 1 case there is a mock implementation that should be async but
it's not. But fixing it changes the behavior, breaking the test. So I'll
fix it in a separate PR.
6. When multiple calls are expected, expect the args for all of them
instead of only the last one.

# Tests

Test only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary