FOLLOW-1337: Use queuedStore for proposalsStore (#5532)
# Motivation

We want to re-enable the update calls that are disabled via
FORCE_CALL_STRATEGY.
But this currently results in a race condition. See
https://github.com/dfinity/nns-dapp/pull/5522 for details.

To fix the race condition, we use a `queuedStore`, similar to
https://github.com/dfinity/nns-dapp/blob/main/frontend/src/lib/stores/icp-account-balances.store.ts

# Changes

1. Change `proposalsStore` to use a `queuedStore`.
2. Change calls to modify `proposalsStore` to use the single-mutation
store as necessary to use a queued store.

# Tests

1. Update tests to use `...ForTesting` mutation functions which hide the
use of a single-mutation store.
2. Resolve the TODO in
`frontend/src/tests/lib/pages/NnsProposals.spec.ts` which was added to
observe the race condition.
3. Add unit tests for `proposalsStore` which check the correct order of
mutations.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary until we actually enable query+update again.