Use a non-aborted SNS as mock data (#5549)
# Motivation

The `aggregatorSnsMockDto` has data for an aborted SNS. This makes it
very non-standard/surprising when used in a test.
While in principal, a test should specify all the fields that it cares
about, I think it's reasonable to expect the test data not to be for an
aborted SNS.

# Changes

1. Use `aggregatorMockSnsesDataDto[4]` instead of
`aggregatorMockSnsesDataDto[7]` as `aggregatorSnsMockDto`.
2. Added a comment explaining why `4`.
3. `export convertDtoToTokenMetadata` to make sure `aggregatorTokenMock`
matches `aggregatorSnsMockDto` (as required by the comment) without
having to hard code the same data.

# Tests

Updated tests that started failing by specifying more data explicitly in
the tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary