Restore changelog after rollback (#5551)
# Motivation

After last week's upgrade, we had to revert to the previous version.
I kept the changelog for the proposal but I think it makes sense to
restore the unreleased changelog so that the changes are reflected in
the next proposal as what has changed since the current prod release.

# Changes

Revert just `CHANGELOG-Nns-Dapp-unreleased.md` from
https://github.com/dfinity/nns-dapp/pull/5536

# Tests

no

# Todos

- [x] Add entry to changelog (if necessary).