Remove mockAuthStoreSubscribe (#5571)
# Motivation

Svelte stores are light weight objects that don't need to be mocked in
tests.
One store that is mocked particularly often is the `authStore`.
For new tests we already use the real `authStore` (mostly via
`resetIdentity` and `setNoIdentity`) but we still have a lot of old test
that mock it.
Keeping bad practices around is bad because it makes it harder to copy
existing code and do the right thing.
It can also be confusing because once the `authStore` is mocked, setting
the real `authStore` has no effect. So mixing old and new practices in
the same file doesn't work.

# Changes

1. Replace `mockAuthStoreSubscribe` with `resetIdentity` and
`mockAuthStoreNoIdentitySubscribe` with `setNoIdentity`.
2. In `frontend/src/tests/lib/services/sns-sale.services.spec.ts`
replace a custom `subscribe` mock with `setForTesting`.
3. Remove `mockAuthStoreSubscribe` and
`mockAuthStoreNoIdentitySubscribe`.

# Tests

Tests only.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary