Enable TVL from nns-dapp (#5541)
# Motivation

The nns-dapp canister now exposes a `get_tvl` method compatible with the
TVL canister.
This means we can set the `nns-dapp` canister ID as the `tvl` canister
ID to get the TVL from the nns-dapp canister.
It also means that we can show TVL on local development.

# Changes

1. Remove `{#if ENABLE_METRICS}` around `TotalValueLocked` and `Metrics`
(the latter takes care of fetching the numbers).
2. Use `OWN_CANISTER_ID` as the `TVL_CANISTER_ID`. On `app` (i.e. Beta)
use the prod nns-dapp because the nns-dapp on the app subnet doesn't get
free exchange rate requests so it won't have a non-zero TVL.
3. Remove `ENABLE_METRICS`.

# Tests

1. Unit tests updated.
2. Golden configs updated.
3. Screenshots updated.
4. Tested manually.

# Todos

- [x] Add entry to changelog (if necessary).