Move isForceCallStrategy and notForceCallStrategy to mockable.constants.ts (#5582)
# Motivation

`isForceCallStrategy` and `notForceCallStrategy` only depend on
`FORCE_CALL_STRATEGY`.
By putting them in a different file, it creates additional dependencies
for workers that depend on anything else in that file.

I noticed this while making a PR to initialize `FORCE_CALL_STRATEGY`
randomly for a percentage rollout and logging what it was initialized
to, because I noticed 2 log lines every time I loaded the app: 1 from
the main app and 1 from a worker.

# Changes

1. Move `isForceCallStrategy` and `notForceCallStrategy` into same file
as `FORCE_CALL_STRATEGY`.
2. Update imports.
3. Add default mock values for the functions to satisfy the compiler.

# Tests

By adding a log statement, tested that
`frontend/src/lib/constants/mockable.constants.ts` is now loaded once
instead of twice.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary