Include actor in bump PR description (#5588)
# Motivation

We have a GitHub workflow to create gix-components or ic-js bump PRs.
Usually the person requesting the bump PR can also approve it
themselves.
But sometimes peterpeterparker requests one and then I approve it.
So it's useful to see who requested the PR.

# Changes

Mention the workflow actor in the PR description.

# Tests

Created https://github.com/dfinity/nns-dapp/pull/5587

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary