Don't hide neurons which only have staked maturity (#5591)
# Motivation

We hide neurons with zero stake.
The exact logic we use for this is that we hide neurons for which the
`stake+maturity < fee` (or `< 0` for SNS).
The problem is that this fails to take staked maturity into accounts.

It's tricky to get a neuron with staked maturity without stake because
if the neuron doesn't have a dissolve delay, the staked maturity is
automatically reverted to normal maturity. So the repro steps are:

1. Stake a neuron
2. Set a dissolve delay on the neuron.
3. Wait for maturity > fee to build up.
4. Dissolve the neuron.
5. Disburse the neuron.
6. Add a dissolve delay to the neuron.
7. Stake the maturity.

In the test environment, steps 2-4 can be replaced with just adding
maturity.

# Changes

Take staked maturity into account for wether to hide NNS and SNS
neurons.

# Tests

1. Unit tests added.
2. Tested manually.

# Todos

- [x] Add entry to changelog (if necessary).