A script to vote on proposals (#5600)
# Motivation

In local testing it can be convenient to quickly vote one or all
proposals through.
Especially since snsdemo8 has a neuron with a lot of voting power.

# Changes

Added a script that can be used to vote on one or all proposals without
having to look up neuron IDs.

# Tests

Tested manually.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary