Ignore imported tokens for redirection when signed out (#5602)
# Motivation

In PR [#5596](https://github.com/dfinity/nns-dapp/pull/5596), we began
checking for the loading of imported tokens to handle navigation from
the wallet page for unknown tokens. This fixed the targeted issue but
unintentionally broke redirection to the tokens page after sign-out. To
restore the sign-out redirection, we also need to consider the login
state.

# Changes

- Ignore imported token readiness when signed out, as we don’t load them
in this state.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.