Make vote script vote in parallel (#5605)
# Motivation

Voting on a lot of proposals can be quite slow.

# Changes

Call the script multiple times in parallel when voting on multiple
proposals.
The `&` at the end of the command runs the command in the background.
At the end we wait for all the processes before exiting the script.

# Tests

Tested manually

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary