Rename DIDC_VERSION to DIDC_RELEASE (#5622)
# Motivation

The current `didc` version (that is the output of `didc --version`) is
`0.4.0` which is installed from release `2024-05-14`.

While trying to debug why the recent "Update sns_aggregator candid
bindings"
[failed](https://github.com/dfinity/nns-dapp/actions/runs/11338847183),
I got very confused because I had a different version installed.
So I want to change our scripts to check that you have the correct
version installed.
For this, I need to add what is the correct version to our repo config.
To obvious name for this is `DIDC_VERSION` but we already use
`DIDC_VERSION` to denote the current release.

So before adding `DIDC_VERSION` we should rename `DIDC_VERSION` to
`DIDC_RELEASE` to more accurately reflect what it is.

# Changes

Rename `DIDC_VERSION` in `dfx.json` (and everywhere that refers to it)
to `DIDC_RELEASE`.

# Tests

CI passes

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary