Menu padding adjustments (and bump gix-components) (#5630)
# Motivation

We had 2 PRs in gix-components
(https://github.com/dfinity/gix-components/pull/501 and
https://github.com/dfinity/gix-components/pull/504) to fix paddings in
the menu.

We plan to make another PR to remove the `MenuBackground` and put the
logos in the menu itself.
This will remove the need to add a bottom margin below the GitHub button
as the logo will then take up its own space.
We'll then also fix the padding above the bottom logo to be 24px as [in
the
design](https://www.figma.com/design/x8dk5axQUf8Bi43uUsqCHc/NNS-Production-%26-Hand-over?node-id=4851-54392&t=08Ps7AaObh2JKt3d-4).
To anticipate that, we already change the bottom margin in this PR as
well to match that.

# Changes

1. Bump gix-components: `npm run update:gix`.
2. Adjust the bottom margin in the `.menu-footer`.

# Tests

Updated screenshots.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary