Fix Hash reload issue (#5625)
# Motivation

There is an existing issue related to stopPropagation onClick events, it
was stopping internal Svelte routing system to catch the routing event
and resolve DOM diff correctly, that is why it was causing a full
reload.

We don't want this to happen in any component who uses Hash that also
need navigation on onclick events.

You can see similar structure in NeuronsTable which has this reload
issue, if you click on the Hash not the row, the page is reloading right
now.

# Changes

PreventDefault is added so onclick on Hash wouldn't trigger routing when
used in a parent which changes the route on click.

# Tests

manually tested

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary