Use page objects for Hash.spec.ts (#5641)
# Motivation

Make the test easier to maintain.
In particular this helps with updating the test dependencies because the
rendered container is no longer the document body so the way this test
tries to find the tooltip doesn't work.

# Changes

1. Use page objects.
2. Expect explicit values instead of constructing values the same way
the code under test does.
3. Escape the element ID passed to `body.querySelectorAll` because
apparently these are not allowed to start with a digit. This could be
done with `CSS.escape` but unfortunately that's not supported in
Playwright tests.

# Tests

The test still passes.
I did not verify that this work with the new dependencies, but if it
doesn't we should fix it in the page object, not in the test.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary