Wait longer for lifecycle in aggregator test (#5654)
# Motivation

The aggregator test on CI [is
flaky](https://github.com/dfinity/nns-dapp/actions/runs/11403403979/job/31730884616?pr=5652).
Waiting for the lifecycle to change timed out because swap canisters no
longer update their state on every heartbeat but only once a minute.

# Changes

Increase the timeout when waiting for the lifecycle to change.

# Tests

Not tested except for on CI.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary