Stop mocking toastsStore (#5666)
# Motivation

To make the tests more robust we should not mock stores but use real
stores.

By doing so, I also found a bug:
In https://github.com/dfinity/nns-dapp/pull/4244 I removed
`sns_amount_not_enough_stake_neuron` because I thought it was not used.
But it turned out it was used and cleaning up the tests exposed that, so
I restored it in this PR as well.

# Changes

1. Stop mocking `toastsStore` functions.
2. Expect the full contents of the store.
3. Expect the store to be empty before it's filled.
4. Expect actual messages instead of label keys.
5. Restore `sns_amount_not_enough_stake_neuron`. I also removed `a
neuron` from the message, in line with the intent of
https://github.com/dfinity/nns-dapp/pull/4244

# Tests

1. The updated tests still pass.
2. In one case the test showed that the toast was showing a label key
instead of a message. The test has been fixed by adding back the removed
message.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary