Stop adding .raw to aggregator URL (#5668)
# Motivation

When connecting from localhost to a non-local environment (such as
DevEnv or mainnet) we add `.raw` to the SNS aggregator URL.
This was originally done to avoid being redirected by the boundary node
to the service worker, because that would result in CORS issues when
connecting from localhost.
But:
1. The boundary node no longer redirects to the service worker
2. Raw URLs stopped working on DevEnv ([slack
thread](https://dfinity.slack.com/archives/CGA566TPV/p1727177718754779))

So adding `.raw` to the aggregator URL is no longer necessary and also
breaks.
Time to remove this code.

# Changes

1. Stop adding `.raw` to the SNS aggregator URL.
2. Remove `addRawToUrl` as it is no longer user.

# Tests

Tested manually:
1. From localhost to localhost
2. From localhost to DevEnv
3. From DevEnv to DevEnv
4. From localhost to mainnet

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary