Replace doMock with mock (#5682)
# Motivation

Because [vi.doMock is not hoisted to the top of the
file](https://vitest.dev/api/vi.html#vi-domock), only the next dynamic
import of the module will be mocked. This could lead to unpredictable
errors when some modules are not mocked (e.g., ReferenceError: Worker is
not defined).

# Changes

- Replace doMock with mock.

# Tests

- Pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary.