Unlock neuron in e2e tests (#5689)
# Motivation

Because neurons are now locked for 7 days when created, we had to
disable the disburse e2e test in
https://github.com/dfinity/nns-dapp/pull/5686

This PR adds a test-only button to unlock a neuron and re-enables the
disburse e2e test and restores the unlocked neuron in the neurons table
screenshot.

# Changes

1. Add a button that uses the test-only `update_neuron` API to unlock a
neuron. The button is added in `NnsNeuronTestnetFunctionsCard` so it
will only appear when testing.
2. Use the button in the disburse e2e test and enable the test again.
3. Use the button in the neurons table e2e test and update the
screenshot.

# Tests

1. Updated tests pass.
2. Unlock button tested manually locally.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary