FOLLOW-244: Add refreshNeuronIfNeeded service (#5696)
# Motivation

Increasing a neurons' stake is a 2 step process:
1. Transfer ICP to the neuron's account.
2. Notify the governance canister to refresh the neurons.

Because this process can be interrupted after step 1, we need to be able
to finish a half completed process.
Currently the nns-dapp canister has a list of neuron accounts for each
user and it watches every ICP ledger transaction to see if it is to a
neuron account and then it refreshes that neuron.
We want to stop keeping this list in the nns-dapp canister and we want
to stop watching every transaction.
So instead we will check from the frontend if a neuron's stake matches
its account's balance and refresh if needed.

This PR just adds a service function to check if the balance matches and
refresh if it doesn't.
It also uses existing `checkedNeuronSubaccountsStore` to avoid checking
too often.

# Changes

1. Add `refreshNeuronIfNeeded` function to `neurons.services.ts`.

# Tests

1. Unit tests added.
2. Tested manually in another branch in which the nns-dapp canister was
changed to no longer perform this actions described above. I also tested
with a neuron from which proposals were created to check that the
expected stake still matches the balance.

# Todos

- [ ] Add entry to changelog (if necessary).
Will add when the service function is used.