Keep default neuron selected on deselect-all (#5707)
# Motivation

When setting neuron visibility of one neuron, you are offered to select
other neurons you might want to change.

<img width="608" alt="image"
src="https://github.com/user-attachments/assets/80003b80-05b5-4752-b01b-52626e2e3480">

If all neurons are selected and you deselect the "Apply to all neurons"
checkbox, all neurons become unselected.

If you arrived at this modal from the details page of one specific
neuron it would be useful to keep that neuron selected when unselecting
all. Because the confirm button is anyway disable when none are selected
and you intended to change that neuron to start with.

# Changes

When deselecting all neurons, if there is a default neuron, keep the
default neuron selected.

# Tests

1. Refactor checkbox checking in unit test.
2. Add missing unit test for deselecting all without default neuron.
3. Add unit test for deselecting all with default neuron.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary