Remove order dependence from AccountDetails.spec.ts (#5716)
# Motivation

The first test in `AccountDetails.spec.ts` sets a value on accounts
stores and the other tests rely on this value being set.
This means that if test are run individually or in a different order,
they fail.

# Changes

Move the setup from the one test to `beforeEach`.

# Tests

The last test was failing before when run as the `.only` test and now it
passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary