Allow dfx-canister-url to make raw local URLs (#5717)
# Motivation

`dfx` 0.24.1 uses ic-http-certification instead of icx-proxy which means
that metrics now need to be requested via a `.raw` URL locally just like
on mainnet.

This breaks
[e2e-test-metrics-present](https://github.com/dfinity/nns-dapp/blob/main/scripts/nns-dapp/e2e-test-metrics-present).
`e2e-test-metrics-present` actually already tries to use a raw URL by
passing `--raw` to `dfx-canister-url`
[here](https://github.com/dfinity/nns-dapp/blob/8e2701a57ed725f30e23aa5277fb179f1deacd99/scripts/nns-dapp/e2e-test-metrics-present#L21).

But `dfx-canister-url` ignore this flag on local.
Since raw URLs are now required on local, we should stop ignoring the
flag.

# Changes

Remove the code that sets `RAW_PART` to the empty string for the local
network.

# Tests

1. Manually ran `scripts/dfx-canister-url --raw nns-dapp` to see that it
now has a `.raw` URL.
2. Ran `scripts/nns-dapp/e2e-test-metrics-present`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary

---------

Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>