FOLLOW-244: Add API to claim a neuron (#5729)
# Motivation

Staking a neuron requires 2 steps: transferring the stake and claiming
the neuron.
Both steps are done in the canister code in ic-js.
But it's possible that the process get interrupted in between.
In this case the nns-dapp canister finishes the process.
We want to move this fallback mechanism to the frontend.
So the frontend should be able to claim a neuron as a separate step.
The governance API already has a `claimOrRefreshNeuron` function but it
requires a `neuronId` as parameter so it can really only be used for
neurons that already exist.

# Changes

1. Add `claimOrRefreshNeuronByMemo` which allows claiming a neuron for
which we do not know the neuron ID yet.
2. Add `claimOrRefreshNeuronByMemo` to api-service as well.

# Tests

1. Unit tests added.
2. Tested end to end in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary