FOLLOW-244: Add service to claim unclaimed neurons (#5731)
# Motivation

Staking a neuron requires 2 steps: transferring the stake and claiming
the neuron.
Both steps are done in ic-js.
But it's possible that the process gets interrupted in between.
In this case the nns-dapp canister finishes the process.
We want to move this fallback mechanism to the frontend.

This PR adds a service function which can detect unclaimed neurons from
transactions and known neuron accounts.

# Changes

1. Add `claimNeuronsIfNeeded` function.

# Tests

1. Add unit tests.
2. Tested manually end-to-end in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet