FOLLOW-244: Do not try to claim the same neuron multiple times in a session (#5733)
# Motivation

When the frontend sees a staking transaction without a corresponding
neuron, it tries to claim the neuron in case claiming it was
interrupted.
We don't need to do this more than once so we store the account
identifier to at least not do it more than once per session.

# Changes

1. If we find a staking transaction, add the neuron account identifier
to `checkedNeuronSubaccountsStore` and if it was already there, don't
try to claim the neuron.

# Tests

1. Unit test added. Tested that the test fails without the new code.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary