Enforces an exact match of replica processes (#5732)
# Motivation

A script runs `pgrep -l replica` to check for replica processes, but it
can't receive false positives from a process called `replicatord`.

```sh
$ pgrep -l replica
781 replicatord
95113 replica
```

# Changes

- The script checks for exact matches with the keyword `replica`.

# Tests

I couldn't run the script that starts the snapshot without this change
because it was complaining about a process (`replicatord`) matching the
pattern. This change allowed me to run it without issues. I also tried
to start a second snapshot, and it correctly informed me that a replica
was already running.

# Todos

- [ ] Add entry to changelog (if necessary) - Not necessary