FOLLOW-244: Remove MultiPartTransactionToBeProcessed::TopUpNeuron (#5735)
# Motivation

https://github.com/dfinity/nns-dapp/pull/5706 removed the functionality
to top up neurons from the nns-dapp canister.
Because `TopUpNeuron` transactions could be on the queue when the
canister is upgraded, I didn't remove the
`MultiPartTransactionToBeProcessed::TopUpNeuron` enum value yet.
Now that the current version on mainnet no longer adds `TopUpNeuron` to
the queue, it should be safe to remove this value from the enum type.

# Changes

1. Remove `TopUpNeuron` from `MultiPartTransactionToBeProcessed`.

# Tests

CI passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary