Remove subscribe from sns.services.spec.ts (#5741)
# Motivation

Some tests in `sns.services.spec.ts` subscribe to `snsAggregatorStore`
and try to get/create a wrapper inside the listener function. But
because the test never unsubscribe, the listeners continue being called
during other tests.
If another test adds different data that the subscribing test doesn't
expect, it looks like one test fails while another test is running.

Tests that subscribe to a store should also unsubscribe. But in this
case, subscribing it totally unnecessary and just makes the test more
complicated.

# Changes

1. Rewrite the tests to not use subscribe.

# Tests

1. The 2 changed tests still pass and fail if the call to
`loadSnsProjects()` is removed.
2. In another branch a new test I tried to add was failing because of
this, and unsubscribing fixed it.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary