FOLLOW-244: Stop claiming neurons in the nns-dapp canister (#5738)
# Motivation

Staking a neuron requires 2 steps: transferring the stake and claiming
the neuron.
Both steps are done in ic-js.
But it's possible that the process gets interrupted in between.
In this case the process needs to be finished later.
This is now done in the frontend so it no longer needs to be done in the
backend.

# Changes

1. Remove logic to detect stake neuron transaction and claim neurons
from the nns-dapp canister code.

# Tests

1. Unit test removed.

# Todos

- [x] Add entry to changelog (if necessary).
existing entry updated