FOLLOW-244: Stop storing neuron accounts in stable memory (#5739)
# Motivation

The nns-dapp backend canister stores neuron accounts to be able to
detect neuron top-up transactions.
The canister no longer does the neuron top-ups so the neuron accounts
aren't used anymore.
So we can stop storing them in stable memory during upgrades.
We can't change the structure of the stable memory so now we just store
an empty map in pre-upgrade, and read nothing (of type `Reserved`) in
post-upgrade. This is forward and backward compatible.

I originally thought changing the encoding will require multiple steps,
but this way it can be done in one step.

There will be additional PRs to remove the `neuron_accounts` field
completely.

# Changes

1. When encoding `AccountsStore`, write and empty
`HashMap<AccountIdentifier, candid::Empty>` instead of
`neuron_accounts`.
2. When decoding `AccountsStore`, ignore whats read in the position of
the `neuron_accounts`. Declaring it as type `candid::Reserved` allows
any data to be read there without error.

# Tests

1. I manually installed a wasm without these changes, created a neuron,
upgraded to a wasm with these changes, created another neuron, and
downgraded again. The neurons were still there and neither the upgrade
nor downgrade caused any errors.
2. Remove `neurons_created_count` as an invariant preserved during
upgrade because it no longer is.

# Todos

- [x] Add entry to changelog (if necessary).