Add beta network to dfx.json (#5743)
# Motivation

We are currently using the `app` network canister for the beta release.
We want to move the beta release to Orbit for more secure management and
make the `app` canister available of ad-hoc testing again.

# Changes

1. Add a `beta` network to `dfx.json` which has all the same values as
`mainnet` and `app` except for the `nns-dapp` canister ID, which is the
one controlled by the NNS team Orbit station.
2. Set `IS_TEST_MAINNET` to true for the `beta` network in addition to
the `app` network so that we show the warning banner.

# Tests

1. Manually tested that the warning banner is show when the network is
set to `beta` or `app`.
2. Ran `DFX_NETWORK=beta ./config.sh` and inspected that the arguments
are as described
[here](https://dfinity.slack.com/archives/D04PTGSF6KX/p1730224706878919).

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary