Remove neuron_accounts field from AccountsStore (#5767)
# Motivation

The nns-dapp canister used to have a list of all the users' neuron
accounts.
This was used to detect neuron top-up transactions as a fallback when
neuron top-up failed in the frontend.
The fallback mechanism has now been moved to the frontend so we stopped
populating the `neuron_accounts` field in the `AccountsStore` in the
nns-dapp canister.

This PR removes the unused `neuron_accounts` field.

A future PR will remove the `neurons_created_count` field from the
stats, as it will now always be 0 because the data from which it was
derived is gone.

# Changes

1. Remove the unused `neuron_accounts` field from `AccountsStore`.

# Tests

1. Stop populating the `neuron_accounts` field in `toy_data.rs`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary