Always build SNS wrappers from aggregator data (#5768)
# Motivation

There are currently 2 ways SNS wrappers can be created:

1. When wrappers are requested, if they are not already cached, a call
is made to SNS-W to get a list of SNSes. Then a call is made to each
root canister to get a list of canister IDs for each SNS. From these
wrappers are created and cached.
2. When SNS aggregator data is loaded, it contains all the canister IDs
of all the SNS canisters and this is used to preemptively populate the
wrappers cache.

Because we always load aggregator data on app load, the only reason #1
happens is because of a race condition and this results in a lot of
unnecessary calls. So we want to always rely on the aggregator data for
creating the SNS wrappers.

# Changes

1. When wrappers are requested, wait for SNS aggregator data to be
loaded and use that to create wrappers.
2. Stop populating the cache when aggregator data is loaded.
3. Remove code that became unused.

# Tests

1. Change how `SnsWrapper` creation is mocked in unit tests and populate
aggregator data in those tests.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/launchpad/

# Todos

- [x] Add entry to changelog (if necessary).