Fix test order dependency in SnsWallet.spec.ts (#5774)
# Motivation

The `'should have canister links in "more" popup'` test in
`SnsWallet.spec.ts` relies on `icrcLedgerApi.queryIcrcBalance` being
mocked. But this happens in the `beforeEach` of a `describe` block which
the test is not a part of.

So the test fails when run separately with `it.only()`.

# Changes

1. Put the test inside the `describe("loading accounts"` block.

# Tests

The test passes when run separately with `it.only`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary