Decode neuron_accounts into original type instead of Reserved (#5805)
# Motivation

The latest nns-dapp upgrade failed because the stable memory couldn't be
decoded.

```
Decoding stable memory failed. Error: "Fail to decode argument 4 from table14 to reserved"
```

In my understanding, anything should be able to be decoded to type
`candid::Reserved` and with local testing this is the case but
apparently there are edge cases where this is not true.

It's still not clear to me why it failed but I was able to reproduce the
failure against golden test data and the changes in this PR fixed it.

I'll try to continue to look into this more to understand what really
happened but meanwhile it's good to have a releasable repo again.

# Changes

Decode unused `neuron_accounts` as its original type
`HashMap<AccountIdentifier, NeuronDetails>` instead of
`candid::Reserve`.

# Tests

1. I tried upgrading nns-dapp with the failed release candidate from
https://nns.ic0.app/proposal/?u=qoctq-giaaa-aaaaa-aaaea-cai&proposal=134171
against the golden NNS state and was able to reproduce the decoding
error.
2. Then I built a wasm with the type change from this PR and was able to
upgrade NNS dapp without errors.

I made the following changes to the golden state test to be able to test
with nns-dapp:
https://github.com/dfinity/ic/pull/2657

# Todos

- [x] Add entry to changelog (if necessary).