Adds nvmrc file with recommended node version (#5831)
# Motivation

The project specifies "recommended" node versions in `package.engines:`
but this is merely a suggestion. By adding an `.nvmrc` file with a
specific version, we can easily load it using `nvm use`.

This also creates a commit history for the supported node versions used
in the project's development.

Note: The nvm module is required.
# Changes

Adds an `.nvmrc` file to the repository with the Node.js LTS version
`v18.20.5`.

# Tests

- Tested manually by running `nvm use`

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.