Filter out SNSes without lifecycle (#5830)
# Motivation

When an SNS swap canister runs out of cycles, the next time the SNS
aggregator tries to fetch the data for this SNS, it will overwrite its
existing data (at least the data that comes from the swap canister) with
empty default data.
This means it will no longer have a value for the `lifecycle` field and
this breaks the code that tries to filter out SNSes which are aborted.

We should probably fix the aggregator to not drop information when the
canister runs out of cycles but this will take longer.
In the meantime it's better to hide one SNS than for the entire NNS dapp
to break.

# Changes

1. Filter out SNSes without `lifecycle` in addition to SNSes with
`lifecycle === SnsSwapLifecycle.Aborted` in the aggregator store.
2. Filter out SNSes without `lifecycle` when creating SNS summaries.

# Tests

1. Unit tests added.
2. Updated `snsWithLifecycle` which took 2 unnamed parameters to take
named parameters.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Yusef Habib <yusef.fernandez@dfinity.org>