NNS1-3450: Extends csv utility function to support labels (#5836)
# Motivation

The CSV table header is currently paired with the keys of the object
from the array being converted into a CSV. This requires either using
"translated" keys or generating a table with very technical names for
the header.

# Changes

- Extends `generateCsvFileToSave` and `convertToCsv` to expect a label
for the headers so they are decouple from the objects keys.

# Tests

- Added unit test

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Prev PR: #5834  | Next PR: #5813