FOLLOW-244: Notify for unfinished canister top-ups from the frontend (#5859)
# Motivation

# Motivation

Topping up a canister is a 2-step process:

1. Transfer ICP to a subaccount of the CMC
2. Notify the CMC of the transfer
Because the process might be interrupted between the 2 steps, the
nns-dapp canister listens for all transactions to see if any might be a
canister top-up transaction by an NNS dapp user.

We want the NNS dapp to stop listening for all transactions so we want
to move this fallback mechanism to the frontend.

# Changes

1. Call the service function that was added in
https://github.com/dfinity/nns-dapp/pull/5857 from `CanisterDetails` to
actually do the notifying if needed.

# Tests

1. Unit tests added.
2. Tested manually by removing the top-up functionality from the backend
canister and reloading the page in the middle of the top-up process.

# Todos

- [x] Add entry to changelog (if necessary).