FOLLOW-244: Stop topping up canisters in the backend (#5862)
# Motivation

Topping up a canisters requires 2 steps: transferring the funds and
notifying the CMC.
Both steps are done in ic-js.
But it's possible that the process gets interrupted in between.
In this case the process needs to be finished later.
This is now also [done in the
frontend](https://github.com/dfinity/nns-dapp/pull/5859) so it no longer
needs to be done in the backend.

# Changes

1. Remove the code to detect canister top-up transactions.
2. Remove the code to notify the CMC of canister top-ups.
3. Keep `MultiPartTransactionToBeProcessed::TopUpCanisterV2` until the
next release, in case there are still instances of it in the queue
during the upgrade.

# Tests

1. Removed a bash script test.
2. Apparently there were no unit tests.
3. Tested manually that the fallback still works, now done by the
frontend.

# Todos

- [x] Add entry to changelog (if necessary).