Give useful output in scripts/vote (#5866)
# Motivation

I was trying to use `scripts/vote` to vote on a proposal on my DevEnv
but it didn't do anything because I was using the wrong identity which
didn't have any neurons.
This was confusing because the script didn't output anything and didn't
give any error.

# Changes

Provide some useful output about the neurons and number of proposals.

# Tests

Tested manually by specifying a wrong identity and a correct identity.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary