NNS1-3480: Fetch all transactions for an account (#5868)
# Motivation

In the scope of feature NNS1-3480, we want to be able to fetch all
transactions for a given `accountId`. The endpoint can't not return all
transactions in one go so we have to keep calling until there are no
more.

It adds a configurable hard limit to the number of times it does the
recursion.

# Changes

- Adds a new function `getAllTransactions` to fetch all transactions for
a given `accountId`

# Tests

- Unit tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Next PR: #5873