NNS1-3480: Get all the transactions for a list of addresses concurrenlty (#5873)
# Motivation

We need to retrieve all transactions for every subaccount associated
with the logged-in user. We want to make these calls concurrently to
minimize the total time, limiting it to the duration of the longest
call. Currently, a call to the Index canister takes approximately 1.2
seconds.

This will be later used by the download functionality that will get all
accounts and the principal and invoke it. Later it will map this
response to feed it into the csv generator.

Todo: The error path needs further development, but it first requires
changes to `getAllTransactionsFromAccountAndIdentity`.


# Changes

- Adds a function to fetch all transactions for a given list of accounts
concurrently.

# Tests

- Added unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Prev. PR: #5868