Add service to load ICP Swap tickers (#5897)
# Motivation

We want to show USD values of assets. For this we want to use data from
ICP Swap.

This PR adds a function to load ticker data and put it into the store.

# Changes

1. Add `loadIcpSwapTickers`. We don't have product requirements yet so
for now we don't do anything with an error except logging it on the
console.

# Tests

1. Added unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet