Add derived store with token prices in ckUSDC (#5899)
# Motivation

We want to show USD values of assets. For this we want to use data from
ICP Swap.

This PR creates a stored derived from the tickers store, which holds
ckUSDC prices of all available tokens based on their ledger canister ID.
This depends on the `ckUSDC_ICP` ticker being available and `XYZ_ICP`
tickers being available for the relevant tokens.

# Changes

1. Add `icpSwapUsdPricesStore`.

# Tests

1. Add unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet