Remove ParticipateSwap transaction type from backend (#5906)
# Motivation


[get_transaction_type](https://github.com/dfinity/nns-dapp/blob/b1bc79f602907794dac188d7a2000fb654e0fe90/rs/backend/src/accounts_store.rs#L888)
never returns `ParticipateSwap`. And as far as I can tell, it never did.
So there is no point in checking if `transaction_type` matches
`ParticipateSwap`.

Also, if it would match `ParticipateSwap`, we wouldn't do anything once
the `MultiPartTransactionToBeProcessed` comes out of the queue.

# Changes

1. Remove `TransactionType::ParticipateSwap`.
2. Stop matching against `TransactionType::ParticipateSwap`.
3. Turn the `match` into a simple `if` since we now only match against a
single remaining type.
4. Stop passing now unused parameters to `process_transaction_type`.
5. Remove `MultiPartTransactionToBeProcessed::ParticipateSwap`.
6. Stop matching against
`MultiPartTransactionToBeProcessed::ParticipateSwap`.

# Tests

Not tested

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary