Add ICP Swap URL to CSP (#5910)
# Motivation

When trying to load the ICP Swap tickers URL, we get an error because of
the Content Security Policy.

# Changes

1. Add the `${{ICP_SWAP_URL}}` placeholder to the content security
policy template such that the canister replaces it with the URL from the
canister args during installation. We do the same for the SNS aggregator
URL.

# Tests

1. Tested manually on DevEnv. Without this we can't load ICP Swap
tickers and with this we can.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet