Render balance in USD in TokenBalanceCell (#5916)
# Motivation

We want to show USD values of assets.

In this PR we went balance in USD in the tokens table, if the feature
flag is enabled.

Price data will not be available until (in another PR) we call
`loadIcpSwapTickers()`.

# Changes

1. In `TokenBalanceCell.svelte`, if `ENABLE_USD_VALUES` is enabled,
render either the USD value of the balance or `$-/-` if the price data
is not available.

# Tests

1. Unit tests added.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

<img width="784" alt="image"
src="https://github.com/user-attachments/assets/ad9320ea-aadd-4f8a-a239-d2153cd9369a">

# Todos

- [ ] Add entry to changelog (if necessary).
not yet