Fix: Pushes abandoned SNS to the bottom of the nsAggregatorStore (#5918)
# Motivation

As discussed
[here](https://dfinity.slack.com/archives/C03H6QEPW5D/p1733263773416409)
and
[here](https://dfinity.slack.com/archives/C039M7YS6F6/p1733302975333649),
we have a situation with the SNS `CYCLES-TRANSFER-STATION`. They have
decided to change all their metadata, resulting in the name and token
being displayed as `---`.

This second PR moves abandoned SNSs to the bottom of both the
`nsAggregatorStore` and any alphabetical sort.

# Changes

- Adds the hidden character `"\u200B"` to the beginning of the name of
an abandoned SNS, so alphabetical sorting pushes them to the end.
- Sorts `nsAggregatorStore` by the temporal parameter `isAbandoned`.

# Tests

- Unit test for the aggregator store derived data to verify that sort
functionality works as expected.

# Todos

- [x] Add entry to changelog (if necessary).

Prev. PR: #5915