chore: configure Vitest globalSetup for timezone (#5929)
# Motivation

We currently rely on
[package.json](https://github.com/dfinity/nns-dapp/blob/main/frontend/package.json#L21)
to set up the timezone for our tests. This works when you run them as a
Node script. However, other environments, like VSC, do not recognize
this configuration, causing time-dependent tests to fail.
Vitest offers a `globalSetup` feature that addresses this issue by
ensuring tests are always executed consistently.

# Changes

- Adds `globalSetup` configuration to the vitest.config.ts file.
- Removes timezone information from Vitest scripts in package.json

# Tests

- Not necessary

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary