NNS1-3480: Exposes transactionDirection from mapIcpTransactionToReport (#5924)
# Motivation

We want to clearly indicate the direction of a transaction without
exposing `from`, `to`, or `selfTransaction` to the consumer. The new
field will contain either `debit` or `credit` values, allowing the
consumer to choose how to display the information, such as using "+" or
"-", or green and red colors.

# Changes

- New field returned by `mapIcpTransactionToReport` indicating the
transaction direction: debit | credit

# Tests

- Updated Unit Tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary