Refresh voting power from details (#6006)
# Motivation

If there are inactive neurons, the user should be able to reset their
voting-power-refreshed timestamp from the neuron details page. In this
PR, we add a ["Confirm Following"
button](https://github.com/dfinity/nns-dapp/pull/5990) to the neuron
status field. It should be shown for all statuses except when the
followings of a neuron have already been reset.

# Changes

- Add the `ConfirmFollowingButton` to the `NnsNeuronRewardStatusAction`
component.

# Tests

- Updated.
- Tested locally:


https://github.com/user-attachments/assets/70d159fc-ee22-408e-b013-097f16a36e54


# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.