refactor: types toast messages  (#6012)
# Motivation

We want to define the types of potential messages that can be passed to
the `toast` store. We can extract this list of potential messages from
`I18n`.


https://github.com/user-attachments/assets/17c7d582-bf7b-42f4-b873-8aee9c2919a5

We can't enforce just `I18n` keys as we have some places were we
programmatically pass the value and we don't have control on the type.

# Changes

- Types `labelKey` based on `I18n` nested keys.

# Tests

- Not necessary

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary