Add HW support to refreshVotingPower (#6013)
# Motivation

Users should be able to refresh the voting power timestamp of a neuron
to avoid missing voting rewards. A new API, `RefreshVotingPower`, is
designed for this purpose. However, currently, the Ledger API does not
support this type of transaction. This PR addresses the issue and
provides a workaround. The `setFollowees` API is used for hardware HW
neurons, as updating following and voting resets the voting power
timestamp in the same way as the `RefreshVotingPower` call.

**Note:** The topic used to confirm the current state of the neuron does
not matter, except for NeuronManagement, which can only be handled by
controllers.

# Changes

- Update `refreshVotingPowerForNeurons` service to call `setFollowees`
with the current value for HW neurons.

# Tests

- Added.
- Tested locally to confirm that it is possible to reset the timestamp
for HW neurons.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.