Show stake in USD on SNS neuron details (#6044)
# Motivation

Want to show the stake in USD on the neuron details page similar to how
we show the account balance on the wallet page.

https://github.com/dfinity/nns-dapp/pull/6039 did it for NNS neurons.
This PR does it for SNS neurons.
And a third PR will make sure token prices are loaded when the neuron
page is deep linked.

# Changes

1. Use `HeadingSubtitleWithUsdValue` in `SnsNeuronPageHeading`. Some of
the slot content is temporarily duplicate between with and without
feature flag.
2. Pass the `ledgerCanisterId` from `SnsNeuronDetail` to
`SnsNeuronPageHeading` as it's needed to find the token price.

# Tests

1. Unit tests added.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neuron/?u=bd3sg-teaaa-aaaaa-qaaba-cai&neuron=d09433a0369029e7b1114df00138c0cf5bbbef6d0d150c0aa6dd0fd67a75f985

# Todos

- [ ] Add entry to changelog (if necessary).
not yet