Nns missing rewards indicator (#6068)
# Motivation

When the user has Nns neurons that haven’t been active and are set to
start losing rewards in 30 days or less, a red circle needs to be
displayed next to the “Neuron Staking” sidebar item and on in the Neuron
Staking page table to indicate that action is required.
In this PR, we introduce the indicator component, which will be used in
the navigation menu and the projects table.

# Changes

- New `NnsNeuronsMissingRewardsBadge` component.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.