Enable ENABLE_USD_VALUES_FOR_NEURONS in e2e tests  (#6052)
# Motivation

We are already showing USD values of token balances. We want to do the
same for neuron balances.
This is currently implemented behind feature flag
`ENABLE_USD_VALUES_FOR_NEURONS`.
In this PR we enable `ENABLE_USD_VALUES_FOR_NEURONS` in affected
end-to-end tests so that we can enable the feature flag on mainnet
config in a follow-up PR.

# Changes

1. Enable `ENABLE_USD_VALUES_FOR_NEURONS` in affected end-to-end tests.
2. Update screenshots.

# Tests

Test-only change.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet