SECFIND-244: Add optional block_index to canister details (#6111)
# Motivation

We want to move the fallback logic for canister creation to the
frontend.
See [this
doc](https://docs.google.com/document/d/1hjMSTzjnVbU9Q4rJk233M3uNcKC-RLhme25nahHpTZg/edit?tab=t.0#heading=h.qtiuv2x3gsjy)
for details.

In this PR we add an optional `block_index` field to canister details
(`AttachCanisterRequest` for setting and `CanisterDetails` for getting)
and support storing the `block_index` in the backend canister.

The `block_index` refers to the index of the funding transaction for
creating the canister.

# Changes

1. Add `block_index` field to `CanisterDetails`, `AttachCanisterRequest`
and `NamedCanister`.
2. In `attach_canister` store the `block_index` for new canisters or
update the `block_index` for existing canister.

# Tests

1. Unit tests added.
2. Tested manually that canister creation still works. The new
functionality is not yet used in the frontend.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet