feat: add accounts-balance loading service for the Portfolio page (#6090)
# Motivation

The new Portfolio page shares some similarities with the existing
`/token` page, as both aim to load the balancers for SNS, ICRP, and BTC.
They load these accounts each time the user navigates to the respective
pages, but they try to avoid loading them multiple times due to
re-renders by tracking the requested accounts.

This first PR extracts some of the logic for use in both pages.

A follow-up PR will implement this logic for the Portfolio page. Later,
it can also be applied to the Token page.

[Ticket](https://dfinity.atlassian.net/browse/NNS1-3520)

# Changes

- `loadSnsAccountsBalances` loads the balances of SNS projects.  
-  `loadAccountsBalances` loadsthe balances of ICRC and ckBTC.  
- A mechanism tracks loaded balances and cleans them when the consumer
page is mounted.

# Tests

- Unit tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Next PR: #6116