NNS1-3504: Split actionable-proposals.derived.ts (#6127)
# Motivation

We want to sort the selectable universes based on their number of
actionable proposals.
For this `selectable-universes.derived.ts` needs to depend on
`actionableProposalCountStore` from `actionable-proposals.derived.ts`.
But this causes a circular dependency because of other derived stores in
`actionable-proposals.derived.ts` which already depend on
`selectable-universes.derived.ts`.

# Changes

1. Move the derived stores in `actionable-proposals.derived.ts` that
depend on `selectable-universes.derived.ts` to a new file named
`actionable-universes.derived.ts`.
2. Update imports.

# Tests

1. Move tests for the moved stores to a corresponding test file.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary