SECFIND-244: Add notifyAndAttachCanisterIfNeeded service (#6130)
# Motivation

We want to recover from interrupted canister creation in the frontend
code.
We want to do this by detecting canister creation funding transactions
without a corresponding canister.
See [this
doc](https://docs.google.com/document/d/1hjMSTzjnVbU9Q4rJk233M3uNcKC-RLhme25nahHpTZg/edit?tab=t.0#heading=h.qtiuv2x3gsjy)
for details.

This PR adds a service function to do this, but does not yet use it.

# Changes

1. Add `notifyAndAttachCanisterIfNeeded` which can be passed a list of
transactions and a list of canisters. If the transactions contain a
canister creation funding transaction that does not correspond to an
existing canister, it will try to recover the canister creation process.

# Tests

1. Added unit tests.
2. Tested manually in a branch with the end-to-end solution.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet