NNS1-3504: Sort selectable universes by actionable proposals (#6131)
# Motivation

On the voting page we want to sort the universes (both in the selector
and in the actionable proposals) based on the number of actionably
proposals or otherwise alphabetically.

Both the selector and the actionable proposals sections are rendered
based on `selectableUniversesStore`.
(This store is also used for other things, such as [finding the
currently selected
universe](https://github.com/dfinity/nns-dapp/blob/58b8d349c2912b678f69b1060342430f6992fb12/frontend/src/lib/components/accounts/IcrcWalletPage.svelte#L166),
which is why it can still also contain non-voting universes such as
ckBTC.)

# Changes

1. Sort `selectableUniversesStore` based on actionable proposal count
and title.

# Tests

1. Simplify the existing tests for `selectableUniversesStore`. Because
of the re-ordering, the tests were confusing in their previous form.
2. Add unit tests with multiple SNS projects.
3. Adapt `ActionableProposals.spec.ts` to the new ordering.
4. Manually tested at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/proposals/?actionable

# Todos

- [x] Add entry to changelog (if necessary).