feat(portfolio): new utility to get top projects for the ProjectsCard (#6159)
# Motivation

The Portfolio page displays a list of projects that behaves differently
based on specific rules.

The `getTopProjects` function encapsulates the following logic:  

- It should return up to four projects.  
- When the user is logged out, the function returns the top four
projects, listing ICP first, followed by three sorted by name. This is
similar to how we do it in the Staking page when the user is logged out.
- When the user is logged in, it returns the projects with the highest
`stakeInUsd`. If ICP is present, it should be the first item in the
list.

# Changes

-  Added a new descending comparator, `compareProjectsByUsdBalance`
- Exported `compareIcpFirst` and `compareByProjectTitle` from
_staking.utils_ to build the sorting comparator.
-  Introduced a new utility, `getTopProjects`

# Tests

-  Created unit tests for `getTopProjects`

# Todos

-  [ ] Add an entry to the changelog (if necessary). Not necessary.
Not necessary.