feat(portfolio): implement ProjectsCard component  (#6166)
# Motivation

The `ProjectsCard` component on the Portfolio page displays the total
USD amount of staked in projects and a list of up to four projects. If
there is a difference of 1 or more with the number of top Tokens it
shows a information message.

Related PRs:
- Get top tokens #6159 

# Changes

- Adds the `ProjectsCard` component.

# Tests

- Unit tests for the component  
-  New page object for the component 
- Manually tested in [E2E
environment](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary