Get network economics service (#6174)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, they may change in the future. Therefore, we want to
retrieve them from the API. In this PR, we add the service and store to
use them in the following PRs.

# Changes

- New `networkEconomicsStore` store.
- New `loadNetworkEconomicsParameters` service (intentionally no error
toasts, because this doesn't block more important activities).

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.