Remove canister creation fallback from backend canister (#6196)
# Motivation

Creating a canister takes 3 steps:
1. Transferring ICP to the CMC.
2. Notifying the CMC to create the canister.
3. Attaching the canister to the user account on the NNS dapp.

If the process is interrupted after step 1 or 2, the user might end up
without ICP and without canister.
To avoid this we have code in the nns-dapp backend canister to detect
transactions from step 1 and perform step 2 and 3.
Since https://github.com/dfinity/nns-dapp/pull/6132 we do this in the
frontend so we no longer need to do it in the backend.

This PR stops doing this in the backend.
This allows a lot of code to be removed from the backend canister. This
will be done in subsequent PRs.

# Changes

1. Stop pushing `CreateCanisterV2` to
`multi_part_transactions_processor` when a canister creation funding
transaction is detected.

# Tests

1. No unit test failed so apparently the code that was removed wasn't
tested.
2. Tested manually that the backend no longer creates the canister after
interrupting the process.
3. Tested manually that the frontend still creates the canister after
interrupting the process.

# Todos

- [x] Add entry to changelog (if necessary).