SECFIND-244: Remove transaction processing from nns-dapp canister (#6203)
# Motivation

The nns-dapp canister processes every transaction from the ICP ledger.
It used to do this to check if there are any unfinished processes for
neuron staking/top-up or canister creation/top-up but these recovery
processes have all been moved to the frontend.
So the backend canister no longer needs to process transactions.

This PR removes the transaction processing logic and related code.
It does not yet remove the transaction queue
(`multi_part_transactions_processor`) because it is stored in stable
memory and needs special care to be removed.
Additionally, `hardware_wallets_and_sub_accounts` and
`block_height_synced_up_to` which are stored in stable memory can be
removed in follow-up PRs.
Another follow-up PR will be made to remove metrics which are no longer
relevant.

NOTE: I removed chunks of code in separate commits and made sure the
code compiles after every commit. This may or may not help in reviewing
the PR.

# Changes

1. Remove code to determine the type of a transaction.
2. Remove code to add transactions to the queue.
3. Remove code to process transactions from the queue.
4. Remove code to fetch transactions from the ICP ledger.
5. Remove code to notify the CMC.
6. Remove code to attach a canister based on a processed transaction
(while keeping the code to attach a canister based on a call to the
nns-dapp from the frontend).
7. Remove code to keep track of which transactions have been processed.
8. Remove code keeping track of transactions which we were not able to
parse.

# Tests

1. Some unit tests have been removed.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/.
Created a new account. Created a canister. Interrupted creating a
canister by reloading the page right after clicking "Confirm". Recovered
from the interruption by going from Canisters to Wallet (to see the
funding transaction) back to Canisters.

# Todos

- [x] Add entry to changelog (if necessary).
not user visible but seemed like a large enough change to mention