Import token by url query parameters (#6213)
# Motivation

To simplify the token import process, the user can import a custom CK
token by clicking a link, and the import token form will be pre-filled
with the parameters from the URL (e.g.
//nns.ic0.app/tokens/?**import-ledger-id**=_m7h5t-euaaa-aaaaa-qabia-cai_&**import-index-id**=_myg3h-jmaaa-aaaaa-qabiq-cai_).
Having the ledger canister ID is sufficient to add an imported token.

Here, we make these parameters available from the pageStore.


# Changes

- Read parameters from URL.
- Open the import modal when the import-ledger-id parameter is in the
URL (not part of this PR, but I didn’t find a better way to test the
changes).

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.