Economics based periodic confirmation utils (#6223)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, they may change in the future. Therefore, we want to
retrieve them from the API.

In this PR, we mark the current periodic confirmation utilities as
deprecated and add sibling utilities that expect the voting power
economics properties to be in the argument. The utilities will be
replaced in upcoming PRs.

# Changes

- Add sibling utils. The only logic change is that the new utilities
assume the neuron is active when a voting power parameter is not
available.

# Tests

- Added. They are similar to the old version utilities.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.