Fix confusing definition of MAX_SUB_ACCOUNT_ID (#6235)
# Motivation

In `create_sub_account` we try subaccount IDs up to (excluding)
`u64::MAX` = 255.
It's excluding because we use `..` and not `..=` in the loop.
This means that the largest subaccount ID we allow is 254 and that's
also the maximum number of subaccounts you can have.
This is surprising and feels like an off-by-one, but this is the
behavior we've had for a long time.

In https://github.com/dfinity/nns-dapp/pull/6224 we introduced a name
(`MAX_SUB_ACCOUNT_ID`) for the upper bound of the loop.
But because the loop is exclusive, this name is wrong.

# Changes

1. Fix the value of `MAX_SUB_ACCOUNT_ID` to be 254 instead of 255.
2. Fix the loop to be inclusive.

# Tests

1. Don't reuse the same constant in the test to make it more obvious
what the actual behavior is. The test passed before and after the
changes so no behavior has changed in this PR.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary