Migrate NnsNeuronRewardStatusAction to voting power economics (#6237)
# Motivation

Currently, the voting power parameters are hardcoded as half a year and
one month. However, actual values are already being retrieved from the
API (they are the same for now but may be updated in the future). The
general idea is that if the parameters are unavailable, we consider the
neuron active.

In this PR, we migrate `NnsNeuronRewardStatusAction` from constants to
the voting power economics store.

# Changes

- Switch from constants to related stores and utilities.
- Do not show the component without voting power economics available (it
is loaded on dapp initialisation).

# Tests

- The old tests should pass.
- Added: should be hidden w/o params.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.