Remove multi_part_transactions_processor from AccountsStore (#6243)
# Motivation

The nns-dapp canister no longer processes transactions so we no longer
need the `MultiPartTransactionProcessor`.
The `MultiPartTransactionProcessor` is stored in stable memory so before
we can remove it completely, we need to deploy a version that accepts
the `MultiPartTransactionProcessor` not being present in stable memory.
This is to make sure that we can always revert back to the previous
WASM.

# Changes

1. Remove `multi_part_transactions_processor` field from
`AccountsStore`.
2. When encoding `AccountsStore` to stable memory, encode a default
`MultiPartTransactionProcessor`.
3. When decoding `AccountsStore` from stable memory, accept anything in
the `MultiPartTransactionProcessor` slot.

# Tests

1. `MultiPartTransactionProcessor` was already not used anymore so no
tests need to be changed.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary