NNS1-3548: Add compare functions for tokens table (#6248)
# Motivation

We want to allow the user to change the order in the tokens table by
clicking on the table headers, similar to the neurons table.
When clicking the "Projects" header, the tokens should be ordered
alphabetically by name.
When clicking the "Balance" header, the tokens should be ordered by:
1. USD value of the balance
2. Tokens without exchange rate but positive balance before tokens with
zero balance.
3. Within tokens with the same (probably 0) USD value, they should be
sorted by:
    1. ICP first
    2. Known ck tokens
    3. Imported tokens
    4. Failed imported tokens
    5. Other tokens
    6. Alphabetically

This PR adds the sorting functions.
It does not add the functionality in the table.

# Changes

1. Add necessary compare functions.

# Tests

1. Unit tests added.
2. Tested manually (with additional changes) at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/tokens/

# Todos

- [ ] Add entry to changelog (if necessary).
not yet