Remove block_height_synced_up_to from AccountsStore (#6254)
# Motivation

The nns-dapp canister no longer processes ICP transactions so we don't
need to keep track of which block height we are at with processing
transactions.

# Changes

1. Remove field `block_height_synced_up_to` from `AccountsStore`.
2. When encoding the `AccountsStore` to stable memory, write `None` in
the slot of `block_height_synced_up_to`.
3. Ignore the slot of `block_height_synced_up_to` when decoding the
`AccountsStore` from stable memory.

# Tests

This was was already not used anymore so no tests needed to be changed.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary