chore: use render utils in test context wrapper (#6263)
# Motivation

Using the `$on` callback for testing is deprecated in Svelte v5 (#6020).
That is why we will use the `events` option of the render function to
replace those and why we are refactoring the code to integrate `$on` in
the test render utils (#6258).

However, without any other changes than replacing the testing-library
render function with the test render utils in the `context-wrapper.mock`
module - in `renderContextWrapper` - few tests start failings.

# Changes

- Replace render of testing-library with utils in context wrapper

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>