refactor: integrate test $on in render utils (#6258)
# Motivation

Using the `$on` callback for testing is deprecated in Svelte v5. To
align with the changes made in Gix-components (see PR
[#544](https://github.com/dfinity/gix-components/pull/554)), we will
adopt the `events` option of the `render` function.

To facilitate the upgrade and minimize changes in the Svelte v5 PR, we
have refactored the test to utilize a utility function.

# Changes

- Extend `render` utility to support events
- Replace all occurences of the keywords `$on()` with the usage of the
utilitity

# Summarized

```
component.$on("something", something)

// =>

render(Component, {events: {something}}
```

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>