Improve canisters e2e test (#6281)
# Motivation

We've observed
[flakiness](https://github.com/dfinity/nns-dapp/actions/runs/13017255077/job/36309718238?pr=6276)
in the canisters end-to-end test.
I was not able to reproduce this, but while looking into the test I
noticed something which seemed brittle.
On line 63, the test gets canister card page objects, and then reuses
those objects on 88 and below, even though the test navigates to a
different page (and back) on the lines in between.
And when it arrives, on line 88, those cards are not even on the page
(it's showing skeleton cards).
The test relies on the fact that getting the name of the card in
`expect(await myCanisterCard.getCanisterName()).toBe(newCanisterName);`
causes the test to wait for the name to exist in the DOM.
This seems very brittle and it would be better to explicitly wait for
the page content the be loaded and not reuse the page objects from
before.

# Changes

1. Wait for content to be loaded after navigating back to the canisters
page.
2. Get new page objects for the canister cards after navigating back to
the canisters page.

# Tests

1. I wasn't able to reproduce the flakiness but the tests still pass as
before.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary