Fix URL parameter double validation (#6280)
# Motivation

The importTokenLedgerId comes from $pageStore (derived from
app/stores/page), which dispatches an update twice on page load
(presumably due to the Svelte hydration mechanism). This results in
reading the canister ID from the URL twice, and if the ID format is
incorrect, the error warning appears twice as well.

To prevent this, we introduce a flags to check the URL params only once.
I also considered adding two derived stores (one per parameter), but
flags offer a leaner solution.

# Changes

- Use local flags to avoid multiple validation calls.

# Tests

- The test was changed and verified that it fails w/o the fix.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.