NNS1-3548: Custom sorting for projects table (#6286)
# Motivation

We want to allow the user to change the order in the projects table by
clicking on the table headers, similar to the neurons table.
When clicking the "Nervous Systems" header, the projects should be
ordered alphabetically by name.
When clicking the "Neurons" header, the projects should be ordered
number of neurons.
When clicking the "Stake" header, the projects should be ordered by:
1. USD value of the stake
2. Projects without exchange rate but positive number of neurons before
projects with zero neurons.
3. Within projects with the same (probably 0) USD stake, ICP should be
sorted before others.

# Changes

1. Change the columns spec to support custom sorting.
2. Connect the `projectsTableOrderStore` to the `ResponsiveTable` to
persist the order between navigations.

# Tests

1. Existing unit tests updated based on the different default ordering.
2. Unit tests added.
3. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/staking/

# Todos

- [x] Add entry to changelog (if necessary).