NNS1-3548: Custom sorting for tokens table (#6295)
# Motivation

We want to allow custom sorting on the tokens table.
But on mobile, the sorting icon clashes with the settings icon for "hide
zero balances".
So for now we only enable sorting on desktop.

# Changes

1. Bind `tokensTableOrderStore` to the tokens table on the tokens page
to make sorting available, as supported by
https://github.com/dfinity/nns-dapp/pull/6273
2. Do not bind any `order` in `NnsAccounts` because we don't want to
support sorting on the accounts page.
3. Pass `disableMobileSorting={true}` in `TokensTable` to disable
sorting on mobile.

# Tests

1. Update existing tests to match the new ordering (while keeping the
intended meaning of the test).
2. Add new unit tests to test the binding to the store.
3. Test that sorting is disabled on mobile.
4. Test that sorting is not enabled on the accounts page.
5. Test that mobile sorting is not disabled on the projects page or the
neurons page.
6. Update screenshots to include the sorting arrow.

# Todos

- [x] Add entry to changelog (if necessary).