Add setIcpSwapUsdPrices test util (#6296)
# Motivation

Setting the `icpSwapTickersStore` in tests can be confusing because we
usually want certain USD prices for certain tokens, but the tickers are
such that they specify how many of a token go in 1 ICP.
This means you have to divide in order to get the actual value you
should set.

# Changes

1. Add a test util function where you can specify the USD values you
want for each token and it will set the tickers store such that the
derived `icpSwapUsdPricesStore` has the given USD prices.

# Tests

1. Unit tests added.
2. Tried to use it in an existing test but will update existing tests in
a separate PR.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary