feat(navigation): add derived store for last tokens/portfolio visit (#6285)
# Motivation

The Accounts page features a custom back button that currently redirects
users to the Tokens page. However, this is no longer always accurate, as
users can now access this page from the Portfolio page.

Since users can also navigate from the Accounts page to the Wallet page,
we cannot rely on a local solution. The Accounts layout will be
unmounted, resulting in the loss of this information.

This second PR follows up on #6274 by creating a derived store that
checks the entry page for the Accounts page, whether it is the Portfolio
or the Tokens page.

A third PR will utilize this derived store in the Accounts layout to
determine the appropriate redirect.

# Changes

- New derived store to calculate the last visited page: Tokens or
Portfolio, defaulting to Tokens if neither was selected.

# Tests

- Added unit tests for the derived store.
- Manually tested
[here](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)
with additional changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary