Use setIcpSwapUsdPrices test util (#6304)
# Motivation

Setting the `icpSwapTickersStore` in tests can be confusing because we
usually want certain USD prices for certain tokens, but the tickers are
such that they specify how many of a token go in 1 ICP.
This means you have to divide in order to get the actual value you
should set.

https://github.com/dfinity/nns-dapp/pull/6296 added
`setIcpSwapUsdPrices` to make this easier.

This PR uses `setIcpSwapUsdPrices` in all the relevant tests.

# Changes

1. Add one more utility function `setIcpPrice` for convenience when
setting just the ICP price.
2. Use `setIcpSwapUsdPrices` or `setIcpPrice` in tests that need to
populate `icpSwapTickersStore`.
3. Update comments as appropriate.

# Tests

1. Unit test added for `setIcpPrice`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary