feat(navigation): implement back navigation for Accounts page (#6303)
# Motivation

The Accounts page features a custom back button that currently redirects
users to the Tokens page. However, this is no longer always accurate, as
users can now access this page from the Portfolio page.

Since users can also navigate from the Accounts page to the Wallet page,
we cannot rely on a local solution. The Accounts layout will be
unmounted, resulting in the loss of this information.

This third PR follows up on #6285 by using the derived store that checks
the entry page for the Accounts page, whether it is the Portfolio or the
Tokens page.

# Changes

- Use `accountsPageOrigin` to navigate to the account's page referrer.

# Tests

- Updated the unit test description for the default case.  
- Added a unit test to verify that the back button redirects to the
portfolio page when the user navigates from it.
- Manually tested
[here](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network).
- Removed the redundant `describe` wrapper since the feature flag is no
longer present.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary