feat(navigation): implement wallet page entry point tracking (#6321)
# Motivation

Following a similar approach to #6285, the Wallets page will feature
multiple return pages based on how it was accessed.

Currently, there are two entry points:
* The Accounts page, when clicking any account except the NNS account.
* The NNS account page, when navigating from the Accounts page to the
NNS account and then to the wallet.

A new entry point will be added to allow navigation back from the
Portfolio page when the user clicks on a token in the `HeldTokensTable`.

A follow-up PR will update the Wallet's layout to make use of this
store.

Note: It is easier to review by hiding the whitespaces
<img width="277" alt="Screenshot 2025-02-03 at 14 31 08"
src="https://github.com/user-attachments/assets/9c28443e-f013-4afa-8684-dcadc2125340"
/>

# Changes

- A new derived store has been created to determine the return
destination from the Wallet page.

# Tests

-  Unit tests have been implemented for this new derived store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary