feat(navigation): implement contextual back navigation for Tokens page (#6329)
# Motivation

The tokens page has three potential entry points:  
- The Accounts page, accessed by clicking any account except the NNS
account.
- The NNS account page, reached by navigating from the Accounts page to
the NNS account and then to the wallet.
- A new entry point will be added to allow navigation back from the
Portfolio page when the user clicks on a token in the HeldTokensTable.

This PR follows up on #6321 by redirecting users from the Tokens page
based on how they entered.

Note: Tests are easier to read by hiding the whitespaces.

# Changes

- Uses the new store `walletPageOrigin` to decide where to navigate back
from Portfolio page.

# Tests

- Added unit test to check that navigation is correct when coming from
Portfolio page.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

---------

Co-authored-by: Max Strasinsky <98811342+mstrasinskis@users.noreply.github.com>