No missing rewards tags for not votable neurons (#6358)
# Motivation

In today's implementation, you can have neurons that are unlocked but
have been inactive for so long that the "confirm following" warning
appears. This is undesired, so the followee confirmation warning
shouldn't be shown if the neuron cannot vote.

This PR ensures that the "missing rewards" tags are not shown for
neurons that do not have enough dissolve delay to vote.

# Changes

- Update tag generation logic.

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.