Remove AccountWrapper (#6362)
# Motivation

https://github.com/dfinity/nns-dapp/pull/6239 removed the
`hardware_wallets_and_sub_accounts` field from `AccountsStore`.
This field was also encoding to and decoded from stable memory.
In order not to violate the skipping quote when decoding
`AccountsStore`, we had to continue decoding it as
`HashMap<AccountIdentifier, AccountWrapper>` until we could encode an
empty structure by deploying a release.

Now that we've had a release we can decode to `candid::Reserved`, encode
to `HashMap::<AccountIdentifier, candid::Empty>` and remove the
`AccountWrapper` type.

# Changes

1. Remove `AccountWrapper`.
2. Encode the `hardware_wallets_and_sub_accounts` slot as
`HashMap::<AccountIdentifier, candid::Empty>`.
3. Decode the `hardware_wallets_and_sub_accounts` slot as
`candid::Reserved`.

# Tests

1. Downgrade-upgrade test passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary