Split config.json from dfx.json (#6346)
# Motivation

For years we've had problems with not being able to run certain `dfx`
commands from the nns-dapp repo directory because `dfx` is confused by
the `.networks.local` entry in our `dfx.json` which will isolate a `dfx`
instance running in that directory from one running from another
directory.

This is, for example, why we `cd $HOME`
[here](https://github.com/dfinity/nns-dapp/blob/acb99323c33d849b969c1e181ad6a6c4b987545a/scripts/dfx-snapshot-start#L59-L60).

This is extra confusing because it's not enforced strictly so only some
commands fail because of this.

For historical reasons we have a lot of configuration in our `dfx.json`
which is not read by `dfx` itself and only by our scripts.

By moving this configuration to a separate file, we can finally remove
the `.networks.local` entry and run all `dfx` commands from the same
directory.

# Changes

1. Split `config.json` from `dfx.json`.
2. Update scripts to read config from `config.json` instead of
`dfx.json`.
3. Stop `cd`ing to another directory before invoking `dfx`.
4. Format `config.json` in `scripts/fmt-json`.

# Tests

1. Manually ran `dfx-snapshot-start` and verified that it works without
`cd`'ing away.
2. Manually ran `scripts/update_ic_commit --crate sns_aggregator
--ic_commit "release-2024-10-11_14-35-overload"`.
3. Manually ran `scripts/update-snsdemo --dir ../../snsdemo/tree2
--release release-2025-02-05`.
4. Manually ran `scripts/setup`.
5. Manually ran `scripts/install-didc`.
6. Manually ran `scripts/ensure-required-didc-version`.

Although there is quite some coverage on CI and I try to test as much as
possible, it's quite possible that one of the periodic workflows will
fail and will have to be fixed. So we should be on the lookout for that.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary