Prepare for PocketIC snapshots (#6367)
# Motivation

We want to start using PocketIC to create snsdemo snapshots.
When running a PocketIC snapshot, there is no `replica` process running.
We use `pgrep replica` in several places to check if a replica is
running.

For now I'm not checking the scripts output so it still calls it a
"replica" when `dfx start` is running, regardless of whether it's
PocketIC or not.

# Changes

1. Where we check `pgrep replica`, also check `pgrep pocket-ic`.
2. In `scripts/nns-dapp/test-proposal-payload` use `dfx info
webserver-port` instead of `dfx info replica-port`, which is required
when running PocketIC and also works fine without it.

# Tests

1. Manually ran `scripts/nns-dapp/test-proposal-payload` with and
without PocketIC snapshot.
2. Relying on CI for the others.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary