No active field for non-votable neurons (#6375)
# Motivation

In today's implementation, you can have neurons that are unlocked but
have been inactive for so long that the "confirm following" warning
appears. This is undesired, so the followee confirmation warning
shouldn't be shown if the neuron cannot vote.

This PR ensures that the neuron activity field is not displayed on the
neuron details page when there is not enough dissolve delay to vote.

# Changes

- Updated visibility logic for the neuron activity field.

# Tests

- Added.
- Tested locally.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary