Advance specific time to pass transition (#6398)
# Motivation

In several tests we use `advanceTime()`, [which
calls](https://github.com/dfinity/nns-dapp/blob/164f78620a9c9e420bd02a661548ad19cb3fe6be/frontend/src/tests/utils/timers.test-utils.ts#L22)
`await vi.runOnlyPendingTimersAsync()` to make a UI transition finish
before continuing with the test.

Using `advanceTime()` without parameter, makes time advance as much as
is necessary to run the currently pending timers.

With Vitest 3, this no longer works in certain cases. I'm not sure why
exactly but it's possible that there or additional timers before the
transition timers, such that only those are passed and we're still
waiting for the transition timers.

Passing an explicit number of milliseconds to `advanceTime()` makes sure
that the transition timers are included in the timers being run.

# Changes

1. Pass explicit number of milliseconds to `advanceTime` in tests for
which this is necessary to make the tests pass if we update to Vitest 3.
2. Also remove one `//await advanceTime(500);` which was commented out
and apparently not necessary.

# Tests

1. Updated tests still pass.
3. Tests also pass in another branch where `vitest` is updated to
`3.0.5`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary