test(portfolio): add e2e tests and screenshots for Portfolio page (#6354)
# Motivation

The new Portfolio page is now the entry point for the nns-dapp. We aim
to maintain UI consistency during development and prevent any visual
bugs on this page. Therefore, we are taking screenshots for the major
viewports and the following scenarios:
-  Landing as an anonymous user  
-  Signed-in user with loaded assets 

It hides the proposals counter badge to prevent false negatives caused
by the parallelism of tests and the creation of dummy proposals by other
tests.

# Changes

- New e2e tests for the portfolio page.

# Tests

- New screenshots for the page

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary