Add comment about increasing account limit (#6437)
# Motivation

We have a hard account limit after which no more accounts can be added.
This is to prevent too many accounts being created at once.
But when we approach that number organically, the number should be
increased.
We have alerts set up to warn us when we approach the limit.
So those alerts should be increased as well when the limit is increased.

# Changes

Add a comment pointing to an example of how to increase the alert
limits.

# Tests

No tested

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary