Avoid failing orphaned promises in AccountMenu.spec.ts (#6448)
# Motivation

`AccountMenu.spec.ts` tries to add event listeners on elements that may
not exist.
If the menu is not opened during a test, the `addEventListener` call
will wait for the element to appear, which will fail.
But because the `addEventListener` call is not awaited, the test passes
before it fails.
However, with the Svelte 5 (and related dependencies) update, this is
caught and causes the test to fail.

# Changes

1. Only add the `addEventListener("click", mockLinkClickEvent)` in tests
that actually click on links and only after the menu is opened.

# Tests

1. Test still passes.
2. Test now also passes in https://github.com/dfinity/nns-dapp/pull/6020
if I add some extra `advanceTime()` calls.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary