CMC notify fallback only once per session (#6449)
# Motivation

If canister creation gets interrupted and next time we see a canister
funding transaction without a corresponding canister, we try to notify
the CMC to create and attach the canister again.

Since this is only a fallback mechanism, we don't need to do this more
than once per transaction per sessions.

Additionally, there is a bug in Svelte that causes `$:` blocks to
sometimes not run, even though a dependency variable has changed. I've
noticed this several times in `NnsWallet`. This bug seems to have been
fixed (or at least changed) in Svelte 5. As a result, it actually tries
to perform the canister creation fallback twice when the account
balances are loaded uncertified and certified. This causes a test
failure.

# Changes

1. Add a store to keep track of block indices for which we've tried
canister notification.
2. Don't try canister notification for the same block index within the
same session.

# Tests

1. Unit tests added.
2. Tested in https://github.com/dfinity/nns-dapp/pull/6020 that it makes
`src/tests/lib/pages/NnsWallet.spec.ts` pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary