chore(metrics): add plausible domain configuration per network (#6450)
# Motivation

We want to integrate Plausible to the nns-dapp. We also want to use two
different accounts one for mainnet and another one for everything else.
For that we need an env. variable.

The rest of the setup can be found in #6445.

Note: the plausible domain is used to send the data to a particular
Plausible account.

# Changes

1. Add an `plausible-domain` in `mainnet` and `general` sections of
`config.json`.
2. Support Plausible Domain in `config.sh`.
3. Create `PLAUSIBLE_DOMAIN` in
`frontend/src/lib/constants/environment.constants.ts`.
4. Ran `./scripts/nns-dapp/test-config --update` to update golden files.

# Tests

1. Ran config scripts and observed:
1. `DFX_NETWORK=mainnet ./config.sh` ->
`VITE_PLAUSIBLE_DOMAIN=nns.ic0.app`
2. `DFX_NETWORK=local ./config.sh` ->
`VITE_PLAUSIBLE_DOMAIN=test.nns.ic0.app`
2. Tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)
with #6445


# Todos

- [ ] Add entry to changelog (if necessary).
Not yet