feat(errors): util for canister out-of-cycles errors (#6465)
# Motivation

We want to track canisters being out of cycles so we can display
appropriate information in the ui. For that we need first to be able to
identify an error as a canister out-of-cycles error.

This PR adds a util to identify such an error. Follow-up PRs will make
use of it to fill the store introduce in #6464.

# Changes

- Util to identify canisters out-of-cycle errors based on the
error.message.

# Tests

- Unit test for the util.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary