Fix bug in DisburseSnsNeuronModal.spec.ts (#6490)
# Motivation

https://github.com/dfinity/nns-dapp/pull/1409 added a test with the
following line in `DisburseSnsNeuronModal.spec.ts`:
```
await waitFor(() => expect(syncSnsAccounts).not.toBeCalled());
```
This doesn't work because if `syncSnsAccounts` has not already been
called, this will pass immediately.
The test will not wait to see if perhaps `syncSnsAccounts` will be
called soon after.

Later https://github.com/dfinity/nns-dapp/pull/4455 refactored the code
to remove `syncSnsAccounts`. It changed the code to call
`loadSnsAccounts` (a subset of the functionality of `syncSnsAccounts`)
instead. And it changed the test to replace `syncSnsAccounts` with
`loadSnsAccounts`.
But `loadSnsAccounts` is in fact called during the test, inside
`executeTransaction` in `DisburseSnsNeuronModal.svelte`.
But this was not caught by the test because of how the `waitFor` line is
broken.

On closer inspection, it seems that this test is a counterpart of the
test before, where it checks that `loadSnsAccounts` is being called,
even without clicking the confirm button. It seems that the next test
should test that this doesn't happen if accounts are already loaded
because `icrcAccountsStore.reset();` is not called.

# Changes

1. Make the `"should not trigger the project account load if already
available"` test more similar to the `"should trigger the project
account load"` test by not clicking the confirm button.
2. Add a comment pointing out that `icrcAccountsStore` is not reset in
`"should trigger the project account load"`.
3. Use `runResolvedPromises` instead of `waitFor` to have the same
conditions in both tests.

# Tests

Before, the test would fail if an extra `await runResolvedPromises();`
was added. Now the test passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary