feat(store): add support for failed actionable SNS proposals in TableProjects (#6502)
# Motivation

Follow up on #6503 to introduce the Failed token when building the
`TableProjects` if the actionable SNS fails.

Subsequent PRs will utilize this new field in the UI to display a
warning message with a tooltip.

Note: A default value of `[]` has been defined for
`failedActionableSnsesStore` in `getTableProjects` to minimize the
changes required for this PR.

<img width="1480" alt="Screenshot 2025-02-24 at 08 20 45"
src="https://github.com/user-attachments/assets/a64ac121-a465-482e-bfcf-804f29633117"
/>

# Changes

- Updates `actionable-sns-proposals.store.ts` to expose the data type of
the `FailedActionableSnsesStore`
- Updates `getNeuronAggregateInfo` to return a payload with a
`FailedTokenValue` for the stake if the actionable proposal failed.

# Tests

- Added new unit test
- Manually tested with https://github.com/dfinity/nns-dapp/pull/6485

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet

Prev. PR: #6503 | Next PR: #6504