Navigate on row click, stake on button click (#6486)
# Motivation

Currently, clicking on a token row on the staking page only opens the
stake-neuron modal, and there is no way to navigate to the neurons page
without knowing the direct URL.

Navigation should still be possible in cases where an issue occurs after
transferring the stake to the neuron subaccount and claiming the neuron
was not completed, as the neurons page in the NNS-dapp attempts to fix
these issues.

# Changes

- Ensure that navigation to the /neurons page is still possible when no
neurons exist.
- Open the Stake modal only when the “Stake TOKEN” button is clicked.

# Tests

- Update tests due to behavior changes.

# Todos

- [x] Add entry to changelog (if necessary).