Responsive table sort control (#6523)
# Motivation

To improve the UX, we are introducing the ability to control table
sorting from the settings popover. This ensures that table visibility
can also be managed on mobile devices, where sorting headers are not
available. With this PR, we add a chip-based sorting control component,
which will be used in the table settings popover.

Demo:
https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/tokens/

# Changes

- Add ResponsiveTableSortControl component.

# Tests

- Chip and ChipGroup POs.
- ResponsiveTableSortControl specs.
- Tested in draft PR: https://github.com/dfinity/nns-dapp/pull/6524
<img width="884" alt="image"
src="https://github.com/user-attachments/assets/b9f3e4da-e496-400f-a971-e0829ce2d849"
/>

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.