Add HideZeroNeuronsStore (#6509)
# Motivation

To improve the user experience, we are adding the ability to hide tokens
for which the user has no neurons (by default, all tokens are visible).
Here, we create a store to save the user’s selection in local storage.

The ticket: https://dfinity.atlassian.net/browse/NNS1-3602

# Changes

- New hideZeroNeuronsStore

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.