feat(navigation): add contextual back navigation for project pages (#6543)
# Motivation

The new Launchpad feature on the `Portfolio` page will display Cards
that direct users to a specific `Project` page. This page includes a
back button in the layout. Currently, this button always redirects the
user to the `Launchpad` page. However, we want to redirect users to
either the `Launchpad` or the `Portfolio` page, depending on their
previous location.

This initial PR adds a derived store to determine the appropriate page
for redirection.

[NNS1-3640](https://dfinity.atlassian.net/browse/NNS1-3640)

# Changes

- Added new derived route for the Projects page.

# Tests

- Added unit test

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

[NNS1-3640]:
https://dfinity.atlassian.net/browse/NNS1-3640?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ