feat(ui): add auto-rotating stacked cards component (#6571)
# Motivation

We want to display all the new launches on the Portfolio page. To
achieve this, we will use a stacked card component that shows an active
card, which changes over time. It also allows users to manually change
the active card.

This PR introduces a new reusable component that will be used later on
the Portfolio page.


https://github.com/user-attachments/assets/c85c5772-11de-4d66-8b7b-91d564cc3c01

[NNS1-3640](https://dfinity.atlassian.net/browse/NNS1-3640)

# Changes

- Adds a new component, `StackedCards`, to render multiple cards that
automatically change over time.

# Tests

- Added unit tests
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

[NNS1-3640]:
https://dfinity.atlassian.net/browse/NNS1-3640?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ