refactor(transactions): reuse transaction fee component in SplitStake modals (#6597)
# Motivation

In the scope of
[NNS1-3671](https://dfinity.atlassian.net/browse/NNS1-3671), we aim to
display the USD value of transaction fees. We already have a component
for this purpose, but it is not utilized consistently. This PR
eliminates redundant one-time implementations and reuses the existing
component. A follow-up PR will introduce the new logic to display the
USD fee value.

# Changes

- Reuse `TransactionFormFee` in the SplitNeuronModals components for
both NNs and SNs.
- Remove redundant translations since they all use the same key.

# Tests

- Test should pass as before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3671]:
https://dfinity.atlassian.net/browse/NNS1-3671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ