feat(ui): create fiat value display component for inputs (#6599)
# Motivation

In the scope of
[NNS1-3671](https://dfinity.atlassian.net/browse/NNS1-3671), we want to
display the USD value of inputs when the user enters a non-fiat value.
This PR introduces a new component that will later be used by the
`AmountInput` to show the fiat value.

<img width="902" alt="Screenshot 2025-03-19 at 13 15 53"
src="https://github.com/user-attachments/assets/aaaf33b9-c718-45df-b2b6-14973f4fea79"
/>

# Changes

- Adds a new component, `AmountInputFiatValue`, to display the USD value
and balance.
- Exposes a CSS variable to style the `font-weight` of `AmountDisplay`
for consumers.

# Tests

- Added unit tests.  
- Manually tested alongside other changes in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

[NNS1-3671]:
https://dfinity.atlassian.net/browse/NNS1-3671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ