feat(inputs): integrate USD value display in amount fields (#6601)
# Motivation

We want to display the USD value of a transaction being done in ICP.
This PR follows up on #6599 and incorporates this new component in the
`AmountInput`. Any component that uses it will automatically show the
USD value. Displaying the balance or maximum value is already available
but will be addressed in a follow-up PR, as it requires changes in the
consumers.


https://github.com/user-attachments/assets/3ee55f1e-164b-4fc9-9232-e676b9d24029

[NNS1-3671](https://dfinity.atlassian.net/browse/NNS1-3671)

# Changes

- Expose the newly added slot Gix component `Input` to the nns-dapp
`Input`.
- Created a new version of `MaxButton` to be used until the designs for
the `DayInput` are ready.
- Display `AmountInputFiatValue` in the newly added slot of
`InputWithError`.

# Tests

- Updated tests to check for the usd value.
- Manually tested with other changes in [devenv]()

# Todos

- [x] Add entry to changelog (if necessary).


[NNS1-3671]:
https://dfinity.atlassian.net/browse/NNS1-3671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ