refactor(forms): relocate balance display in amount inputs (#6606)
# Motivation

Following the work on
[NNS1-3671](https://dfinity.atlassian.net/browse/NNS1-3671), we want to
display the balance or maximum amount a user can use in a transaction
form closer to the input field and the maximum button.

Before:
<img width="695" alt="Screenshot 2025-03-20 at 07 35 16"
src="https://github.com/user-attachments/assets/7719fb29-d6b5-40fc-80a3-1077b9e102d0"
/>

After:
<img width="681" alt="Screenshot 2025-03-20 at 07 35 30"
src="https://github.com/user-attachments/assets/58741598-c11f-484a-bae4-337ee6bed66f"
/>


# Changes

- Extend `TransactionFromAccount` to show or hide the balance based on
the presence of the `Token` property. This change is backward
compatible, as the field behaves as it did for all consumers that still
provide the property.
- Update `NnsStakeNeuron` and `TransactionForm` to pass the balance to
the `AmountInput` field instead of to the `TransactionFromAccount`.
- Update translation key.

# Tests

- Updated tests to check for the balance.

# Todos

- [x] Add entry to changelog (if necessary).

[NNS1-3671]:
https://dfinity.atlassian.net/browse/NNS1-3671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ