build: svelte v5 (#6020)
# Motivation

We want to stay up-to-date with the tooling for best practices and
security reasons.

# Notes

We keep, when possible, the deprecated `ComponentType` in the source
code because if we move to the new types `Component`, component
`ResponsiveTyleRow` starts throwing a typing error in its template
`Error: Argument of type 'string' is not assignable to parameter of type
'never'.`. (see
[job](https://github.com/dfinity/nns-dapp/actions/runs/13107187031/job/36563870978))

Three screenshots were updated in this PR but, visually do no seem to
contains any changes.

# TODOs

Few features and tests need to be adapted for Svelte v5. Those leftovers
were marked with `// TODO: Svelte v5 migration` and have to be resolved
at latest before the release which will contains the upgrade but, after
this PR is merged.

# Changes

- Bump Gix-cmp for Svelte v5, Svelte, Vite and Testling-library
- Update component types in Tests (to resolve issue `Type
'LegacyComponentType' is not assignable to type 'typeof
SvelteComponent`).
- Remove `svelte/internal` in vitest config to fix `Error: Your
application, or one of its dependencies, imported from
'svelte/internal', which was a private module used by Svelte 4
components that no longer exists in Svelte 5. It is not intended to be
public API. If you're a library author and you used 'svelte/internal'
deliberately, please raise an issue on
https://github.com/sveltejs/svelte/issues detailing your use case.`
- Remove global override of `requestAnimationFrame` which leads to error
`Cannot assign to read only property 'requestAnimationFrame' of object
'#<Object>'` and has been resolved in testing library few month ago
according this
[comment](https://github.com/testing-library/svelte-testing-library/issues/206#issuecomment-1942132052).
- Patch configuration for IC-js libraries `spyOn` that ends in error
with `Cannot redefine property`
https://github.com/vitest-dev/vitest/issues/5625#issuecomment-2078969371
- Mock `animate`
(https://github.com/testing-library/svelte-testing-library/issues/284#issuecomment-2651620976)
- Update tests using `$on` and `$set` as documented
https://svelte.dev/docs/svelte/v5-migration-guide#Components-are-no-longer-classes
- Rename test files having to use `$state` for test purpose because
"rune is only available inside `.svelte` and `.svelte.js/ts` files"

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: David de Kloet <david.de.kloet@dfinity.org>
Co-authored-by: Max Strasinsky <maksims.strasinskis@dfinity.org>