refactor(validation): adds inline error message for neuron split modal (#6614)
# Motivation

Similar to #6602, we want to implement client validation to check
whether the value provided by the user for splitting a neuron is valid.
If it is not valid, we will display a message below the input field
instead of using a toast message.

In this PR, we address both flows: NNS and SNS for splitting neurons.

Before:


https://github.com/user-attachments/assets/756f468c-a134-462e-91da-8161642d8eae

After:


https://github.com/user-attachments/assets/291ae9e7-13d7-4eea-9ea8-f2da0796e6f6

[NNS1-3690](https://dfinity.atlassian.net/browse/NNS1-3690)

# Changes

- Refactor `isValidInputForm` from a utility function to a type
predicate. This change helps TypeScript recognize that `amount` is
defined when `isValidInputForm` returns true.
- Move the validation to the `split` closure for better clarity.  
- Remove the toaster message, as it was unnecessary; the button was
disabled when `isValidInputForm` was false. Additionally, it wasn't
cover by tests.
- Reuse the toaster message as an input error message.

# Tests

- Refactor `isValidInputForm` to ensure compatibility with the new
function signature.
- Add a test to verify that the error message is displayed correctly.

# Todos

- [x] Add entry to changelog (if necessary).

[NNS1-3690]:
https://dfinity.atlassian.net/browse/NNS1-3690?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ