feat(navigation): add contextual back navigation for proposals (#6632)
# Motivation

The Proposal page includes a back button in its layout. Currently, there
are two ways to access this page:
-  From the Proposals page for all proposals.  
-  From the Launchpad page to create a new SNS proposal.  

We will introduce a third option when the Portfolio page displays the
Proposal cards. We aim to implement a mechanism similar to #6543 to
determine the appropriate navigation based on how the user accessed the
page.

A follow-up PR will utilize this new store from the route's layout.

[NNS1-3638](https://dfinity.atlassian.net/browse/NNS1-3638)

# Changes

- A new derived store will define the navigation path back from the
Proposal page.

# Tests

Add unit tests for the new derived store.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary.

[NNS1-3638]:
https://dfinity.atlassian.net/browse/NNS1-3638?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ