feat(api): add option to include large proposal fields (#6647)
# Motivation

For the upcoming SNS proposals on the Portfolio page, we want to display
these proposals along with their SNS icons. To achieve this, we need to
disable the `omitLargeFields` flag in the request. This flag was
introduced in #5187 to address issues with payloads that are too large
when using filters on the `/proposals` page. This PR adds a flag to
request the entire payload. Since it will be used to fetch proposals
related to the topic `SnsAndCommunityFund`, it seems safe.

Note: In the past, issues arose when the payload exceeded approximately
3MB (3145728 bytes). Our current payload is around 60k, which would
allow for up to 50 proposals, while the estimate is to have no more than
5 at a time.

# Changes

- Exposes flag in `queryProposals`.

# Tests

- Adds unit test to check that property is part of request payload.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary