feat(services): expose large fields flag for SNS proposals (#6677)
# Motivation

For the upcoming SNS proposals on the Portfolio page, we want to display
these proposals along with their SNS icons. To achieve this, we need to
disable the `omitLargeFields` flag in the request. This PR follows up on
the work done in #6647 and exposes the flag in the service.

The service lacked tests, so this PR adds the minimum necessary to
ensure that the flag functions as intended while maintaining the
previous behavior as the default.

[NNS1-3638](https://dfinity.atlassian.net/browse/NNS1-3638)

# Changes

- Exposes flag in `loadProposalsSnsCF`.

# Tests

- Adds unit test to check that property is properly forwarded to
`queryProposals`.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

[NNS1-3638]:
https://dfinity.atlassian.net/browse/NNS1-3638?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ