chore(deps): bump ic-js and fix update list_proposals response structure (#6704)
# Motivation

In the scope of
[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666), we need to
filter proposals by topic. To achieve this, we must upgrade to the
latest version of [ic-js](https://github.com/dfinity/ic-js/pull/886).
The automatic job #6703 cannot upgrade the ic-js version to the latest
because it introduced a new field in the response of `list_proposals`,
which needs to be updated in the nns-dapp mock responses.

# Changes

- Checked out #6703 and created a new branch.  
- Updated the response for the `list_proposals` mocks

# Tests

- For now, tests should pass as before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>