feat(store): add topic filtering support to SNS filters (#6732)
# Motivation

We want to enable users to filter SNS proposals by topic. This filtering
will be saved in local storage to enhance the user experience. This PR
extends the `sns-filters.store` to support filtering by topic. Follow-up
PRs will utilize this new entry in the store to filter proposals and
update the store based on user interactions with the new topic filter
modal.

[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666)

# Changes

- Extends `sns-filters.store` with `setTopics ` and `setCheckTopics`.

# Tests

- Adds new tests to cover the latest entries in the store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ