feat(proposals): add SNS topic filtering support store (#6740)
# Motivation

We want users to filter proposals by topic; however, not all SNS may
have upgraded to the latest version, so not all will support this
feature. Therefore, the canister added a new field to the
[response](https://github.com/dfinity/ic/blob/579b8ba3a31341f354f4ddb3d60ac44548a91bc2/rs/sns/governance/canister/governance.did#L383)
that can be used to determine which filter to display to the user:
topics or types.

This PR introduces a new store to track those projects that don't
support yet this new API.

Note: As of today, most existing projects support the new API. The
approach is to track those that do not support it and update the UI
accordingly.

[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666)

# Changes

- New store to track a list of all sns that don't support the new API.

# Tests

- Unit tests for the new store.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ