fix(sns-topics): Hide remove button from the legacy followee tag (#6786)
# Motivation

Currently, there is a way to skip the confirmation step for the
"Catch-all" legacy following by clicking the **X icon** on the
"Catch-all" following tag instead of using the **"Deactivate
catch-all"** button. This is not the desired flow.

[NNS1-3791](https://dfinity.atlassian.net/browse/NNS1-3791).


![image](https://github.com/user-attachments/assets/8867d890-69de-48ac-b6de-48b6a3ee722d)

# Changes

- Removed the ability to delete the "Catch-all" legacy following using
the X icon on its tag, in order to enforce the confirmation step.

# Tests

- Added.
- Verified locally that the button is not being rendered.

<img width="564" alt="image"
src="https://github.com/user-attachments/assets/af4013e6-ab8b-451a-b48a-541b05076a46"
/>

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3791]:
https://dfinity.atlassian.net/browse/NNS1-3791?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ