feat(topics): implement topic priority sorting in store (#6788)
# Motivation

We want to apply the same sorting used for the proposals filter #6745 in
other parts of the application.

This PR implements the sorting logic directly in the derived store.
Since there is no need to retain the original data, we can apply it
directly to the store and avoid duplication.

The sorting logic is as follows:
- First, critical topics are sorted alphabetically.
- Second, non-critical topics are sorted alphabetically.

[NNS1-3792](https://dfinity.atlassian.net/browse/NNS1-3792)


https://github.com/user-attachments/assets/0660382f-3027-4f7b-b32f-c5dc0e42bfa5

# Changes

- Introduce sorting logic to `createSnsTopicsProjectStore`.

# Tests

- Add test to validate new sorting logic.
- Fix existing tests to easily visualize sorting.
- Tested in beta.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3792]:
https://dfinity.atlassian.net/browse/NNS1-3792?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ