Display vote delegation state in NNS neurons table (#6826)
# Motivation

To improve the visibility of a user’s vote delegation state, it should
be displayed directly in the neuron table by adding a new column: “Vote
Delegation”. This PR introduces a new function to calculate NNS neuron
vote delegation state.

# Changes

- New `getNnsNeuronVoteDelegationState` util.
- Add `voteDelegationState` field for NNS version of TableNeuron which
automatically makes the column to appear in the NNS neurons table.
- Make `voteDelegationState` not optional, since both SNS and NNS tables
now supports it.
- Update screenshots because of new field.

# Tests

- Tested `getNnsNeuronVoteDelegationState` util.
- Tested that the NNS version of `TableNeuron` contains
`voteDelegationState`.

# Todos

- [x] Add entry to changelog (if necessary).