refactor(api): migrate getLedgerId to next-gen index canister (#6830)
# Motivation

The ic-js library
[introduced](https://github.com/dfinity/ic-js/pull/562) the
next-generation version of ICRC index canisters. The nns-dapp was using
the old implementation until the new one was introduced in #6815, which
added a new method. The goal is to migrate existing methods to the
next-generation version and deprecate the old one from ic-js.

# Changes

- Migrate the `getLedgerId` API method to use the next-generation index
canister.

# Tests

- Update `getLedgerId` tests to use mocks and spies for the
next-generation canisters.
- Refactor `getTransactions` tests to follow the same structure as the
other tests.

# Todos

- [ ] Add an entry to the changelog (if necessary)
Not necessary.