refactor(api): migrate getTransactions to next-gen index canister (#6832)
# Motivation

The ic-js library
[introduced](https://github.com/dfinity/ic-js/pull/562) the
next-generation version of ICRC index canisters. The nns-dapp was using
the old implementation until the new one was introduced in #6815, which
added a new method. The goal is to migrate existing methods to the
next-generation version and deprecate the old one from ic-js.

# Changes

- Migrate the `getTransactions` API method to use the next-generation
index canister.
- Remove `indexCanister` as it is no longer used.

# Tests

- Update `getTransactions` tests to use mocks and spies for the
next-generation canisters.

# Todos

-  [ ] Add an entry to the changelog (if necessary)
Not necessary.