feat(display): implement hidden balance formatting in the USD value banner (#6850)
# Motivation

We want to introduce a toggle that allows users to hide or show their
balances on the main pages. This PR modifies the `UsdValueHeadless`
component to format the amount either as a number or the hidden
characters.


https://github.com/user-attachments/assets/1dacdc62-5afd-488c-93f4-7516c42592a9

[NNS1-3721](https://dfinity.atlassian.net/browse/NNS1-3721)

Note: This PR does not yet provide a way to toggle the hidden/show
value.

# Changes

- Format amounts based on the hide balance toggle.

# Tests

- Unit test to cover the hidden value.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3721]:
https://dfinity.atlassian.net/browse/NNS1-3721?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ