fix(scroll): upgrade gix to restore infinite scrolling (#6857)
# Motivation

We want to upgrade the version of Gix in the nns-dapp to address an
issue with Infinite Scrolling. The latest version of GIX introduced a
breaking change to `InfiniteScroll` when migrating it to Svelte 5.

This PR builds on top of #6859 and refactors how these components
consume the `InfiniteScroll`.

Note: This upgrade fixes the issue of all lists not fetching next pages.

Before:

Not all votes are being loaded: 


https://github.com/user-attachments/assets/4b71cf51-d18f-427e-a9f7-97f67adc4fa4

Not all proposals for a project are being loaded by scrolling:


https://github.com/user-attachments/assets/b81273ff-3d87-4b7d-a881-4fe0cb39c2b9


After:

All votes are being loaded: 


https://github.com/user-attachments/assets/bcf125c9-f567-48d7-9d30-b3b4c29b0291

All proposals for a project are being loaded by scrolling:


https://github.com/user-attachments/assets/6d9ba8f7-db19-414a-bcb8-24e0b7048159

# Changes

- Upgrade the GIX component library to the latest version.  
- Update the consumers of `InfiniteScroll` to the new interface.

# Tests

- Pipeline should work as before.
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network/portfolio/).

# Todos

- [x] Add entry to changelog (if necessary).