feat(menu): create balance privacy toggle component (#6865)
# Motivation

We want to introduce a toggle that allows users to hide or show their
balances on the main pages. This PR creates a new component that will be
displayed in the menu entries.

Note: The [Figma
designs](https://www.figma.com/design/29OdZHbNaOk9RkrfESucjb/NNS-1.03-WIP?node-id=12650-33136&t=dnaTDwnUEaOXBmlH-4)
display a differently styled checkbox. This change in the checkbox
design should be implemented consistently across all instances.

Note: This video is taken from a PR with additional changes.


https://github.com/user-attachments/assets/39949c19-66d2-4992-9d22-6d98ac5c5ff7

[NNS1-3721](https://dfinity.atlassian.net/browse/NNS1-3721)

# Changes

- New component to be displayed in the Menu with a toggle to change
between privacy modes.

# Tests

- Add a unit test to cover the new component.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.

[NNS1-3721]:
https://dfinity.atlassian.net/browse/NNS1-3721?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ