feat(disburse-maturity): api (#6872)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR introduces a basic API to disburse maturity to the user’s main
account only (sub-account and external address support planned for
later).

# Changes

- Add `disburseMaturity` api.
- Add `disburseMaturity` api-service which additionally invalidates the
cache.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.