feat(disburse-maturity): service (#6884)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR introduces a basic service to disburse maturity to the user’s
main account only (sub-account and external address support planned for
later).

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- Add disburseMaturity service.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet