feat(proposals): add collapsible vote condition status component (#6888)
# Motivation

We want to change how the voting results are displayed for a proposal.
This PR adds a new component based on `VotesResultsMajorityDescription`
to render a collapsible area with a title and an icon. The icon will
change based on the condition, and the content can be collapsed or
uncollapsed from the outside.

Note: `VotesResultsMajorityDescription` will be remove in a later
iteration.

Note 2: Screenshot has been taken with other changes not present in this
PR.

![Screenshot 2025-05-27 at 08 24
26](https://github.com/user-attachments/assets/5b6252bc-2830-4779-bd03-eee9bd7c2937)

[NNS1-3753](https://dfinity.atlassian.net/browse/NNS1-3753)

# Changes

- A new component to display the status of a condition for a proposal to
pass.

# Tests

- Unit tests for the new component.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3753]:
https://dfinity.atlassian.net/browse/NNS1-3753?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ