feat(disburse-maturity): Display "disburse" button (#6904)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR adds (behind the flag) the "Disburse" button to the neuron page.

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- New `NnsDisburseMaturityButton` component.
- Display disburse maturity button on page.
- Added a new modal key "disburse-maturity" to display a modal (comes
later).

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet