chore(cleanup): remove import token feature flag (#6912)
# Motivation

The nns-dapp uses feature flags to introduce new features into the
`main` branch without disrupting existing flows. Some of these feature
flags have been in the code for a long time, so we want to remove them.

This PR removes the feature flags introduced in #6209.

# Changes

- Removes `ENABLE_IMPORT_TOKEN_BY_URL` feature flag.
- Eliminates tests that check the UI behavior when the feature flag is
off.
- Updates tests to reflect the UI behavior when the feature flag is on,
as this is the default behavior.

# Tests

- The pipeline should pass as before.
- Tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.