feat(disburse-maturity): Disbursements in progress modal (#6919)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR adds a modal that contains the current disbursements in progress
status.


![image](https://github.com/user-attachments/assets/03136a11-436a-4bdd-9dae-93c2c9dbc232)

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- New `NnsActiveDisbursementModal` and `NnsActiveDisbursementEntry`
components.
- Displaying the modal.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet