feat(disburse-maturity): UI improvements (#6928)
# Motivation

To provide more context about the maturity disbursement process, this PR
updates the links to the relevant documentation pages (separating NNS
and SNS).

**Drive-by change**: The active disbursement count has been removed from
the modal title. It didn’t add much value, as there can be at most 10
active disbursements, all of which are already displayed within the
modal and this solves the issue when there is only a single entry.

| Before | After |
|--------|--------|
| ![Screenshot 2025-06-06 at 14 05
40](https://github.com/user-attachments/assets/f15a1407-5abf-424b-afe8-23a4841bd4b5)
| ![Screenshot 2025-06-06 at 14 08
30](https://github.com/user-attachments/assets/a5c0ce21-f410-479d-ada8-e56debc6a2ba)
|

# Changes

- Updated maturity description links. Separate copies for NNS and SNS
are now used to allow greater flexibility in the description text.
- Remove active disbursements count from title.

# Tests

- Added for the dynamic changes.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?