chore(cleanup): remove portfolio page feature flag (#7034)
# Motivation

The nns-dapp uses feature flags to introduce new features into the
`main` branch without disrupting existing flows. Some of these feature
flags have been in the code for a long time, so we want to remove them.

This PR removes the feature flags introduced in #6036.

# Changes

- Removes `ENABLE_PORTFOLIO_PAGE` feature flag.
- Eliminates tests that check the UI behavior when the feature flag is
off.
- Updates tests to reflect the UI behavior when the feature flag is on,
as this is the default behavior.

# Tests

- The pipeline should pass as before.
- Tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?