feat(tooltips): add configurable icon size (#7096)
# Motivation

The nns-dapp displays tooltips with an icon in various parts of the
application. The icon has a predefined size of 20px, which is no longer
compatible with the new font sizes defined in the app. This PR allows
the icon size to be set as a prop, enabling consumers to define it as
needed.

[NNS1-3922](https://dfinity.atlassian.net/browse/NNS1-3922)

# Changes

- Expose a new prop to define the icon size, with a default value of 20.

# Tests

- All existing tests should pass as before.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3922]:
https://dfinity.atlassian.net/browse/NNS1-3922?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ