feat(portfolio): integrate staking power information cards (#7088)
# Motivation

The nns-dapp will show staking power information on the portfolio page
using the new ApyCard. However, this requires a substantial amount of
data to load, which may lead to issues. This PR loads the new data and
displays new cards on the portfolio page. These cards are conditionally
rendered based on a feature flag.


https://github.com/user-attachments/assets/6aa0d42f-b430-49a1-a9e4-af26ac2643f8

[NNS1-3922](https://dfinity.atlassian.net/browse/NNS1-3922)

# Changes

- Retrieve `stakingRewardData`
- Render `ApyCard` along with a fallback version for error and loading
states.

# Tests

- Unit tests
- Manually tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3922]:
https://dfinity.atlassian.net/browse/NNS1-3922?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Francesco <francesco.rizzi@dfinity.org>