feat(launchpad2): Display sns weekly activity (#7112)
# Motivation

When user commitment is not available or does not exist, the proposal
weekly activity should be displayed in the third stat column of the
project card.

Currently, a dummy value is used for this field.  
This PR replaces the dummy value with a calculated value based on the
SNS project metrics.

# Changes

- Add new utility: `snsProjectWeeklyProposalActivity`.
- Replace the dummy value in the project card.

# Tests

- Updated.

<img width="1440" height="788" alt="image"
src="https://github.com/user-attachments/assets/47b0b894-ccc0-4d85-b0a1-58d70b133896"
/>


# Todos

- [ ] Accessibility (a11y) – any impact?
- [ ] Changelog – is it needed?