perf(apy): debounce reward calculation (#7120)
# Motivation

The new APY card estimates rewards based on a simulation of potential
earnings over a year. This simulation must be conducted for each project
and neuron. The more projects and neurons a user has, the longer the
simulation will take. The calculation is triggered as data sources
become available, leading to multiple executions in a cascading effect.
This PR minimizes this effect by debouncing the function call until the
data sources are ready.

[NNS1-3922](https://dfinity.atlassian.net/browse/NNS1-3922)

# Changes

- Debounce call to `getStakingRewardData`.

# Tests

- Update tests by advancing timers and waiting for a tick.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3922]:
https://dfinity.atlassian.net/browse/NNS1-3922?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ