chore(deps): update sns_aggregator candid bindings (#7218)
# Motivation
A newer release of the internet computer is available.
We would like to parse all the latest SNS data. To do so, we update the
candid interfaces for the SNS aggregator.
Even with no changes, just updating the reference is good practice.

Note: This PR was spun off from #7123 

Additionally, this PR fixes an issue with the aggregator-test job, which
was failing due to the following line that attempts to retrieve the
latest version's wasm:

```bash
curl -sL https://github.com/dfinity/nns-dapp/releases/download/aggregator-prod/sns_aggregator.wasm.gz >"$PROD_WASM"
```

The problem is that this version returns a 404 error because of an issue
with the release:
https://github.com/dfinity/nns-dapp/releases/tag/untagged-a9cd9aed86e4d5861ab6

Therefore, we will target the latest nightly release until we have a new
production version.

# Changes
* Update the version of `IC_COMMIT_FOR_SNS_AGGREGATOR` specified in
`config.json`.
* Updated the `sns_aggregator` candid files to the versions in that
commit.
* Updated the Rust code derived from `.did` files in the aggregator.

# Tests
- [ ] Please check the API updates for any breaking changes that affect
our code.
  Breaking changes are:
    * New mandatory fields
    * Removing mandatory fields
    * Renaming fields
    * Changing the type of a field
    * Adding new variants

---------

Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>