fix(neurons): correct minimum top-up stake validation (#7371)
# Motivation

The nns-dapp has a neuron top-up flow. This flow currently does not work
for empty neurons with maturity when the user inputs 1 ICP. The issue
arises because it checks if `amount_to_stake + neuron_stake > 1`, which
is incorrect. It should be `>= 1` since the transaction fees are
deducted from the account, not from the amount to stake.

The issue was introduced
[here](https://github.com/dfinity/nns-dapp/pull/1424/files#diff-b570daeb2a65094f5bd1d1b823bcf3025de7af68555903e0c2061d2f8b029e8fR736)

# Changes

- Changes the check for the minimum stake amount to be greater than or
equal to the 1.

# Tests

- Unit test for the minimum amount.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?