feat(utils): extend reporting utils for custom date range (#7390)
# Motivation

The nns-dapp will introduce a date range selector for the Reporting
Transactions feature. This initial PR prepares some of the utils to
support this new case.

Note: The implementation uses the local date and time to accommodate the
user's timezone. It is then converted to UTC in nanoseconds, as this is
the format supported by the API.

[NNS1-4134](https://dfinity.atlassian.net/browse/NNS1-4134)

# Changes

- Extend `formatDateCompact` to support delimiters on the dates.
- Extend `convertPeriodToNanosecondRange` to support custom date range
with provided `from` and `to`.

# Tests

- Added unit tests to both utils.
- Manually tested in a larger PR #7374 

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-4134]:
https://dfinity.atlassian.net/browse/NNS1-4134?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ