feat(transactions): implement memo field utils (#7399)
# Motivation

Support the transaction memo for both ICP and ICRC1 addresses for ICP
transaction flows. This PR adds utils to validate the provided memo:
* The `memo` field for ICP account transactions is an unsigned 64-bit
integer.
* The `icrc1Memo` field for ICRC1 account transactions is a byte array
with a length of up to 32 bytes.

# Changes

- Updated `mapIcpTransactionToUi` to prepare the `memo` field for the UI
when the value is present. It check first for `icrc1Memo` as this value
is optional, if not present it uses the value from the `memo` field that
is always present.
- Added `isValidIcpMemo`, `isValidIcrc1Memo`, and
`validateTransactionMemo` to validate the different types of memos.

# Tests

- Added unit tests for the utils
- Tested the completed feature in #7351.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?