feat(commands): add transaction memo toggle option (#7402)
# Motivation

Support the transaction memo for both ICP and ICRC1 addresses for ICP
transaction flows. This PR adds the options to the Command Palette for
users to turn it on/off. The feature will be added in a follow-up PR.

Note: The feature is not persisted in localStorage. Users will have to
turn it on on every session.

# Changes

- Add a memo option to the Command Palette.
- Change side: Move some translations to the translations file.

# Tests

- Added unit tests for the change.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?