feat(follow-nns-topics): following by topics after neuron stake (#7492)
# Motivation

To improve the voting delegation user experience, the NNS following
workflow will be redesigned to allow delegating voting by multiple
topics and to align it with the SNS version.

This PR replaces the last step of the staking neuron flow to utilize the
new follow-by-topics steps instead of the current single-step following
(behind the ENABLE_NNS_TOPICS feature flag).

https://dfinity.atlassian.net/browse/NNS1-4250

# Changes

- Replace the last step of the staking neuron flow to the new
follow-by-topics steps.

# Tests

- Added.
- Tested manually.


https://github.com/user-attachments/assets/4b451332-4464-4f16-8de6-35d51dfca0c4

# Todos

- [ ] Accessibility (a11y) – any impact?
- [ ] Changelog – is it needed?