feat(follow-nns-topics): following by topics flow (#7490)
# Motivation

To improve the voting delegation user experience, the NNS following
workflow will be redesigned to allow delegating voting by multiple
topics and to align it with the SNS version.

This PR updates the following card to make the topic definitions modal
accessible and opens the new follow-by-topics modal instead of the
current one (behind the `ENABLE_NNS_TOPICS` feature flag).

https://dfinity.atlassian.net/browse/NNS1-4250

# Changes

- Add a `topic-definitions-button` to the following card that opens the
modal.
- Open the new `FollowNnsNeuronsByTopicModal` instead of
`FollowNeuronsModal` when `ENABLE_NNS_TOPICS` is enabled.

# Tests

- Added.
- Tested manually.


https://github.com/user-attachments/assets/b128d6d2-a754-4228-b47e-92e31fa33cb2

# Todos

- [ ] Accessibility (a11y) – any impact?
- [ ] Changelog – is it needed?