feat(reporting): add ck-token transaction export flow integration (#7455)
# Motivation

A new option to export transactions from ck-tokens will be added to the
Reporting page. This PR adds the logic necessary to fetch all the data
for all the ck-tokens and generate a report with this information. A
follow-up, PR will introduce the option on the UI to select this new
option.

[NNS1-4231](https://dfinity.atlassian.net/browse/NNS1-4231)

# Changes

- Introduced a ck flow for the report generation based on a prop defined
by the parent.

# Tests

- Added tests to validate that functions are being called as expected.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed? - Not yet

[NNS1-4231]:
https://dfinity.atlassian.net/browse/NNS1-4231?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ