feat(follow-nns-topics): wizard step neuron (#7480)
# Motivation

To improve the voting delegation user experience, the NNS following
workflow will be redesigned to allow delegating voting by multiple
topics and to align it with the SNS version.

This PR introduces the second step in the Follow Neuron workflow, which
provides an interface for choosing which neuron to follow. These
components will also be used in the neuron staking flow for the same
purpose.

https://dfinity.atlassian.net/browse/NNS1-4250

# Changes

- Add `KnownNeuronFollowByTopicsItem` component.
- Add `FollowNnsNeuronsByTopicStepNeuron` component (duplicates a lot of
logic from `NewFolloweeModal`).

# Tests

- Added.
- Manually tested

<img width="627" height="766" alt="image"
src="https://github.com/user-attachments/assets/872a4cca-e60d-4420-9a54-81aef75f0555"
/>


# Todos

- [ ] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?