fix(following): recalculate parent collapsible when nested topics expand (#7522)
# Motivation

The new NNS neuron following flow has an issue when interacting with
other topics in the select topics modal step. #7521 updated the Gix
version to provide a way to control the recalculation of the collapsible
container from the outside.

This PR exposes the function of the outer collapsible to the inner
collapsible, allowing the child to invoke it when it expands. This
forces the parent to recalculate the space needed to display the
children.

Before:


https://github.com/user-attachments/assets/319efaef-2454-41fb-82a9-58ae22124c6d

After:


https://github.com/user-attachments/assets/62428c75-ab0f-4cdd-ac4a-d40f44bfe811

[NNS1-4251](https://dfinity.atlassian.net/browse/NNS1-4251)

# Changes

- Force the recalculation of container heights when interacting with
collapsible topics.

# Tests

- Visually tested

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-4251]:
https://dfinity.atlassian.net/browse/NNS1-4251?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ