refactor(fiat): add abstracted tickers store and service (#7546)
# Motivation

The nns-dapp currently has only one provider for ICP tickers. If there
is an issue with this provider, the dapp cannot display fiat values. To
prevent such situations, a secondary provider will be added. To keep
changes minimal, this first PR introduces a new store that abstracts the
tickers from the providers. This store is aliased from the old one. A
follow-up will remove the existing stores and replace them with the new
one.

[NNS1-4238](https://dfinity.atlassian.net/browse/NNS1-4238)

# Changes

- Introduced a `tickers` store to hold a map of canister IDs to USD
prices.
- Introduced a `tickers` service to fetch tickers from providers,
initially using the existing one.
- Introduced a ipcSwap provider to map from their data model to the
tickers model.

# Tests

- Updated tests.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed? Not yet.


[NNS1-4238]:
https://dfinity.atlassian.net/browse/NNS1-4238?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ