refactor(pricing): migrate to abstracted tickers store and setters (#7550)
# Motivation

The nns-dapp currently has only one provider for ICP tickers. If there
is an issue with this provider, the dapp cannot display fiat values. To
prevent such situations, a secondary provider will be added. This PR
builds on top of #7546 and refactors all uses of the old stores and
utils to the new one.

No logical changes in this PR.

[NNS1-4238](https://dfinity.atlassian.net/browse/NNS1-4238)

# Changes

- Replaced `icpSwapUsdPricesStore` with `tickersStore`.
- Replaced `setIcpSwapUsdPrices` with `setTickers`

# Tests

- CI should pass as before.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed? Not yet.


[NNS1-4238]:
https://dfinity.atlassian.net/browse/NNS1-4238?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ