feat(flags): disable NNS topics following feature flag (#7568)
# Motivation

The follow Neuron flow for NNS neurons was updated to match the user
experience of the SNS flow. This PR disables the feature flag.

Note: Since the feature was activated on #7540, the snapshot has been
updated with a new version of the governance canister. This version is
more restrictive when following neurons, which caused attempts to follow
a non-existing neuron `123` to fail the end-to-end test.

[NNS1-4191](https://dfinity.atlassian.net/browse/NNS1-4191)

# Changes

- Disable `ENABLE_NNS_TOPICS`  feature flag.
- Ran `./scripts/nns-dapp/test-config --update` to update golden config
files.

# Tests

- Verified locally that the feature is disabled by default.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?

[NNS1-4191]:
https://dfinity.atlassian.net/browse/NNS1-4191?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ