feat(frontend): upgrade ic-js and resolve breaking changes (#7599)
# Motivation

As
[mentionned](https://github.com/dfinity/nns-dapp/pull/7582#issuecomment-3546099291)
in the automatic PR that bumps ic-js, the latest version of the
libraries contain breaking changes.

On one end, we upgraded to the latest version of `@icp-sdk/bindgen`
which e.g. generate types as `Uint8Array` instead of union `Uint8Array |
number[]`. On the other and more minor changes, `canisterStatus`
requires an object as params.

# Changes

- Run `npm run upgrade:next`
- Resolve breaking changes