feat(frontend): use @icp-sdk/canisters with embedded cmc (#7612)
# Motivation

We want to embed the code in `@icp-sdk/canisters` instead of
re-exporting peer dependencies on legacy libraries or, conversely,
re-exporting `@icp-sdk/canisters` as a peer from the legacy
`@dfinity/...` dependencies.

We also want to use this opportunity to migrate the types to a DID and
IDL namespace approach.

We are going to make one major release, but we'll integrate the work
step by step, starting with the CMC.

# Changes

- Upgrade `@icp-sdk/canisters@next`
- Update types for DID namespace