feat(frontend): use icp-sdk/canisters with embedded ic-mgmt and camel case (#7615)
# Motivation

Same as #7612 but, for the `@icp-sdk/canisters/ic-management`.

In addition, we noticed inconsistency in the naming between upper case
and camel case (Icrc, Sns but CMC and ICManagement). Therefore we opted
for taking the opportunity of the upcoming breaking change (DID
namespace) to also resolve this "issue". We picked camel case, that's
why `ICManagementCanister` becomes `IcManagementCanister` and likewise
for `CMCCanister` which becomes `CmcCanister`.

# Changes

- Bump next
- Update naming as explained above
- Refactor also the naming of the custom `ICManagementCanister` class of
NNS dapp for consistency

# Notes

Could have shipped a separate PR for the later but, feel like it did not
arm to have it included here. If you which, I can revert and provide a
separate PR. Just let know.