feat(commands): add ICRC-1 account encoder utility to command palette (#7618)
# Motivation

The nns-dapp transaction forms accept both address formats: ICP and
ICRC. For the latter, users must provide the textual encoding of the
account. The
[ic-js](https://github.com/dfinity/icp-js-canisters/blob/2047aee31fabc8f69fef64c8ca43da897331cc92/packages/ledger-icrc/src/utils/ledger.utils.ts#L27)
library exposes a method for this, but no user interfaces are available.

This PR adds a utility to the Command Palette for generating such a
string.


https://github.com/user-attachments/assets/99be56d8-8210-4bac-8034-58098e0ceeaf

# Changes

- Introduced the concept of _utils_ to the Command Palette.
- Added a util to encode an ICRC-1 account.

# Tests

- Manually tested.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?