feat(frontend): use ic-js next with DID namespace for ledgers, Nns and Sns (#7630)
# Motivation

We want to move the source of of the ic-js legacy libraries to the
multi-entries library `@icp-sdk/canisters`.
At the same time, we also want to bring consistency and start using a
pattern we called "DID namespace usage".

Unfortunatelly, because of interdepency between ledgers used in Nns and
Sns libs, we had to pack everything at once.

This PRs bumps the latest next libs and adapt the interfaces.

NO functional changes.

# Notes

Few imports are updated for ICP and ICRC ledgers but, at the end of the
day, 90% of the imports modified are related to Sns.

# Changes

- `npm run upgrade:next`
- Update imports and usage for DID namespace