feat(frontend): use ic-js next + ledger and nns canisters renamed (prefixed) (#7631)
# Motivation

Parts of next ic-js is a bit of clean-up that includes renaming:
- `IcrcIndexNg...` to `IcrcIndex...`
- `LedgerCanister` to `IcpLedgerCanister`
- `IndexCanister` to `IcpIndexCanister`
- `GovernanceCanister` to `NnsGovernanceCanister`
- `TestGovernanceCanister` to `NnsTestGovernanceCanister`
- `GenesisTokenCanister` to `NnsGenesisTokenCanister`

This PR bumps next ic-js that includes those changes.

# Notes

FI team confirmed that they are not aware of any known Index canisters
that would not use the "index-ng" WASM. Likewise all SNSes have been
migrated to "index-ng" a year ago or so and the original "non-ng index"
source code has been removed from the IC repo. Therefore for consistency
and to make the tooling comprehensive, we deleted the old
implementation, kept "ng" but renamed it to `IcrcIndexCanister`.