chore(flags): remove ENABLE_DISBURSE_MATURITY feature flag and related tests (#7639)
# Motivation

The nns-dapp uses feature flags to introduce new features into the
`main` branch without disrupting existing flows. Some of these feature
flags have been in the code for a long time, so we want to remove them.

This PR removes the feature flags introduced in #6942.

# Changes

- Removes `ENABLE_DISBURSE_MATURITY` feature flag.
- Eliminates tests that check the UI behavior when the feature flag is
off.
- Updated configuration files with `./scripts/nns-dapp/test-config
--update`
- Sidetrack: Remove `swapCanisterId` from `createDummyProposals` as it
is not being used.

# Tests

- The pipeline should pass as before.

# Todos

- [x] Add entry to changelog (if necessary).
Not necessary.