chore(ci): npm run prepare (#7654)
# Motivation

When I prevented npm to run scripts in Juno and OISY
(https://github.com/dfinity/oisy-wallet/pull/10916) I faced the similar
issue as currently in #7653 - some checks and ci steps were no running
because they did not found the tsconfig. This happens because preventing
npm to run scripts on npm install has for results that svelte does not
run its scripts that create the `.svelte-kit` folder which contains the
`tsconfig` that is inherited from the configuration.

That's why I propose to resolve the issue by running `npm run prepare`
after `npm ci` in the CI.